Crate psoc6_01_pac

Peripheral access API for PSOC6_02 microcontrollers (generated using svd2rust v0.19.0 ( ))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Modules

backup
SRSS Backup Domain

cpuss
CPU subsystem (CPUSS)

csd0
Capsense Controller

dmac
DMAC

dw0
Datawire Controller

efuse
EFUSE MXS40 registers

fault
Fault structures

flashc
Flash controller

generic
Common register and bit access and modify traits

gpio
GPIO port control/configuration

hsiom
High Speed IO Matrix (HSIOM)

i2s0
I2S registers

ipc
IPC

lcd0
LCD Controller Block

lpcomp
Low Power Comparators

pass
PASS top-level MMIO (DSABv2, INTR)

pdm0
PDM registers

peri
Peripheral interconnect

peri_ms
Peripheral interconnect, master interface

profile
Energy Profiler IP

prot
Protection

sar
SAR ADC with Sequencer

scb0
Serial Communications Block (SPI/UART/I2C)

scb5
Serial Communications Block (SPI/UART/I2C)

sdhc0
SD/eMMC Host Controller

smartio
Programmable IO configuration

smif0
Serial Memory Interface

srss
SRSS Core Registers

tcpwm0
Timer/Counter/PWM

usbfs0
USB Host and Device Controller

Structs

BACKUP
SRSS Backup Domain

CBP
Cache and branch predictor maintenance operations

CPUID
CPUID

CPUSS
CPU subsystem (CPUSS)

CSD0
Capsense Controller

CorePeripherals
Core peripherals

DCB
Debug Control Block

DMAC
DMAC

DW0
Datawire Controller

DW1
Datawire Controller

DWT
Data Watchpoint and Trace unit

EFUSE
EFUSE MXS40 registers

FAULT
Fault structures

FLASHC
Flash controller

FPB
Flash Patch and Breakpoint unit

FPU
Floating Point Unit

GPIO
GPIO port control/configuration

HSIOM
High Speed IO Matrix (HSIOM)

I2S0
I2S registers

I2S1
I2S registers

IPC
IPC

ITM
Instrumentation Trace Macrocell

LCD0
LCD Controller Block

LPCOMP
Low Power Comparators

MPU
Memory Protection Unit

NVIC
Nested Vector Interrupt Controller

PASS
PASS top-level MMIO (DSABv2, INTR)

PDM0
PDM registers

PERI
Peripheral interconnect

PERI_MS
Peripheral interconnect, master interface

PROFILE
Energy Profiler IP

PROT
Protection

Peripherals
All the peripherals

SAR
SAR ADC with Sequencer

SCB
System Control Block

SCB0
Serial Communications Block (SPI/UART/I2C)

SCB1
Serial Communications Block (SPI/UART/I2C)

SCB2
Serial Communications Block (SPI/UART/I2C)

SCB3
Serial Communications Block (SPI/UART/I2C)

SCB4
Serial Communications Block (SPI/UART/I2C)

SCB5
Serial Communications Block (SPI/UART/I2C)

SCB6
Serial Communications Block (SPI/UART/I2C)

SCB7
Serial Communications Block (SPI/UART/I2C)

SCB8
Serial Communications Block (SPI/UART/I2C)

SCB9
Serial Communications Block (SPI/UART/I2C)

SCB10
Serial Communications Block (SPI/UART/I2C)

SCB11
Serial Communications Block (SPI/UART/I2C)

SCB12
Serial Communications Block (SPI/UART/I2C)

SDHC0
SD/eMMC Host Controller

SDHC1
SD/eMMC Host Controller

SMARTIO
Programmable IO configuration

SMIF0
Serial Memory Interface

SRSS
SRSS Core Registers

SYST
SysTick: System Timer

TCPWM0
Timer/Counter/PWM

TCPWM1
Timer/Counter/PWM

TPIU
Trace Port Interface Unit

USBFS0
USB Host and Device Controller

Enums

Interrupt
Enumeration of all the interrupts.

Constants

NVIC_PRIO_BITS
Number available in the NVIC for configuring priority