Module psoc6_01_pac::backup::alm2_time

Alarm 2 Seconds, Minute, Hours, Day of Week

Structs

ALM2_TIME_SPEC
Alarm 2 Seconds, Minute, Hours, Day of Week

ALM_DAY_EN_R
Field ALM_DAY_EN reader - Alarm Day of the Week enable: 0=ignore, 1=match

ALM_DAY_EN_W
Field ALM_DAY_EN writer - Alarm Day of the Week enable: 0=ignore, 1=match

ALM_DAY_R
Field ALM_DAY reader - Alarm Day of the week in BCD, 1-7 It is up to the user to define the meaning of the values, but 1=Monday is recommended

ALM_DAY_W
Field ALM_DAY writer - Alarm Day of the week in BCD, 1-7 It is up to the user to define the meaning of the values, but 1=Monday is recommended

ALM_HOUR_EN_R
Field ALM_HOUR_EN reader - Alarm hour enable: 0=ignore, 1=match

ALM_HOUR_EN_W
Field ALM_HOUR_EN writer - Alarm hour enable: 0=ignore, 1=match

ALM_HOUR_R
Field ALM_HOUR reader - Alarm hours in BCD, value depending on 12/24HR mode 12HR: [5]:0=AM, 1=PM, [4:0]=1-12 24HR: [5:0]=0-23

ALM_HOUR_W
Field ALM_HOUR writer - Alarm hours in BCD, value depending on 12/24HR mode 12HR: [5]:0=AM, 1=PM, [4:0]=1-12 24HR: [5:0]=0-23

ALM_MIN_EN_R
Field ALM_MIN_EN reader - Alarm minutes enable: 0=ignore, 1=match

ALM_MIN_EN_W
Field ALM_MIN_EN writer - Alarm minutes enable: 0=ignore, 1=match

ALM_MIN_R
Field ALM_MIN reader - Alarm minutes in BCD, 0-59

ALM_MIN_W
Field ALM_MIN writer - Alarm minutes in BCD, 0-59

ALM_SEC_EN_R
Field ALM_SEC_EN reader - Alarm second enable: 0=ignore, 1=match

ALM_SEC_EN_W
Field ALM_SEC_EN writer - Alarm second enable: 0=ignore, 1=match

ALM_SEC_R
Field ALM_SEC reader - Alarm seconds in BCD, 0-59

ALM_SEC_W
Field ALM_SEC writer - Alarm seconds in BCD, 0-59

R
Register ALM2_TIME reader

W
Register ALM2_TIME writer