Module psoc6_01_pac::backup::rtc_time

Calendar Seconds, Minutes, Hours, Day of Week

Structs

CTRL_12HR_R
Field CTRL_12HR reader - Select 12/24HR mode: 1=12HR, 0=24HR

CTRL_12HR_W
Field CTRL_12HR writer - Select 12/24HR mode: 1=12HR, 0=24HR

R
Register RTC_TIME reader

RTC_DAY_R
Field RTC_DAY reader - Calendar Day of the week in BCD, 1-7 It is up to the user to define the meaning of the values, but 1=Monday is recommended

RTC_DAY_W
Field RTC_DAY writer - Calendar Day of the week in BCD, 1-7 It is up to the user to define the meaning of the values, but 1=Monday is recommended

RTC_HOUR_R
Field RTC_HOUR reader - Calendar hours in BCD, value depending on 12/24HR mode 0=24HR: [21:16]=0-23 1=12HR: [21]:0=AM, 1=PM, [20:16]=1-12

RTC_HOUR_W
Field RTC_HOUR writer - Calendar hours in BCD, value depending on 12/24HR mode 0=24HR: [21:16]=0-23 1=12HR: [21]:0=AM, 1=PM, [20:16]=1-12

RTC_MIN_R
Field RTC_MIN reader - Calendar minutes in BCD, 0-59

RTC_MIN_W
Field RTC_MIN writer - Calendar minutes in BCD, 0-59

RTC_SEC_R
Field RTC_SEC reader - Calendar seconds in BCD, 0-59

RTC_SEC_W
Field RTC_SEC writer - Calendar seconds in BCD, 0-59

RTC_TIME_SPEC
Calendar Seconds, Minutes, Hours, Day of Week

W
Register RTC_TIME writer