Module psoc6_01_pac::cpuss::ap_ctl

Access port control

Structs

AP_CTL_SPEC
Access port control

CM0_DISABLE_R
Field CM0_DISABLE reader - Disables the CM0 AP interface: ‘0’: Enabled. ‘1’: Disabled. Typically, this field is set by the Cypress boot code with information from eFUSE. The access port is only enabled when CM0_DISABLE is ‘0’ and CM0_ENABLE is ‘1’.

CM0_DISABLE_W
Field CM0_DISABLE writer - Disables the CM0 AP interface: ‘0’: Enabled. ‘1’: Disabled. Typically, this field is set by the Cypress boot code with information from eFUSE. The access port is only enabled when CM0_DISABLE is ‘0’ and CM0_ENABLE is ‘1’.

CM0_ENABLE_R
Field CM0_ENABLE reader - Enables the CM0 AP interface: ‘0’: Disabled. ‘1’: Enabled.

CM0_ENABLE_W
Field CM0_ENABLE writer - Enables the CM0 AP interface: ‘0’: Disabled. ‘1’: Enabled.

CM4_DISABLE_R
Field CM4_DISABLE reader - Disables the CM4 AP interface: ‘0’: Enabled. ‘1’: Disabled. Typically, this field is set by the Cypress boot code with information from eFUSE. The access port is only enabled when CM4_DISABLE is ‘0’ and CM4_ENABLE is ‘1’.

CM4_DISABLE_W
Field CM4_DISABLE writer - Disables the CM4 AP interface: ‘0’: Enabled. ‘1’: Disabled. Typically, this field is set by the Cypress boot code with information from eFUSE. The access port is only enabled when CM4_DISABLE is ‘0’ and CM4_ENABLE is ‘1’.

CM4_ENABLE_R
Field CM4_ENABLE reader - Enables the CM4 AP interface: ‘0’: Disabled. ‘1’: Enabled.

CM4_ENABLE_W
Field CM4_ENABLE writer - Enables the CM4 AP interface: ‘0’: Disabled. ‘1’: Enabled.

R
Register AP_CTL reader

SYS_DISABLE_R
Field SYS_DISABLE reader - Disables the system AP interface: ‘0’: Enabled. ‘1’: Disabled. Typically, this field is set by the Cypress boot code with information from eFUSE. The access port is only enabled when SYS_DISABLE is ‘0’ and SYS_ENABLE is ‘1’.

SYS_DISABLE_W
Field SYS_DISABLE writer - Disables the system AP interface: ‘0’: Enabled. ‘1’: Disabled. Typically, this field is set by the Cypress boot code with information from eFUSE. The access port is only enabled when SYS_DISABLE is ‘0’ and SYS_ENABLE is ‘1’.

SYS_ENABLE_R
Field SYS_ENABLE reader - Enables the system AP interface: ‘0’: Disabled. ‘1’: Enabled.

SYS_ENABLE_W
Field SYS_ENABLE writer - Enables the system AP interface: ‘0’: Disabled. ‘1’: Enabled.

W
Register AP_CTL writer