Module psoc6_01_pac::cpuss::cm0_system_int_ctl

CM0+ system interrupt control

Structs

CM0_SYSTEM_INT_CTL_SPEC
CM0+ system interrupt control

CPU_INT_IDX_R
Field CPU_INT_IDX reader - CPU interrupt index (legal range [0, 7]). This field specifies to which CPU interrupt the system interrupt is mapped. E.g., if CPU_INT_IDX is ‘6’, the system interrupt is mapped to CPU interrupt ‘6’. Note: it is possible to map multiple system interrupts to the same CPU interrupt. It is advised to assign different priorities to the CPU interrupts and to assign system interrupts to CPU interrupts accordingly.

CPU_INT_IDX_W
Field CPU_INT_IDX writer - CPU interrupt index (legal range [0, 7]). This field specifies to which CPU interrupt the system interrupt is mapped. E.g., if CPU_INT_IDX is ‘6’, the system interrupt is mapped to CPU interrupt ‘6’. Note: it is possible to map multiple system interrupts to the same CPU interrupt. It is advised to assign different priorities to the CPU interrupts and to assign system interrupts to CPU interrupts accordingly.

CPU_INT_VALID_R
Field CPU_INT_VALID reader - Interrupt enable: ‘0’: Disabled. The system interrupt will NOT be mapped to any CPU interrupt. ‘1’: Enabled. The system interrupt is mapped on CPU interrupt CPU_INT_IDX. Note: the CPUs have dedicated XXX_SYSTEM_INT_CTL registers. In other words, the CPUs can use different CPU interrupts for the same system interrupt. However, typically only one of the CPUs will have the ENABLED field of a specific system interrupt set to ‘1’.

CPU_INT_VALID_W
Field CPU_INT_VALID writer - Interrupt enable: ‘0’: Disabled. The system interrupt will NOT be mapped to any CPU interrupt. ‘1’: Enabled. The system interrupt is mapped on CPU interrupt CPU_INT_IDX. Note: the CPUs have dedicated XXX_SYSTEM_INT_CTL registers. In other words, the CPUs can use different CPU interrupts for the same system interrupt. However, typically only one of the CPUs will have the ENABLED field of a specific system interrupt set to ‘1’.

R
Register CM0_SYSTEM_INT_CTL[%s] reader

W
Register CM0_SYSTEM_INT_CTL[%s] writer