Module psoc6_01_pac::csd0::adc_ctl

ADC Control

Structs

ADC_CTL_SPEC
ADC Control

ADC_MODE_R
Field ADC_MODE reader - Enable ADC measurement. When enabled the ADC sequencer will be started when the main sequencer goes to the SAMPLE_NORM state

ADC_MODE_W
Field ADC_MODE writer - Enable ADC measurement. When enabled the ADC sequencer will be started when the main sequencer goes to the SAMPLE_NORM state

ADC_TIME_R
Field ADC_TIME reader - ADC timing -1 in csd_sense clock cycles (actual time is ADC_TIME+1 cycles), either used to discharge Cref1&2, or as the aperture to capture the input voltage on Cref1&2

ADC_TIME_W
Field ADC_TIME writer - ADC timing -1 in csd_sense clock cycles (actual time is ADC_TIME+1 cycles), either used to discharge Cref1&2, or as the aperture to capture the input voltage on Cref1&2

R
Register ADC_CTL reader

W
Register ADC_CTL writer

Enums

ADC_MODE_A
Enable ADC measurement. When enabled the ADC sequencer will be started when the main sequencer goes to the SAMPLE_NORM state