Module psoc6_01_pac::csd0::config

Configuration and Control

Structs

CONFIG_SPEC
Configuration and Control

CSX_DUAL_CNT_R
Field CSX_DUAL_CNT reader - Enable the use of two counters for MUTUAL cap sensing mode (CSX), do not use when MUTUAL_CAP=0

CSX_DUAL_CNT_W
Field CSX_DUAL_CNT writer - Enable the use of two counters for MUTUAL cap sensing mode (CSX), do not use when MUTUAL_CAP=0

DSI_COUNT_SEL_R
Field DSI_COUNT_SEL reader - Select what to output on the dsi_count bus.

DSI_COUNT_SEL_W
Field DSI_COUNT_SEL writer - Select what to output on the dsi_count bus.

DSI_SAMPLE_EN_R
Field DSI_SAMPLE_EN reader - Enables the use of the dsi_sample_in input instead of the comparator output to strobe COUNTER.

DSI_SAMPLE_EN_W
Field DSI_SAMPLE_EN writer - Enables the use of the dsi_sample_in input instead of the comparator output to strobe COUNTER.

DSI_SENSE_EN_R
Field DSI_SENSE_EN reader - Enables the use of the dsi_sense_in input instead of the internally generated modulation signal to drive csd_sense and csd_shield signals.

DSI_SENSE_EN_W
Field DSI_SENSE_EN writer - Enables the use of the dsi_sense_in input instead of the internally generated modulation signal to drive csd_sense and csd_shield signals.

ENABLE_R
Field ENABLE reader - Master enable of the CSDv2 IP. Must be set to 1 for any CSDv2, ADC or IDAC operation to function. When 0 all analog components will be off and all switches will be open.

ENABLE_W
Field ENABLE writer - Master enable of the CSDv2 IP. Must be set to 1 for any CSDv2, ADC or IDAC operation to function. When 0 all analog components will be off and all switches will be open.

FILTER_DELAY_R
Field FILTER_DELAY reader - This value determines the number of cycles that the digital filter makes the CSDCMP output ignored while the counter counts and IDAC is on. When set to 0 the digital filter is off. When set to any other value the ignoring will last for FILTER_DELAY clk_csd cycles after the start of each measurement and from the first comparator trip to the end of each measurement.

FILTER_DELAY_W
Field FILTER_DELAY writer - This value determines the number of cycles that the digital filter makes the CSDCMP output ignored while the counter counts and IDAC is on. When set to 0 the digital filter is off. When set to any other value the ignoring will last for FILTER_DELAY clk_csd cycles after the start of each measurement and from the first comparator trip to the end of each measurement.

FULL_WAVE_R
Field FULL_WAVE reader - Enables full wave cap sensing mode

FULL_WAVE_W
Field FULL_WAVE writer - Enables full wave cap sensing mode

IREF_SEL_R
Field IREF_SEL reader - Select Iref supply.

IREF_SEL_W
Field IREF_SEL writer - Select Iref supply.

LP_MODE_R
Field LP_MODE reader - Select the power mode for the CSD components (REFGEN, AMBUF, CSDCMP, HSCMP): 0: High Power mode 1: Low Power mode

LP_MODE_W
Field LP_MODE writer - Select the power mode for the CSD components (REFGEN, AMBUF, CSDCMP, HSCMP): 0: High Power mode 1: Low Power mode

MUTUAL_CAP_R
Field MUTUAL_CAP reader - Enables mutual cap sensing mode

MUTUAL_CAP_W
Field MUTUAL_CAP writer - Enables mutual cap sensing mode

R
Register CONFIG reader

SAMPLE_SYNC_R
Field SAMPLE_SYNC reader - Enables double synchronizing of sample input from DSI (only relevant when DSI_SAMPLE_EN=1).

SAMPLE_SYNC_W
Field SAMPLE_SYNC writer - Enables double synchronizing of sample input from DSI (only relevant when DSI_SAMPLE_EN=1).

SENSE_EN_R
Field SENSE_EN reader - Enables the sense modulator output. 0: all switches, static or dynamic, are open and IDAC in CSD mode is off 1: switches and IDAC can be closed/on as per MMIO setting and CSD sequencer.

SENSE_EN_W
Field SENSE_EN writer - Enables the sense modulator output. 0: all switches, static or dynamic, are open and IDAC in CSD mode is off 1: switches and IDAC can be closed/on as per MMIO setting and CSD sequencer.

SHIELD_DELAY_R
Field SHIELD_DELAY reader - Selects the delay by which csd_shield is delayed relative to csd_sense.

SHIELD_DELAY_W
Field SHIELD_DELAY writer - Selects the delay by which csd_shield is delayed relative to csd_sense.

W
Register CONFIG writer

Enums

CSX_DUAL_CNT_A
Enable the use of two counters for MUTUAL cap sensing mode (CSX), do not use when MUTUAL_CAP=0

DSI_COUNT_SEL_A
Select what to output on the dsi_count bus.

FULL_WAVE_A
Enables full wave cap sensing mode

IREF_SEL_A
Select Iref supply.

MUTUAL_CAP_A
Enables mutual cap sensing mode

SHIELD_DELAY_A
Selects the delay by which csd_shield is delayed relative to csd_sense.