Module psoc6_01_pac::csd0::csdcmp

CSD Comparator configuration

Structs

AZ_EN_R
Field AZ_EN reader - Auto-Zero enable, allow the Sequencer to Auto-Zero this component

AZ_EN_W
Field AZ_EN writer - Auto-Zero enable, allow the Sequencer to Auto-Zero this component

CMP_MODE_R
Field CMP_MODE reader - Select which signal to output on dsi_sample_out.

CMP_MODE_W
Field CMP_MODE writer - Select which signal to output on dsi_sample_out.

CMP_PHASE_R
Field CMP_PHASE reader - Select in what phase(s) the comparator is active, typically set to match the BAL_MODE of the used IDAC. Note, this also determines when a bad conversion is detected, namely at the beginning and end of the comparator active phase (also taking into account FILTER_DELAY and non-overlap).

CMP_PHASE_W
Field CMP_PHASE writer - Select in what phase(s) the comparator is active, typically set to match the BAL_MODE of the used IDAC. Note, this also determines when a bad conversion is detected, namely at the beginning and end of the comparator active phase (also taking into account FILTER_DELAY and non-overlap).

CSDCMP_EN_R
Field CSDCMP_EN reader - CSD Comparator Enable

CSDCMP_EN_W
Field CSDCMP_EN writer - CSD Comparator Enable

CSDCMP_SPEC
CSD Comparator configuration

FEEDBACK_MODE_R
Field FEEDBACK_MODE reader - This bit controls whether the output directly from the comparator (csdcmp_out) or the flopped version (csdcmp_out_ff) is used. For CSD operation, the selected signal controls the IDAC(s), in GP mode the signal goes out on dsi_sample_out.

FEEDBACK_MODE_W
Field FEEDBACK_MODE writer - This bit controls whether the output directly from the comparator (csdcmp_out) or the flopped version (csdcmp_out_ff) is used. For CSD operation, the selected signal controls the IDAC(s), in GP mode the signal goes out on dsi_sample_out.

POLARITY_SEL_R
Field POLARITY_SEL reader - Select which IDAC polarity to use to detect CSDCMP triggering

POLARITY_SEL_W
Field POLARITY_SEL writer - Select which IDAC polarity to use to detect CSDCMP triggering

R
Register CSDCMP reader

W
Register CSDCMP writer

Enums

CMP_MODE_A
Select which signal to output on dsi_sample_out.

CMP_PHASE_A
Select in what phase(s) the comparator is active, typically set to match the BAL_MODE of the used IDAC. Note, this also determines when a bad conversion is detected, namely at the beginning and end of the comparator active phase (also taking into account FILTER_DELAY and non-overlap).

CSDCMP_EN_A
CSD Comparator Enable

FEEDBACK_MODE_A
This bit controls whether the output directly from the comparator (csdcmp_out) or the flopped version (csdcmp_out_ff) is used. For CSD operation, the selected signal controls the IDAC(s), in GP mode the signal goes out on dsi_sample_out.

POLARITY_SEL_A
Select which IDAC polarity to use to detect CSDCMP triggering