Struct psoc6_01_pac::csd0::idacb::W

pub struct W(_);

Register IDACB writer

Implementations

impl W

pub fn val(&mut self) -> VAL_W<'_>

Bits 0:6 - Current value setting for this IDAC (7 bits).

pub fn pol_dyn(&mut self) -> POL_DYN_W<'_>

Bit 7 - Polarity is dynamic, this bit does not influence the logic in the SoftIP, it only goes to the HardIP.

pub fn polarity(&mut self) -> POLARITY_W<'_>

Bits 8:9 - Selects the polarity of the IDAC (sensing operation). Normally the actual polarity depends on this bit, optionally mixed with DSI (see DSI_CTRL_EN) and if LEG1_EN==1 and LEG1_MODE==CSD also mixed with the CSD configuration and operation. In mutual cap mode however (see config.mutual_cap) the polarity of the IDAC is controlled by csd_sense. If LEG3_EN=1 (the other two legs must be off) then the ADC sequencer controls the IDACB polarity, optionally mixed with DSI.

pub fn bal_mode(&mut self) -> BAL_MODE_W<'_>

Bits 10:11 - same as corresponding IDACA Balancing mode

pub fn leg1_mode(&mut self) -> LEG1_MODE_W<'_>

Bits 16:17 - Controls the usage mode of LEG1 and the Polarity bit

pub fn leg2_mode(&mut self) -> LEG2_MODE_W<'_>

Bits 18:19 - Controls the usage mode of LEG2

pub fn dsi_ctrl_en(&mut self) -> DSI_CTRL_EN_W<'_>

Bit 21 - Mix DSI inputs with MMIO controls or not (before getting mixed with CSD controls if enabled) 0: no DSI control IDACB_POLARITY = IDACB.POLARITY IDACB_LEG1_EN = IDACB.LEG1_EN IDACB_LEG2_EN = IDACB.LEG2_EN IDACB_LEG3_EN = IDACB.LEG3_EN 1: Mix MMIO with DSI control IDACB_POLARITY = IDACB.POLARITY EXOR dsi_idacb_pol IDACB_LEG1_EN = IDACB.LEG1_EN AND dsi_idacb_leg1_en IDACB_LEG2_EN = IDACB.LEG2_EN AND dsi_idacb_leg2_en IDACB_LEG3_EN = IDACB.LEG3_EN AND dsi_idacb_leg3_en

pub fn range(&mut self) -> RANGE_W<'_>

Bits 22:23 - IDAC multiplier

pub fn leg1_en(&mut self) -> LEG1_EN_W<'_>

Bit 24 - output enable for leg 1 to CSDBUSB or CSDBUSA

pub fn leg2_en(&mut self) -> LEG2_EN_W<'_>

Bit 25 - output enable for leg 2 to CSDBUSB or CSDBUSA

pub fn leg3_en(&mut self) -> LEG3_EN_W<'_>

Bit 26 - output enable for leg3 to CSDBUSC, only allowed when RANGE = IDAC_LO. When this bit is set both other legs should be off. Note that leg3 can only be used for ADC mode, not GP mode. Which means that leg3 can only be on when the ADC Sequencer is in the ADC_measure or Calib_measure state. In those states leg3 is controlled by the ADC configuration and the HSCMP output. In addition this leg3 enable bit can optionally be mixed with DSI (see DSI_CTRL_EN). When LEG3_EN=1 also the IDACB polarity is controlled by the ADC sequencer.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.