Module psoc6_01_pac::csd0::refgen

Reference Generator configuration

Structs

BYPASS_R
Field BYPASS reader - Bypass selected input reference unbuffered to Vrefhi

BYPASS_W
Field BYPASS writer - Bypass selected input reference unbuffered to Vrefhi

GAIN_R
Field GAIN reader - Select resistor string tap for feedback, 0= minimum vout, 31= maximum vout = vrefhi -> gain=1 (only works if the resistor string is enabled; RES_EN=1)

GAIN_W
Field GAIN writer - Select resistor string tap for feedback, 0= minimum vout, 31= maximum vout = vrefhi -> gain=1 (only works if the resistor string is enabled; RES_EN=1)

R
Register REFGEN reader

REFGEN_EN_R
Field REFGEN_EN reader - Reference Generator Enable

REFGEN_EN_W
Field REFGEN_EN writer - Reference Generator Enable

REFGEN_SPEC
Reference Generator configuration

RES_EN_R
Field RES_EN reader - Resistor string enable; 0= open switch on top of the resistor string (Vreflo=Vssa)

RES_EN_W
Field RES_EN writer - Resistor string enable; 0= open switch on top of the resistor string (Vreflo=Vssa)

VDDA_EN_R
Field VDDA_EN reader - Close Vdda switch to top of resistor string (or Vrefhi?)

VDDA_EN_W
Field VDDA_EN writer - Close Vdda switch to top of resistor string (or Vrefhi?)

VREFLO_INT_R
Field VREFLO_INT reader - Ouput the resistor string tap either to Vreflo (0) or Vreflo_int (1).

VREFLO_INT_W
Field VREFLO_INT writer - Ouput the resistor string tap either to Vreflo (0) or Vreflo_int (1).

VREFLO_SEL_R
Field VREFLO_SEL reader - Select resistor string tap for Vreflo/Vreflo_int, 0= minimum vout, 31= maximum vout = vrefhi (only works if the resistor string is enabled; RES_EN=1)

VREFLO_SEL_W
Field VREFLO_SEL writer - Select resistor string tap for Vreflo/Vreflo_int, 0= minimum vout, 31= maximum vout = vrefhi (only works if the resistor string is enabled; RES_EN=1)

W
Register REFGEN writer

Enums

REFGEN_EN_A
Reference Generator Enable