Module psoc6_01_pac::dmac::ch::ctl

Channel control

Structs

B_R
Field B reader - Non-bufferable/bufferable access control: ‘0’: non-bufferable. ‘1’: bufferable. This field is used to indicate to an AMBA bridge that a write transaction can complete without waiting for the destination to accept the write transaction data. All transactions for this channel uses the B field for the non-bufferable/bufferable access control (‘hprot[2]’).

B_W
Field B writer - Non-bufferable/bufferable access control: ‘0’: non-bufferable. ‘1’: bufferable. This field is used to indicate to an AMBA bridge that a write transaction can complete without waiting for the destination to accept the write transaction data. All transactions for this channel uses the B field for the non-bufferable/bufferable access control (‘hprot[2]’).

CTL_SPEC
Channel control

ENABLED_R
Field ENABLED reader - Channel enable: ‘0’: Disabled. The channel’s trigger is ignored and the channel cannot be made pending and therefore cannot be made active. If a pending channel is disabled, the channel is made non pending. If the activate channel is disabled, the channel is de-activated (bus transactions are completed). ‘1’: Enabled. SW sets this field to ‘1’ to enable a specific channel. HW sets this field to ‘0’ when an error interrupt cause is activated.

ENABLED_W
Field ENABLED writer - Channel enable: ‘0’: Disabled. The channel’s trigger is ignored and the channel cannot be made pending and therefore cannot be made active. If a pending channel is disabled, the channel is made non pending. If the activate channel is disabled, the channel is de-activated (bus transactions are completed). ‘1’: Enabled. SW sets this field to ‘1’ to enable a specific channel. HW sets this field to ‘0’ when an error interrupt cause is activated.

NS_R
Field NS reader - Secure/on-secure access control: ‘0’: secure. ‘1’: non-secure. This field is set with the secure/non-secure access control of the transaction that writes this register; i.e. the access control is inherited from the write transaction and not specified by the transaction write data. All transactions for this channel use the NS field for the secure/non-secure access control (‘hprot[4]’).

NS_W
Field NS writer - Secure/on-secure access control: ‘0’: secure. ‘1’: non-secure. This field is set with the secure/non-secure access control of the transaction that writes this register; i.e. the access control is inherited from the write transaction and not specified by the transaction write data. All transactions for this channel use the NS field for the secure/non-secure access control (‘hprot[4]’).

PC_R
Field PC reader - Protection context. This field is set with the protection context of the transaction that writes this register; i.e. the context is inherited from the write transaction and not specified by the transaction write data. All transactions for this channel uses the PC field for the protection context.

PC_W
Field PC writer - Protection context. This field is set with the protection context of the transaction that writes this register; i.e. the context is inherited from the write transaction and not specified by the transaction write data. All transactions for this channel uses the PC field for the protection context.

PRIO_R
Field PRIO reader - Channel priority: ‘0’: highest priority. ‘1’ ‘2’ ‘3’: lowest priority. Channels with the same priority constitute a priority group and within this priority group, the following ‘roundrobin’ arbitration is applied. A ‘round’ consists of a contiguous sequence of channel activations, within this priority group, without any repetition. Within a round, higher priority is given to the lower channel indices. The notion of a round guarantees that within a group, higher channel indices do not yield to lower indices indefinitely.

PRIO_W
Field PRIO writer - Channel priority: ‘0’: highest priority. ‘1’ ‘2’ ‘3’: lowest priority. Channels with the same priority constitute a priority group and within this priority group, the following ‘roundrobin’ arbitration is applied. A ‘round’ consists of a contiguous sequence of channel activations, within this priority group, without any repetition. Within a round, higher priority is given to the lower channel indices. The notion of a round guarantees that within a group, higher channel indices do not yield to lower indices indefinitely.

P_R
Field P reader - User/privileged access control: ‘0’: user mode. ‘1’: privileged mode. This field is set with the user/privileged access control of the transaction that writes this register; i.e. the access control is inherited from the write transaction and not specified by the transaction write data. All transactions for this channel use the P field for the user/privileged access control (‘hprot[1]’).

P_W
Field P writer - User/privileged access control: ‘0’: user mode. ‘1’: privileged mode. This field is set with the user/privileged access control of the transaction that writes this register; i.e. the access control is inherited from the write transaction and not specified by the transaction write data. All transactions for this channel use the P field for the user/privileged access control (‘hprot[1]’).

R
Register CTL reader

W
Register CTL writer