Module psoc6_01_pac::dw0::ch_struct::ch_ctl

Channel control

Structs

B_R
Field B reader - Non-bufferable/bufferable access control: ‘0’: non-bufferable. ‘1’: bufferable. This field is used to indicate to an AMBA bridge that a write transaction can complete without waiting for the destination to accept the write transaction data. All transactions for this channel uses the B field for the non-bufferable/bufferable access control (‘hprot[2]’).

B_W
Field B writer - Non-bufferable/bufferable access control: ‘0’: non-bufferable. ‘1’: bufferable. This field is used to indicate to an AMBA bridge that a write transaction can complete without waiting for the destination to accept the write transaction data. All transactions for this channel uses the B field for the non-bufferable/bufferable access control (‘hprot[2]’).

CH_CTL_SPEC
Channel control

ENABLED_R
Field ENABLED reader - Channel enable: ‘0’: Disabled. The channel’s trigger is ignored and the channel cannot be made pending and therefore cannot be made active. If a pending channel is disabled, the channel is made non pending. If the activate channel is disabled, the channel is de-activated (bus transactions are completed). ‘1’: Enabled. SW sets this field to ‘1’ to enable a specific channel. HW sets this field to ‘0’ on an error interrupt cause (the specific error is specified by CH_STATUS.INTR_CAUSE).

ENABLED_W
Field ENABLED writer - Channel enable: ‘0’: Disabled. The channel’s trigger is ignored and the channel cannot be made pending and therefore cannot be made active. If a pending channel is disabled, the channel is made non pending. If the activate channel is disabled, the channel is de-activated (bus transactions are completed). ‘1’: Enabled. SW sets this field to ‘1’ to enable a specific channel. HW sets this field to ‘0’ on an error interrupt cause (the specific error is specified by CH_STATUS.INTR_CAUSE).

NS_R
Field NS reader - Secure/on-secure access control: ‘0’: secure. ‘1’: non-secure. This field is set with the secure/non-secure access control of the transaction that writes this register; i.e. the ‘write data’ is ignored and instead the access control is inherited from the write transaction (note the field attributes should be HW:RW, SW:R). All transactions for this channel use the NS field for the secure/non-secure access control (‘hprot[4]’).

NS_W
Field NS writer - Secure/on-secure access control: ‘0’: secure. ‘1’: non-secure. This field is set with the secure/non-secure access control of the transaction that writes this register; i.e. the ‘write data’ is ignored and instead the access control is inherited from the write transaction (note the field attributes should be HW:RW, SW:R). All transactions for this channel use the NS field for the secure/non-secure access control (‘hprot[4]’).

PC_R
Field PC reader - Protection context. This field is set with the protection context of the transaction that writes this register; i.e. the ‘write data’ is ignored and instead the context is inherited from the write transaction (note the field attributes should be HW:RW, SW:R). All transactions for this channel uses the PC field for the protection context.

PC_W
Field PC writer - Protection context. This field is set with the protection context of the transaction that writes this register; i.e. the ‘write data’ is ignored and instead the context is inherited from the write transaction (note the field attributes should be HW:RW, SW:R). All transactions for this channel uses the PC field for the protection context.

PREEMPTABLE_R
Field PREEMPTABLE reader - Specifies if the channel is preemptable. ‘0’: Not preemptable. ‘1’: Preemptable. This field allows higher priority pending channels (from a higher priority group; i.e. an active channel can NOT be preempted by a pending channel in the same priority group) to preempt the active channel in between ‘single transfers’ (a 1D transfer consists out of X_COUNT single transfers; a 2D transfer consists out of X_COUNT*Y_COUNT single transfers). Preemption will NOT affect the pending status of channel. As a result, after completion of a higher priority activated channel, the current channel may be reactivated.

PREEMPTABLE_W
Field PREEMPTABLE writer - Specifies if the channel is preemptable. ‘0’: Not preemptable. ‘1’: Preemptable. This field allows higher priority pending channels (from a higher priority group; i.e. an active channel can NOT be preempted by a pending channel in the same priority group) to preempt the active channel in between ‘single transfers’ (a 1D transfer consists out of X_COUNT single transfers; a 2D transfer consists out of X_COUNT*Y_COUNT single transfers). Preemption will NOT affect the pending status of channel. As a result, after completion of a higher priority activated channel, the current channel may be reactivated.

PRIO_R
Field PRIO reader - Channel priority: ‘0’: highest priority. ‘1’ ‘2’ ‘3’: lowest priority. Channels with the same priority constitute a priority group. Priority decoding determines the highest priority pending channel. This channel is determined as follows. First, the highest priority group with pending channels is identified. Second, within this priority group, round robin arbitration is applied. Round robin arbitration (within a priority group) gives the highest priority to the lower channel indices (within the priority group).

PRIO_W
Field PRIO writer - Channel priority: ‘0’: highest priority. ‘1’ ‘2’ ‘3’: lowest priority. Channels with the same priority constitute a priority group. Priority decoding determines the highest priority pending channel. This channel is determined as follows. First, the highest priority group with pending channels is identified. Second, within this priority group, round robin arbitration is applied. Round robin arbitration (within a priority group) gives the highest priority to the lower channel indices (within the priority group).

P_R
Field P reader - User/privileged access control: ‘0’: user mode. ‘1’: privileged mode. This field is set with the user/privileged access control of the transaction that writes this register; i.e. the ‘write data’ is ignored and instead the access control is inherited from the write transaction (note the field attributes should be HW:RW, SW:R). All transactions for this channel use the P field for the user/privileged access control (‘hprot[1]’).

P_W
Field P writer - User/privileged access control: ‘0’: user mode. ‘1’: privileged mode. This field is set with the user/privileged access control of the transaction that writes this register; i.e. the ‘write data’ is ignored and instead the access control is inherited from the write transaction (note the field attributes should be HW:RW, SW:R). All transactions for this channel use the P field for the user/privileged access control (‘hprot[1]’).

R
Register CH_CTL reader

W
Register CH_CTL writer