Module psoc6_01_pac::dw0::ch_struct::ch_status

Channel status

Structs

CH_STATUS_SPEC
Channel status

INTR_CAUSE_R
Field INTR_CAUSE reader - Specifies the source of the interrupt cause: ‘0’: No interrupt generated ‘1’: Interrupt based on transfer complettion configuration based on INTR_TYPE ‘2’: Source transfer bus error ‘3’: Destination transfer bus error ‘4’: Source address misalignment ‘5’: Destination address misalignment ‘6’: Current descriptor pointer is null ‘7’: Active channel is disabled ‘8’: Descriptor bus error ‘9’-‘15’: Not used. For error related interrupt causes (INTR_CAUSE is ‘2’, ‘3’, …, ‘8’), the channel is disabled (HW sets CH_CTL.ENABLED to ‘0’).

PENDING_R
Field PENDING reader - Specifies pending DW channels; i.e. enabled channels whose trigger got activated. This field includes all channels that are in the pending state (not scheduled) or active state (scheduled and performing data transfer(s)).

R
Register CH_STATUS reader