Module psoc6_01_pac::dw0::crc_pol_ctl

CRC polynomial control

Structs

CRC_POL_CTL_SPEC
CRC polynomial control

POLYNOMIAL_R
Field POLYNOMIAL reader - CRC polynomial. The polynomial is represented WITHOUT the high order bit (this bit is always assumed ‘1’). The polynomial should be aligned/shifted such that the more significant bits (bit 31 and down) contain the polynomial and the less significant bits (bit 0 and up) contain padding ’0’s. Some frequently used polynomials: - CRC32: POLYNOMIAL is 0x04c11db7 (x^32 + x^26 + x^23 + x^22 + x^16 + x^12 + x^11 + x^10 + x^8 + x^7 + x^5 + x^4 + x^2 + x + 1). - CRC16: POLYNOMIAL is 0x80050000 (x^16 + x^15 + x^2 + 1, shifted by 16 bit positions). - CRC16 CCITT: POLYNOMIAL is 0x10210000 (x^16 + x^12 + x^5 + 1, shifted by 16 bit positions).

POLYNOMIAL_W
Field POLYNOMIAL writer - CRC polynomial. The polynomial is represented WITHOUT the high order bit (this bit is always assumed ‘1’). The polynomial should be aligned/shifted such that the more significant bits (bit 31 and down) contain the polynomial and the less significant bits (bit 0 and up) contain padding ’0’s. Some frequently used polynomials: - CRC32: POLYNOMIAL is 0x04c11db7 (x^32 + x^26 + x^23 + x^22 + x^16 + x^12 + x^11 + x^10 + x^8 + x^7 + x^5 + x^4 + x^2 + x + 1). - CRC16: POLYNOMIAL is 0x80050000 (x^16 + x^15 + x^2 + 1, shifted by 16 bit positions). - CRC16 CCITT: POLYNOMIAL is 0x10210000 (x^16 + x^12 + x^5 + 1, shifted by 16 bit positions).

R
Register CRC_POL_CTL reader

W
Register CRC_POL_CTL writer