Module psoc6_01_pac::efuse::cmd

Command

Structs

BIT_ADDR_R
Field BIT_ADDR reader - Bit address. This field specifies a bit within a Byte.

BIT_ADDR_W
Field BIT_ADDR writer - Bit address. This field specifies a bit within a Byte.

BIT_DATA_R
Field BIT_DATA reader - Bit data. This field specifies the bit value that is to be programmed into the eFUSE macro array. The address of the bit is specified by the BIT_ADDR, BYTE_ADDR, and MACRO_ADDR fields. This bit is a don’t care for the MXS40 Macro.

BIT_DATA_W
Field BIT_DATA writer - Bit data. This field specifies the bit value that is to be programmed into the eFUSE macro array. The address of the bit is specified by the BIT_ADDR, BYTE_ADDR, and MACRO_ADDR fields. This bit is a don’t care for the MXS40 Macro.

BYTE_ADDR_R
Field BYTE_ADDR reader - Byte address. This field specifies a Byte within a eFUSE macro (each macro has 32 B).

BYTE_ADDR_W
Field BYTE_ADDR writer - Byte address. This field specifies a Byte within a eFUSE macro (each macro has 32 B).

CMD_SPEC
Command

MACRO_ADDR_R
Field MACRO_ADDR reader - Macro address. This field specifies an eFUSE macro.

MACRO_ADDR_W
Field MACRO_ADDR writer - Macro address. This field specifies an eFUSE macro.

R
Register CMD reader

START_R
Field START reader - FW sets this field to ‘1’ to start a program operation. HW sets this field to ‘0’ to indicate that the operation has completed. Note: it is good practice to verify the result of a program operation by reading back a programmed eFUSE memory location. Programming can only change an eFUSE memory bit from ‘0’ to ‘1’; i.e. a programming operation is a ‘one-off’ operation for each eFUSE memory bit: once a bit is changed to ‘1’, it can NEVER be changed back to ‘0’ as a hardware fuse is blown. Programming a memory bit to ‘1’ requires blowing a fuse and requires an eFUSE macro operation. Therefore, this programmiong operation takes time (as specified by the SEQ_PROGRAM_CTL reguisters). Programming amemory bit to ‘0’ does not require an eFUSE macro operation (it is the default eFUSE macro state). Therefore, this programming operation is almost instantaneous. Note: during a program operation, a read operation can not be performed. An AHB-Lite read transfer to the eFUSE memory during a program operation results in an AHB-Lite bus error.

START_W
Field START writer - FW sets this field to ‘1’ to start a program operation. HW sets this field to ‘0’ to indicate that the operation has completed. Note: it is good practice to verify the result of a program operation by reading back a programmed eFUSE memory location. Programming can only change an eFUSE memory bit from ‘0’ to ‘1’; i.e. a programming operation is a ‘one-off’ operation for each eFUSE memory bit: once a bit is changed to ‘1’, it can NEVER be changed back to ‘0’ as a hardware fuse is blown. Programming a memory bit to ‘1’ requires blowing a fuse and requires an eFUSE macro operation. Therefore, this programmiong operation takes time (as specified by the SEQ_PROGRAM_CTL reguisters). Programming amemory bit to ‘0’ does not require an eFUSE macro operation (it is the default eFUSE macro state). Therefore, this programming operation is almost instantaneous. Note: during a program operation, a read operation can not be performed. An AHB-Lite read transfer to the eFUSE memory during a program operation results in an AHB-Lite bus error.

W
Register CMD writer