Module psoc6_01_pac::flashc::flash_ctl

Control

Structs

FLASH_CTL_SPEC
Control

MAIN_BANK_MODE_R
Field MAIN_BANK_MODE reader - Specifies bank mode of FLASH macro main array. 0: Single bank mode. 1: Dual bank mode.

MAIN_BANK_MODE_W
Field MAIN_BANK_MODE writer - Specifies bank mode of FLASH macro main array. 0: Single bank mode. 1: Dual bank mode.

MAIN_ECC_EN_R
Field MAIN_ECC_EN reader - Enable ECC checking for FLASH main interface: 0: Disabled. ECC checking/reporting on FLASH main interface is disabled. No correctable or non-correctable faults are reported. 1: Enabled.

MAIN_ECC_EN_W
Field MAIN_ECC_EN writer - Enable ECC checking for FLASH main interface: 0: Disabled. ECC checking/reporting on FLASH main interface is disabled. No correctable or non-correctable faults are reported. 1: Enabled.

MAIN_ECC_INJ_EN_R
Field MAIN_ECC_INJ_EN reader - Enable error injection for FLASH main interface. When’1’, the parity (ECC_CTL.PARITY[7:0]) is used for a load from the ECC_CTL.WORD_ADDR[23:0] word address.

MAIN_ECC_INJ_EN_W
Field MAIN_ECC_INJ_EN writer - Enable error injection for FLASH main interface. When’1’, the parity (ECC_CTL.PARITY[7:0]) is used for a load from the ECC_CTL.WORD_ADDR[23:0] word address.

MAIN_ERR_SILENT_R
Field MAIN_ERR_SILENT reader - Specifies bus transfer behavior for a non-recoverable error on the FLASH macro main interface (either a non-correctable ECC error, a FLASH macro main interface internal error, a FLASH macro main interface memory hole access): 0: Bus transfer has a bus error. 1: Bus transfer does NOT have a bus error; i.e. the error is ‘silent’ In either case, the erroneous FLASH macro data is returned by the bus master interface. The erroneous data is NOT placed in a bus master interface’s cache and/or buffer. This field is ONLY used by CPU bus transfers. Non-CPU bus transfers always have a bus transfer with a bus error, in case of a non-recoverable error. Note: All CPU bus masters have dedicated status registers (CM0_STATUS and CM4_STATUS) to register the occurrence of FLASH macro main interface internal errors (non-correctable ECC errors and memory hole errors are NOT registered). Note: fault reporting can be used to identify the error that occurred: - FLASH macro main interface internal error. - FLASH macro main interface non-recoverable ECC error. - FLASH macro main interface recoverable ECC error. - FLASH macro main interface memory hole error.

MAIN_ERR_SILENT_W
Field MAIN_ERR_SILENT writer - Specifies bus transfer behavior for a non-recoverable error on the FLASH macro main interface (either a non-correctable ECC error, a FLASH macro main interface internal error, a FLASH macro main interface memory hole access): 0: Bus transfer has a bus error. 1: Bus transfer does NOT have a bus error; i.e. the error is ‘silent’ In either case, the erroneous FLASH macro data is returned by the bus master interface. The erroneous data is NOT placed in a bus master interface’s cache and/or buffer. This field is ONLY used by CPU bus transfers. Non-CPU bus transfers always have a bus transfer with a bus error, in case of a non-recoverable error. Note: All CPU bus masters have dedicated status registers (CM0_STATUS and CM4_STATUS) to register the occurrence of FLASH macro main interface internal errors (non-correctable ECC errors and memory hole errors are NOT registered). Note: fault reporting can be used to identify the error that occurred: - FLASH macro main interface internal error. - FLASH macro main interface non-recoverable ECC error. - FLASH macro main interface recoverable ECC error. - FLASH macro main interface memory hole error.

MAIN_MAP_R
Field MAIN_MAP reader - Specifies mapping of FLASH macro main array. 0: Mapping A. 1: Mapping B. This field is only used when MAIN_BANK_MODE is ‘1’ (dual bank mode).

MAIN_MAP_W
Field MAIN_MAP writer - Specifies mapping of FLASH macro main array. 0: Mapping A. 1: Mapping B. This field is only used when MAIN_BANK_MODE is ‘1’ (dual bank mode).

MAIN_WS_R
Field MAIN_WS reader - FLASH macro main interface wait states: ‘0’: 0 wait states. … ‘15’: 15 wait states

MAIN_WS_W
Field MAIN_WS writer - FLASH macro main interface wait states: ‘0’: 0 wait states. … ‘15’: 15 wait states

R
Register FLASH_CTL reader

W
Register FLASH_CTL writer

WORK_BANK_MODE_R
Field WORK_BANK_MODE reader - Specifies bank mode of FLASH macro work array. 0: Single bank mode. 1: Dual bank mode.

WORK_BANK_MODE_W
Field WORK_BANK_MODE writer - Specifies bank mode of FLASH macro work array. 0: Single bank mode. 1: Dual bank mode.

WORK_ECC_EN_R
Field WORK_ECC_EN reader - Enable ECC checking for FLASH work interface: 0: Disabled. ECC checking/reporting on FLASH work interface is disabled. No correctable or non-correctable faults are reported. 1: Enabled.

WORK_ECC_EN_W
Field WORK_ECC_EN writer - Enable ECC checking for FLASH work interface: 0: Disabled. ECC checking/reporting on FLASH work interface is disabled. No correctable or non-correctable faults are reported. 1: Enabled.

WORK_ECC_INJ_EN_R
Field WORK_ECC_INJ_EN reader - Enable error injection for FLASH work interface. When’1’, the parity (ECC_CTL.PARITY[6:0]) is used for a load from the ECC_CTL.WORD_ADDR[23:0] word address.

WORK_ECC_INJ_EN_W
Field WORK_ECC_INJ_EN writer - Enable error injection for FLASH work interface. When’1’, the parity (ECC_CTL.PARITY[6:0]) is used for a load from the ECC_CTL.WORD_ADDR[23:0] word address.

WORK_ERR_SILENT_R
Field WORK_ERR_SILENT reader - Specifies bus transfer behavior for a non-recoverable error on the FLASH macro work interface (either a non-correctable ECC error, a FLASH macro work interface internal error, a FLASH macro work interface memory hole access): 0: Bus transfer has a bus error. 1: Bus transfer does NOT have a bus error; i.e. the error is ‘silent’ In either case, the erroneous FLASH macro data is returned by the bus master interface. The erroneous data is NOT placed in a bus master interface’s cache and/or buffer. This field is ONLY used by CPU bus transfers. Non-CPU bus transfers always have a bus transfer with a bus error, in case of a non-recoverable error. Note: All CPU bus masters have dedicated status registers (CM0_STATUS and CM4_STATUS) to register the occurrence of FLASH macro work interface internal errors (non-correctable ECC errors and memory hole errors are NOT registered). Note: fault reporting can be used to identify the error that occurred: - FLASH macro work interface internal error. - FLASH macro work interface non-recoverable ECC error. - FLASH macro work interface recoverable ECC error. - FLASH macro work interface memory hole error.

WORK_ERR_SILENT_W
Field WORK_ERR_SILENT writer - Specifies bus transfer behavior for a non-recoverable error on the FLASH macro work interface (either a non-correctable ECC error, a FLASH macro work interface internal error, a FLASH macro work interface memory hole access): 0: Bus transfer has a bus error. 1: Bus transfer does NOT have a bus error; i.e. the error is ‘silent’ In either case, the erroneous FLASH macro data is returned by the bus master interface. The erroneous data is NOT placed in a bus master interface’s cache and/or buffer. This field is ONLY used by CPU bus transfers. Non-CPU bus transfers always have a bus transfer with a bus error, in case of a non-recoverable error. Note: All CPU bus masters have dedicated status registers (CM0_STATUS and CM4_STATUS) to register the occurrence of FLASH macro work interface internal errors (non-correctable ECC errors and memory hole errors are NOT registered). Note: fault reporting can be used to identify the error that occurred: - FLASH macro work interface internal error. - FLASH macro work interface non-recoverable ECC error. - FLASH macro work interface recoverable ECC error. - FLASH macro work interface memory hole error.

WORK_MAP_R
Field WORK_MAP reader - Specifies mapping of FLASH macro work array. 0: Mapping A. 1: Mapping B. This field is only used when WORK_BANK_MODE is ‘1’ (dual bank mode).

WORK_MAP_W
Field WORK_MAP writer - Specifies mapping of FLASH macro work array. 0: Mapping A. 1: Mapping B. This field is only used when WORK_BANK_MODE is ‘1’ (dual bank mode).