Struct psoc6_01_pac::flashc::flash_ctl::W

pub struct W(_);

Register FLASH_CTL writer

Implementations

impl W

pub fn main_ws(&mut self) -> MAIN_WS_W<'_>

Bits 0:3 - FLASH macro main interface wait states: ‘0’: 0 wait states. … ‘15’: 15 wait states

pub fn main_map(&mut self) -> MAIN_MAP_W<'_>

Bit 8 - Specifies mapping of FLASH macro main array. 0: Mapping A. 1: Mapping B. This field is only used when MAIN_BANK_MODE is ‘1’ (dual bank mode).

pub fn work_map(&mut self) -> WORK_MAP_W<'_>

Bit 9 - Specifies mapping of FLASH macro work array. 0: Mapping A. 1: Mapping B. This field is only used when WORK_BANK_MODE is ‘1’ (dual bank mode).

pub fn main_bank_mode(&mut self) -> MAIN_BANK_MODE_W<'_>

Bit 12 - Specifies bank mode of FLASH macro main array. 0: Single bank mode. 1: Dual bank mode.

pub fn work_bank_mode(&mut self) -> WORK_BANK_MODE_W<'_>

Bit 13 - Specifies bank mode of FLASH macro work array. 0: Single bank mode. 1: Dual bank mode.

pub fn main_ecc_en(&mut self) -> MAIN_ECC_EN_W<'_>

Bit 16 - Enable ECC checking for FLASH main interface: 0: Disabled. ECC checking/reporting on FLASH main interface is disabled. No correctable or non-correctable faults are reported. 1: Enabled.

pub fn main_ecc_inj_en(&mut self) -> MAIN_ECC_INJ_EN_W<'_>

Bit 17 - Enable error injection for FLASH main interface. When’1’, the parity (ECC_CTL.PARITY[7:0]) is used for a load from the ECC_CTL.WORD_ADDR[23:0] word address.

pub fn main_err_silent(&mut self) -> MAIN_ERR_SILENT_W<'_>

Bit 18 - Specifies bus transfer behavior for a non-recoverable error on the FLASH macro main interface (either a non-correctable ECC error, a FLASH macro main interface internal error, a FLASH macro main interface memory hole access): 0: Bus transfer has a bus error. 1: Bus transfer does NOT have a bus error; i.e. the error is ‘silent’ In either case, the erroneous FLASH macro data is returned by the bus master interface. The erroneous data is NOT placed in a bus master interface’s cache and/or buffer. This field is ONLY used by CPU bus transfers. Non-CPU bus transfers always have a bus transfer with a bus error, in case of a non-recoverable error. Note: All CPU bus masters have dedicated status registers (CM0_STATUS and CM4_STATUS) to register the occurrence of FLASH macro main interface internal errors (non-correctable ECC errors and memory hole errors are NOT registered). Note: fault reporting can be used to identify the error that occurred: - FLASH macro main interface internal error. - FLASH macro main interface non-recoverable ECC error. - FLASH macro main interface recoverable ECC error. - FLASH macro main interface memory hole error.

pub fn work_ecc_en(&mut self) -> WORK_ECC_EN_W<'_>

Bit 20 - Enable ECC checking for FLASH work interface: 0: Disabled. ECC checking/reporting on FLASH work interface is disabled. No correctable or non-correctable faults are reported. 1: Enabled.

pub fn work_ecc_inj_en(&mut self) -> WORK_ECC_INJ_EN_W<'_>

Bit 21 - Enable error injection for FLASH work interface. When’1’, the parity (ECC_CTL.PARITY[6:0]) is used for a load from the ECC_CTL.WORD_ADDR[23:0] word address.

pub fn work_err_silent(&mut self) -> WORK_ERR_SILENT_W<'_>

Bit 22 - Specifies bus transfer behavior for a non-recoverable error on the FLASH macro work interface (either a non-correctable ECC error, a FLASH macro work interface internal error, a FLASH macro work interface memory hole access): 0: Bus transfer has a bus error. 1: Bus transfer does NOT have a bus error; i.e. the error is ‘silent’ In either case, the erroneous FLASH macro data is returned by the bus master interface. The erroneous data is NOT placed in a bus master interface’s cache and/or buffer. This field is ONLY used by CPU bus transfers. Non-CPU bus transfers always have a bus transfer with a bus error, in case of a non-recoverable error. Note: All CPU bus masters have dedicated status registers (CM0_STATUS and CM4_STATUS) to register the occurrence of FLASH macro work interface internal errors (non-correctable ECC errors and memory hole errors are NOT registered). Note: fault reporting can be used to identify the error that occurred: - FLASH macro work interface internal error. - FLASH macro work interface non-recoverable ECC error. - FLASH macro work interface recoverable ECC error. - FLASH macro work interface memory hole error.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.