Module psoc6_01_pac::flashc::fm_ctl::ana_ctl0

Analog control 0

Structs

ANA_CTL0_SPEC
Analog control 0

CSLDAC_R
Field CSLDAC reader - Trimming of common source line DAC.

CSLDAC_W
Field CSLDAC writer - Trimming of common source line DAC.

FLIP_AMUXBUS_AB_R
Field FLIP_AMUXBUS_AB reader - Flips amuxbusa and amuxbusb 0: amuxbusa, amuxbusb 1: amuxbusb, amuxbusb

FLIP_AMUXBUS_AB_W
Field FLIP_AMUXBUS_AB writer - Flips amuxbusa and amuxbusb 0: amuxbusa, amuxbusb 1: amuxbusb, amuxbusb

MDAC_R
Field MDAC reader - Trimming of the output margin Voltage as a function of Vpos and Vneg.

MDAC_W
Field MDAC writer - Trimming of the output margin Voltage as a function of Vpos and Vneg.

NDAC_MIN_R
Field NDAC_MIN reader - NDAC staircase min value

NDAC_MIN_W
Field NDAC_MIN writer - NDAC staircase min value

PDAC_MIN_R
Field PDAC_MIN reader - PDAC staircase min value

PDAC_MIN_W
Field PDAC_MIN writer - PDAC staircase min value

R
Register ANA_CTL0 reader

SCALE_PRG_PEOFF_R
Field SCALE_PRG_PEOFF reader - PROG&PRE_PROG: Scale for R_GRANT_DELAY on PE OFF transition: 00: 0.125uS 01: 1uS 10: 10uS 11: 100uS

SCALE_PRG_PEOFF_W
Field SCALE_PRG_PEOFF writer - PROG&PRE_PROG: Scale for R_GRANT_DELAY on PE OFF transition: 00: 0.125uS 01: 1uS 10: 10uS 11: 100uS

SCALE_PRG_PEON_R
Field SCALE_PRG_PEON reader - PROG&PRE_PROG: Scale for R_GRANT_DELAY on PE On transition: 00: 0.125uS 01: 1uS 10: 10uS 11: 100uS

SCALE_PRG_PEON_W
Field SCALE_PRG_PEON writer - PROG&PRE_PROG: Scale for R_GRANT_DELAY on PE On transition: 00: 0.125uS 01: 1uS 10: 10uS 11: 100uS

SCALE_PRG_SEQ01_R
Field SCALE_PRG_SEQ01 reader - PROG&PRE_PROG: Scale for R_GRANT_DELAY on seq0-seq1 transition: 00: 0.125uS 01: 1uS 10: 10uS 11: 100uS

SCALE_PRG_SEQ01_W
Field SCALE_PRG_SEQ01 writer - PROG&PRE_PROG: Scale for R_GRANT_DELAY on seq0-seq1 transition: 00: 0.125uS 01: 1uS 10: 10uS 11: 100uS

SCALE_PRG_SEQ12_R
Field SCALE_PRG_SEQ12 reader - PROG&PRE_PROG: Scale for R_GRANT_DELAY on seq1-seq2 transition: 00: 0.125uS 01: 1uS 10: 10uS 11: 100uS

SCALE_PRG_SEQ12_W
Field SCALE_PRG_SEQ12 writer - PROG&PRE_PROG: Scale for R_GRANT_DELAY on seq1-seq2 transition: 00: 0.125uS 01: 1uS 10: 10uS 11: 100uS

SCALE_PRG_SEQ23_R
Field SCALE_PRG_SEQ23 reader - PROG&PRE_PROG: Scale for R_GRANT_DELAY on seq2-seq3 transition: 00: 0.125uS 01: 1uS 10: 10uS 11: 100uS

SCALE_PRG_SEQ23_W
Field SCALE_PRG_SEQ23 writer - PROG&PRE_PROG: Scale for R_GRANT_DELAY on seq2-seq3 transition: 00: 0.125uS 01: 1uS 10: 10uS 11: 100uS

SCALE_SEQ30_R
Field SCALE_SEQ30 reader - PROG&PRE_PROG& ERASE: Scale for R_GRANT_DELAY on seq3-seq0 transition: 00: 0.125uS 01: 1uS 10: 10uS 11: 100uS

SCALE_SEQ30_W
Field SCALE_SEQ30 writer - PROG&PRE_PROG& ERASE: Scale for R_GRANT_DELAY on seq3-seq0 transition: 00: 0.125uS 01: 1uS 10: 10uS 11: 100uS

W
Register ANA_CTL0 writer