Module psoc6_01_pac::flashc::fm_ctl::cal_ctl3

Cal control osc trim bits, idac, sdac, itim

Structs

BGHI_EN_HV_R
Field BGHI_EN_HV reader - 0: Normal (Automatic change over from HI to LO) 1: Force enable HI Bandgap When both BGLO_EN_HV and BGHI_EN_HV are HIGH, only BGHI output is used and turbo_hv_n pulse is active

BGHI_EN_HV_W
Field BGHI_EN_HV writer - 0: Normal (Automatic change over from HI to LO) 1: Force enable HI Bandgap When both BGLO_EN_HV and BGHI_EN_HV are HIGH, only BGHI output is used and turbo_hv_n pulse is active

BGLO_EN_HV_R
Field BGLO_EN_HV reader - 0: Normal (Automatic change over from HI to LO) 1: Force enable LO Bandgap

BGLO_EN_HV_W
Field BGLO_EN_HV writer - 0: Normal (Automatic change over from HI to LO) 1: Force enable LO Bandgap

CAL_CTL3_SPEC
Cal control osc trim bits, idac, sdac, itim

CL_ISO_DIS_HV_R
Field CL_ISO_DIS_HV reader - 0: The internal logic controls the CL isolation 1: Forces CL bypass

CL_ISO_DIS_HV_W
Field CL_ISO_DIS_HV writer - 0: The internal logic controls the CL isolation 1: Forces CL bypass

FDIV_TRIM_HV_R
Field FDIV_TRIM_HV reader - FDIV_TRIM_HV[1:0]: Assuming oscillator frequency of 8MHz in standby. Following are the clock frequencies seen by doubler 00: F = 1MHz 01: F = 0.5MHz 10: F = 2MHz 11: F = 4MHz

FDIV_TRIM_HV_W
Field FDIV_TRIM_HV writer - FDIV_TRIM_HV[1:0]: Assuming oscillator frequency of 8MHz in standby. Following are the clock frequencies seen by doubler 00: F = 1MHz 01: F = 0.5MHz 10: F = 2MHz 11: F = 4MHz

IPREF_TC_HV_R
Field IPREF_TC_HV reader - 0: Increases the IPREF Tempco by subtracting ICREF from IPREF - IPREF internal will be 0.5uA 1: Reduces the IPREF Tempco without subtracting ICREF from IPREF - IPREF internal will be 1uA

IPREF_TC_HV_W
Field IPREF_TC_HV writer - 0: Increases the IPREF Tempco by subtracting ICREF from IPREF - IPREF internal will be 0.5uA 1: Reduces the IPREF Tempco without subtracting ICREF from IPREF - IPREF internal will be 1uA

IREF_SEL_HV_R
Field IREF_SEL_HV reader - Current reference: 0: internal current reference 1: external current reference

IREF_SEL_HV_W
Field IREF_SEL_HV writer - Current reference: 0: internal current reference 1: external current reference

LP_ULP_SW_HV_R
Field LP_ULP_SW_HV reader - LP<–>ULP switch for trim signals: 0: LP 1: ULP

LP_ULP_SW_HV_W
Field LP_ULP_SW_HV writer - LP<–>ULP switch for trim signals: 0: LP 1: ULP

OSC_RANGE_TRIM_HV_R
Field OSC_RANGE_TRIM_HV reader - 0: Oscillator High Frequency Range 1: Oscillator Low Frequency range

OSC_RANGE_TRIM_HV_W
Field OSC_RANGE_TRIM_HV writer - 0: Oscillator High Frequency Range 1: Oscillator Low Frequency range

OSC_TRIM_HV_R
Field OSC_TRIM_HV reader - Flash macro pump clock trim control.

OSC_TRIM_HV_W
Field OSC_TRIM_HV writer - Flash macro pump clock trim control.

R
Register CAL_CTL3 reader

REG_ACT_HV_R
Field REG_ACT_HV reader - 0: VBST regulator will operate in active/standby mode based on control signal. 1: Forces the VBST regulator in active mode all the time

REG_ACT_HV_W
Field REG_ACT_HV writer - 0: VBST regulator will operate in active/standby mode based on control signal. 1: Forces the VBST regulator in active mode all the time

R_GRANT_EN_HV_R
Field R_GRANT_EN_HV reader - 0: r_grant handshake disabled, r_grant always 1. 1: r_grand handshake enabled

R_GRANT_EN_HV_W
Field R_GRANT_EN_HV writer - 0: r_grant handshake disabled, r_grant always 1. 1: r_grand handshake enabled

TURBO_PULSEW_HV_R
Field TURBO_PULSEW_HV reader - Turbo pulse width trim (Typical) 00: 40 us 01: 20 us 10: 15 us 11: 8 us

TURBO_PULSEW_HV_W
Field TURBO_PULSEW_HV writer - Turbo pulse width trim (Typical) 00: 40 us 01: 20 us 10: 15 us 11: 8 us

VDDHI_HV_R
Field VDDHI_HV reader - 0: vdd < 2.3V 1: vdd >= 2.3V ‘0’ setting can used for vdd > 2.3V also, but with a current penalty.

VDDHI_HV_W
Field VDDHI_HV writer - 0: vdd < 2.3V 1: vdd >= 2.3V ‘0’ setting can used for vdd > 2.3V also, but with a current penalty.

VPROT_ACT_HV_R
Field VPROT_ACT_HV reader - Forces VPROT in active mode all the time

VPROT_ACT_HV_W
Field VPROT_ACT_HV writer - Forces VPROT in active mode all the time

VREF_SEL_HV_R
Field VREF_SEL_HV reader - Voltage reference: 0: internal bandgap reference 1: external voltage reference

VREF_SEL_HV_W
Field VREF_SEL_HV writer - Voltage reference: 0: internal bandgap reference 1: external voltage reference

W
Register CAL_CTL3 writer