Module psoc6_01_pac::flashc::fm_ctl::cal_ctl4

Cal Control Vlim, SA, fdiv, reg_act

Structs

AUTO_HVPULSE_HV_R
Field AUTO_HVPULSE_HV reader - 0: HV Pulse controlled by FW 1: HV Pulse controlled by Hardware

AUTO_HVPULSE_HV_W
Field AUTO_HVPULSE_HV writer - 0: HV Pulse controlled by FW 1: HV Pulse controlled by Hardware

CAL_CTL4_SPEC
Cal Control Vlim, SA, fdiv, reg_act

FM_READY_DEL_ULP_HV_R
Field FM_READY_DEL_ULP_HV reader - 00: Default : delay 1ns 01: Delayed by 1.5us 10: Delayed by 2.0us 11: Delayed by 2.5us

FM_READY_DEL_ULP_HV_W
Field FM_READY_DEL_ULP_HV writer - 00: Default : delay 1ns 01: Delayed by 1.5us 10: Delayed by 2.0us 11: Delayed by 2.5us

IDAC_ULP_HV_R
Field IDAC_ULP_HV reader - Sets the sense current reference offset value. Refer to trim tables for details.

IDAC_ULP_HV_W
Field IDAC_ULP_HV writer - Sets the sense current reference offset value. Refer to trim tables for details.

ITIM_ULP_HV_R
Field ITIM_ULP_HV reader - Trimming of timing current

ITIM_ULP_HV_W
Field ITIM_ULP_HV writer - Trimming of timing current

R
Register CAL_CTL4 reader

READY_RESTART_N_HV_R
Field READY_RESTART_N_HV reader - Toggle: 1–>0, ready goes low, ready will remain low as long as the bit is low. Toggle the bit back to 1 to activate the ready logic. To be used by API only.

READY_RESTART_N_HV_W
Field READY_RESTART_N_HV writer - Toggle: 1–>0, ready goes low, ready will remain low as long as the bit is low. Toggle the bit back to 1 to activate the ready logic. To be used by API only.

SDAC_ULP_HV_R
Field SDAC_ULP_HV reader - Sets the sense current reference temp slope. Refer to trim tables for details.

SDAC_ULP_HV_W
Field SDAC_ULP_HV writer - Sets the sense current reference temp slope. Refer to trim tables for details.

SPARE451_ULP_HV_R
Field SPARE451_ULP_HV reader - N/A

SPARE451_ULP_HV_W
Field SPARE451_ULP_HV writer - N/A

UGB_EN_HV_R
Field UGB_EN_HV reader - UGB enable in TM control

UGB_EN_HV_W
Field UGB_EN_HV writer - UGB enable in TM control

VBST_S_DIS_HV_R
Field VBST_S_DIS_HV reader - 0: VBST_S voltage for each sector to allow VBST level to be dropped to VCC during Erase in the selected sector, reducing coupling to GBL. 1: VBST_S voltage for each sector stays at VBST level during Erase in the selected sector.

VBST_S_DIS_HV_W
Field VBST_S_DIS_HV writer - 0: VBST_S voltage for each sector to allow VBST level to be dropped to VCC during Erase in the selected sector, reducing coupling to GBL. 1: VBST_S voltage for each sector stays at VBST level during Erase in the selected sector.

VLIM_TRIM_ULP_HV_R
Field VLIM_TRIM_ULP_HV reader - VLIM_TRIM[1:0]: 00: V2 = 650mV 01: V2 = 600mV 10: V2 = 750mV 11: V2 = 700mV

VLIM_TRIM_ULP_HV_W
Field VLIM_TRIM_ULP_HV writer - VLIM_TRIM[1:0]: 00: V2 = 650mV 01: V2 = 600mV 10: V2 = 750mV 11: V2 = 700mV

W
Register CAL_CTL4 writer