Module psoc6_01_pac::flashc::fm_ctl::fm_ctl

Flash macro control

Structs

DAA_MUX_SEL_R
Field DAA_MUX_SEL reader - Direct memory cell access address.

DAA_MUX_SEL_W
Field DAA_MUX_SEL writer - Direct memory cell access address.

FM_CTL_SPEC
Flash macro control

FM_MODE_R
Field FM_MODE reader - Requires (IF_SEL|WR_EN)=1 Flash macro mode selection

FM_MODE_W
Field FM_MODE writer - Requires (IF_SEL|WR_EN)=1 Flash macro mode selection

FM_SEQ_R
Field FM_SEQ reader - Requires (IF_SEL|WR_EN)=1 Flash macro sequence selection

FM_SEQ_W
Field FM_SEQ writer - Requires (IF_SEL|WR_EN)=1 Flash macro sequence selection

IF_SEL_R
Field IF_SEL reader - Interface selection. Specifies the interface that is used for flash memory read operations: 0: R interface is used (default value). In this case, the flash memory address is provided as part of the R signal interface. 1: C interface is used. In this case, the flash memory address is provided by FM_MEM_ADDR (the page address) and by the C interface access offset in the FM_MEM_DATA structure. Note: IF_SEL and WR_EN cannot be changed at the same time

IF_SEL_W
Field IF_SEL writer - Interface selection. Specifies the interface that is used for flash memory read operations: 0: R interface is used (default value). In this case, the flash memory address is provided as part of the R signal interface. 1: C interface is used. In this case, the flash memory address is provided by FM_MEM_ADDR (the page address) and by the C interface access offset in the FM_MEM_DATA structure. Note: IF_SEL and WR_EN cannot be changed at the same time

R
Register FM_CTL reader

W
Register FM_CTL writer

WR_EN_R
Field WR_EN reader - 0: normal mode 1: Fm Write Enable Note: IF_SEL and WR_EN cannot be changed at the same time

WR_EN_W
Field WR_EN writer - 0: normal mode 1: Fm Write Enable Note: IF_SEL and WR_EN cannot be changed at the same time