Module psoc6_01_pac::flashc::fm_ctl::rgrant_delay_prg

R-grant delay for program

Structs

HV_PARAMS_LOADED_R
Field HV_PARAMS_LOADED reader - 0: HV Pulse common params not loaded 1: HV Pulse common params loaded: r-grant delays, r-grant scale, prescaler, timer values for seq1,seq2_pre, seq2_post, seq3

HV_PARAMS_LOADED_W
Field HV_PARAMS_LOADED writer - 0: HV Pulse common params not loaded 1: HV Pulse common params loaded: r-grant delays, r-grant scale, prescaler, timer values for seq1,seq2_pre, seq2_post, seq3

R
Register RGRANT_DELAY_PRG reader

RGRANT_DELAY_CLK_R
Field RGRANT_DELAY_CLK reader - Frequency divider from clk_t to create the 8MHz reference clock for R_grant delay The value of this field is the integer result of ‘clk_t frequency / 8’. Example: for clk_t=100 this field is INT(100/8) =12. This field is updated at runtime with the ’SW_RGRANT_DELAY_CLK ’ value from the HV parameters table

RGRANT_DELAY_CLK_W
Field RGRANT_DELAY_CLK writer - Frequency divider from clk_t to create the 8MHz reference clock for R_grant delay The value of this field is the integer result of ‘clk_t frequency / 8’. Example: for clk_t=100 this field is INT(100/8) =12. This field is updated at runtime with the ’SW_RGRANT_DELAY_CLK ’ value from the HV parameters table

RGRANT_DELAY_PRG_SEQ12_R
Field RGRANT_DELAY_PRG_SEQ12 reader - PROG&PRE_PROG: R-grant blocking delay on seq1-seq2 transition. Scale = ANA_CTL0.SCALE_SEQ12 When = 0 R_GRANT_DELAY control is disabled when IF_SEL=1 R_GRANT_DELAY control is disabled

RGRANT_DELAY_PRG_SEQ12_W
Field RGRANT_DELAY_PRG_SEQ12 writer - PROG&PRE_PROG: R-grant blocking delay on seq1-seq2 transition. Scale = ANA_CTL0.SCALE_SEQ12 When = 0 R_GRANT_DELAY control is disabled when IF_SEL=1 R_GRANT_DELAY control is disabled

RGRANT_DELAY_PRG_SEQ23_R
Field RGRANT_DELAY_PRG_SEQ23 reader - PROG&PRE_PROG: R-grant blocking delay on seq2-seq3 transition. Scale = ANA_CTL0.SCALE_SEQ23 When = 0 R_GRANT_DELAY control is disabled when IF_SEL=1 R_GRANT_DELAY control is disabled

RGRANT_DELAY_PRG_SEQ23_W
Field RGRANT_DELAY_PRG_SEQ23 writer - PROG&PRE_PROG: R-grant blocking delay on seq2-seq3 transition. Scale = ANA_CTL0.SCALE_SEQ23 When = 0 R_GRANT_DELAY control is disabled when IF_SEL=1 R_GRANT_DELAY control is disabled

RGRANT_DELAY_PRG_SPEC
R-grant delay for program

RGRANT_DELAY_SEQ30_R
Field RGRANT_DELAY_SEQ30 reader - PROG&PRE_PROG & ERASE: R-grant blocking delay on seq3-seq0 transition. Scale = ANA_CTL0.SCALE_SEQ30 When = 0 R_GRANT_DELAY control is disabled when IF_SEL=1 R_GRANT_DELAY control is disabled

RGRANT_DELAY_SEQ30_W
Field RGRANT_DELAY_SEQ30 writer - PROG&PRE_PROG & ERASE: R-grant blocking delay on seq3-seq0 transition. Scale = ANA_CTL0.SCALE_SEQ30 When = 0 R_GRANT_DELAY control is disabled when IF_SEL=1 R_GRANT_DELAY control is disabled

W
Register RGRANT_DELAY_PRG writer