Struct psoc6_01_pac::flashc::fm_ctl::status::R

pub struct R(_);

Register STATUS reader

Implementations

impl R

pub fn timer_enabled(&self) -> TIMER_ENABLED_R

Bit 0 - This is the timer_en bit set by writing a ‘1’ in the TIMER_CTL bit 31. It is reset by HW when the timer expires 0: timer not running 1: Timer is enabled and not expired yet

pub fn hv_regs_isolated(&self) -> HV_REGS_ISOLATED_R

Bit 1 - Indicates the isolation status at HV trim and redundancy registers inputs 0: Not isolated, writing permitted 1: isolated writing disabled

pub fn illegal_hvop(&self) -> ILLEGAL_HVOP_R

Bit 2 - Indicates a bulk, sector erase, program has been requested when axa=1 0: no error 1: illegal HV operation error

pub fn turbo_n(&self) -> TURBO_N_R

Bit 3 - After FM power up indicates the analog blocks currents are boosted to faster reach their functional state.. Used in the testchip boot only as an ‘FM READY’ flag. 0: turbo mode 1: normal mode

pub fn wr_en_mon(&self) -> WR_EN_MON_R

Bit 4 - FM_CTL.WR_EN bit after being synchronized in clk_r domain

pub fn if_sel_mon(&self) -> IF_SEL_MON_R

Bit 5 - FM_CTL.IF_SEL bit after being synchronized in clk_r domain

pub fn timer_status(&self) -> TIMER_STATUS_R

Bit 6 - The actual timer state sync-ed in clk_c domain: 0: timer is not running: 1: timer is running;

pub fn r_grant_delay_status(&self) -> R_GRANT_DELAY_STATUS_R

Bit 7 - 0: R_GRANT_DELAY timer is not running 1: R_GRANT_DELAY timer is running

pub fn fm_busy(&self) -> FM_BUSY_R

Bit 8 - 0’: FM not busy 1: FM BUSY : R_GRANT is 0 as result of a busy request from FM ready, or from HV operations.

pub fn fm_ready(&self) -> FM_READY_R

Bit 9 - 0: FM not ready 1: FM ready

pub fn pos_pump_vlo(&self) -> POS_PUMP_VLO_R

Bit 10 - POS pump VLO

pub fn neg_pump_vhi(&self) -> NEG_PUMP_VHI_R

Bit 11 - NEG pump VHI

pub fn rww(&self) -> RWW_R

Bit 12 - FM Type (Read While Write or Not Read While Write): 0: Non RWW FM Type 1: RWW FM Type

pub fn max_dout_width(&self) -> MAX_DOUT_WIDTH_R

Bit 13 - Internal memory core max data out size (number of data out bits per column): 0: x128 bits 1: x256 bits

pub fn sector0_sr(&self) -> SECTOR0_SR_R

Bit 14 - 0: Sector 0 does not contain special rows. The special rows are located in separate special sectors. 1: Sector 0 contains special rows

pub fn reset_mm(&self) -> RESET_MM_R

Bit 15 - Test_only, internal node: mpcon reset_mm

pub fn row_odd(&self) -> ROW_ODD_R

Bit 16 - Test_only, internal node: mpcon row_odd

pub fn row_even(&self) -> ROW_EVEN_R

Bit 17 - Test_only, internal node: mpcon row_even

pub fn hvop_sub_sector_n(&self) -> HVOP_SUB_SECTOR_N_R

Bit 18 - Test_only, internal node: mpcon bk_subb

pub fn hvop_sector(&self) -> HVOP_SECTOR_R

Bit 19 - Test_only, internal node: mpcon bk_sec

pub fn hvop_bulk_all(&self) -> HVOP_BULK_ALL_R

Bit 20 - Test_only, internal node: mpcon bk_all

pub fn cbus_ra_match(&self) -> CBUS_RA_MATCH_R

Bit 21 - Test_only, internal node: mpcon ra match

pub fn cbus_red_row_en(&self) -> CBUS_RED_ROW_EN_R

Bit 22 - Test_only, internal node: mpcon red_row_en

pub fn rq_error(&self) -> RQ_ERROR_R

Bit 23 - Test_only, internal node: rq_error sync-de in clk_c domain

pub fn pump_pdac(&self) -> PUMP_PDAC_R

Bits 24:27 - Test_only, internal node: regif pdac outputs to pos pump

pub fn pump_ndac(&self) -> PUMP_NDAC_R

Bits 28:31 - Test_only, internal node: regif ndac outputs to pos pump