Module psoc6_01_pac::flashc::fm_ctl::timer_clk_ctl

Timer prescaler (clk_t to timer clock frequency divider)

Structs

R
Register TIMER_CLK_CTL reader

RGRANT_DELAY_PRG_PEOFF_R
Field RGRANT_DELAY_PRG_PEOFF reader - PROG&PRE_PROG: R-grant blocking delay on PE OFF. Scale = ANA_CTL0.SCALE_PEOFF When = 0 R_GRANT_DELAY control is disabled when IF_SEL=1 R_GRANT_DELAY control is disabled

RGRANT_DELAY_PRG_PEOFF_W
Field RGRANT_DELAY_PRG_PEOFF writer - PROG&PRE_PROG: R-grant blocking delay on PE OFF. Scale = ANA_CTL0.SCALE_PEOFF When = 0 R_GRANT_DELAY control is disabled when IF_SEL=1 R_GRANT_DELAY control is disabled

RGRANT_DELAY_PRG_PEON_R
Field RGRANT_DELAY_PRG_PEON reader - PROG&PRE_PROG: R-grant blocking delay on PE ON. Scale = ANA_CTL0.SCALE_PEON When = 0 R_GRANT_DELAY control is disabled when IF_SEL=1 R_GRANT_DELAY control is disabled

RGRANT_DELAY_PRG_PEON_W
Field RGRANT_DELAY_PRG_PEON writer - PROG&PRE_PROG: R-grant blocking delay on PE ON. Scale = ANA_CTL0.SCALE_PEON When = 0 R_GRANT_DELAY control is disabled when IF_SEL=1 R_GRANT_DELAY control is disabled

RGRANT_DELAY_PRG_SEQ01_R
Field RGRANT_DELAY_PRG_SEQ01 reader - PROG&PRE_PROG: R-grant blocking delay on seq0-seq1 transition. Scale = ANA_CTL0.SCALE_SEQ01 When = 0 R_GRANT_DELAY control is disabled when IF_SEL=1 R_GRANT_DELAY control is disabled

RGRANT_DELAY_PRG_SEQ01_W
Field RGRANT_DELAY_PRG_SEQ01 writer - PROG&PRE_PROG: R-grant blocking delay on seq0-seq1 transition. Scale = ANA_CTL0.SCALE_SEQ01 When = 0 R_GRANT_DELAY control is disabled when IF_SEL=1 R_GRANT_DELAY control is disabled

TIMER_CLK_CTL_SPEC
Timer prescaler (clk_t to timer clock frequency divider)

TIMER_CLOCK_FREQ_R
Field TIMER_CLOCK_FREQ reader - Clk_t frequency divider to provide the 1MHz reference clock for the Regif Timer. Equal to the frequency in MHz of the timer clock ‘clk_t’. Example: if ‘clk_t’ has a frequency of 4 MHz then this field value is ‘4’ Max clk_t frequency = 100MHz. This field is updated at runtime with the ’SW_TIMER_CLOCK_FREQ ’ value from the HV parameters table

TIMER_CLOCK_FREQ_W
Field TIMER_CLOCK_FREQ writer - Clk_t frequency divider to provide the 1MHz reference clock for the Regif Timer. Equal to the frequency in MHz of the timer clock ‘clk_t’. Example: if ‘clk_t’ has a frequency of 4 MHz then this field value is ‘4’ Max clk_t frequency = 100MHz. This field is updated at runtime with the ’SW_TIMER_CLOCK_FREQ ’ value from the HV parameters table

W
Register TIMER_CLK_CTL writer