Module psoc6_01_pac::flashc::fm_ctl::timer_ctl

Timer control

Structs

ACLK_EN_R
Field ACLK_EN reader - ACLK enable (generates a single cycle pulse for the FM): 0: disabled 1: enabled. SW set this field to ‘1’ to generate a single cycle pulse. HW sets this field to ‘0’ when the pulse is generated.

ACLK_EN_W
Field ACLK_EN writer - ACLK enable (generates a single cycle pulse for the FM): 0: disabled 1: enabled. SW set this field to ‘1’ to generate a single cycle pulse. HW sets this field to ‘0’ when the pulse is generated.

AUTO_SEQUENCE_R
Field AUTO_SEQUENCE reader - 1’: Starts1 the HV automatic sequencing Cleared by HW

AUTO_SEQUENCE_W
Field AUTO_SEQUENCE writer - 1’: Starts1 the HV automatic sequencing Cleared by HW

PERIOD_R
Field PERIOD reader - Timer period in either microseconds (SCALE is ‘0’) or 100’s of microseconds (SCALE is ‘1’) multiples.

PERIOD_W
Field PERIOD writer - Timer period in either microseconds (SCALE is ‘0’) or 100’s of microseconds (SCALE is ‘1’) multiples.

PRE_PROG_CSL_R
Field PRE_PROG_CSL reader - 0: CSL lines driven by CSL_DAC 1: CSL lines driven by VNEG_G

PRE_PROG_CSL_W
Field PRE_PROG_CSL writer - 0: CSL lines driven by CSL_DAC 1: CSL lines driven by VNEG_G

PRE_PROG_R
Field PRE_PROG reader - 1 during pre-program operation

PRE_PROG_W
Field PRE_PROG writer - 1 during pre-program operation

PUMP_EN_R
Field PUMP_EN reader - Pump enable: 0: disabled 1: enabled (also requires FM_CTL.IF_SEL to be’1’, this additional restriction is required to prevent non intentional clearing of the FM). SW sets this field to ‘1’ to generate a single PE pulse. HW clears this field when timer is expired.

PUMP_EN_W
Field PUMP_EN writer - Pump enable: 0: disabled 1: enabled (also requires FM_CTL.IF_SEL to be’1’, this additional restriction is required to prevent non intentional clearing of the FM). SW sets this field to ‘1’ to generate a single PE pulse. HW clears this field when timer is expired.

R
Register TIMER_CTL reader

SCALE_R
Field SCALE reader - Timer tick scale: 0: 1 microsecond. 1: 100 microseconds.

SCALE_W
Field SCALE writer - Timer tick scale: 0: 1 microsecond. 1: 100 microseconds.

TIMER_CTL_SPEC
Timer control

TIMER_EN_R
Field TIMER_EN reader - Timer enable: 0: disabled 1: enabled. SW sets this field to ‘1’ to start the timer. HW sets this field to ‘0’ when the timer is expired.

TIMER_EN_W
Field TIMER_EN writer - Timer enable: 0: disabled 1: enabled. SW sets this field to ‘1’ to start the timer. HW sets this field to ‘0’ when the timer is expired.

W
Register TIMER_CTL writer