Module psoc6_01_pac::flashc::fm_ctl::wait_ctl

Wait State control

Structs

DRMM_R
Field DRMM reader - 0: Normal 1: Test mode to enable Margin mode for 2 rows at a time

DRMM_W
Field DRMM writer - 0: Normal 1: Test mode to enable Margin mode for 2 rows at a time

FM_RWW_MODE_R
Field FM_RWW_MODE reader - 00: Full CBUS MODE 01: RWW 10: RWW. R_GRANT is stalling r_bus for the whole program/erase duration

FM_RWW_MODE_W
Field FM_RWW_MODE writer - 00: Full CBUS MODE 01: RWW 10: RWW. R_GRANT is stalling r_bus for the whole program/erase duration

LV_SPARE_1_R
Field LV_SPARE_1 reader - Spare register

LV_SPARE_1_W
Field LV_SPARE_1 writer - Spare register

MBA_R
Field MBA reader - 0: Normal 1: Test mode to enable Master Bulk Access which allows both normal rows and redundant rows to be erased / programmed in one HV cycle (Bulk / Sector Erase and Sector Program).

MBA_W
Field MBA writer - 0: Normal 1: Test mode to enable Master Bulk Access which allows both normal rows and redundant rows to be erased / programmed in one HV cycle (Bulk / Sector Erase and Sector Program).

PL_SOFT_SET_EN_R
Field PL_SOFT_SET_EN reader - Page latch soft set enable, 0 = disabled, 1 = enabled (at end of seq_2), taken care in API

PL_SOFT_SET_EN_W
Field PL_SOFT_SET_EN writer - Page latch soft set enable, 0 = disabled, 1 = enabled (at end of seq_2), taken care in API

R
Register WAIT_CTL reader

W
Register WAIT_CTL writer

WAIT_CTL_SPEC
Wait State control

WAIT_FM_HV_RD_R
Field WAIT_FM_HV_RD reader - Number of C interface wait cycles (on ‘clk_c’) for a read from the high Voltage page latches. Common for reading HV Page Latches and the DATA_COMP_RESULT bit

WAIT_FM_HV_RD_W
Field WAIT_FM_HV_RD writer - Number of C interface wait cycles (on ‘clk_c’) for a read from the high Voltage page latches. Common for reading HV Page Latches and the DATA_COMP_RESULT bit

WAIT_FM_HV_WR_R
Field WAIT_FM_HV_WR reader - Number of C interface wait cycles (on ‘clk_c’) for a write to the high Voltage page latches.

WAIT_FM_HV_WR_W
Field WAIT_FM_HV_WR writer - Number of C interface wait cycles (on ‘clk_c’) for a write to the high Voltage page latches.

WAIT_FM_MEM_RD_R
Field WAIT_FM_MEM_RD reader - Number of C interface wait cycles (on ‘clk_c’) for a read from the memory

WAIT_FM_MEM_RD_W
Field WAIT_FM_MEM_RD writer - Number of C interface wait cycles (on ‘clk_c’) for a read from the memory