Module psoc6_01_pac::flashc::fm_sram_ecc_ctl3

eCT Flash SRAM ECC control 3

Structs

ECC_ENABLE_R
Field ECC_ENABLE reader - ECC generation/check enable for eCT Flash SRAM memory.

ECC_ENABLE_W
Field ECC_ENABLE writer - ECC generation/check enable for eCT Flash SRAM memory.

ECC_INJ_EN_R
Field ECC_INJ_EN reader - eCT Flash SRAM ECC error injection test enable. Follow the steps below for ECC logic test: 1. Write corrupted or uncorrupted 39-bit data to FM_SRAM_ECC_CTL0/1 registers. 2. Set the ECC_INJ_EN bit to ‘1’. 3. Confirm that the bit ECC_TEST_FAIL is ‘0’. If this is not the case, start over at item 1 because the eCT Flash was not idle. 4. Check the corrected data in FM_SRAM_ECC_CTL2. 5. Confirm that fault was reported to fault structure, and check syndrome (only applicable if corrupted data was written in step 1). 6. If not finished, start over at 1 with different data.

ECC_INJ_EN_W
Field ECC_INJ_EN writer - eCT Flash SRAM ECC error injection test enable. Follow the steps below for ECC logic test: 1. Write corrupted or uncorrupted 39-bit data to FM_SRAM_ECC_CTL0/1 registers. 2. Set the ECC_INJ_EN bit to ‘1’. 3. Confirm that the bit ECC_TEST_FAIL is ‘0’. If this is not the case, start over at item 1 because the eCT Flash was not idle. 4. Check the corrected data in FM_SRAM_ECC_CTL2. 5. Confirm that fault was reported to fault structure, and check syndrome (only applicable if corrupted data was written in step 1). 6. If not finished, start over at 1 with different data.

ECC_TEST_FAIL_R
Field ECC_TEST_FAIL reader - Status of ECC test. 1 : ECC test failed because eCT Flash macro is busy and using the SRAM. 0: ECC was performed.

FM_SRAM_ECC_CTL3_SPEC
eCT Flash SRAM ECC control 3

R
Register FM_SRAM_ECC_CTL3 reader

W
Register FM_SRAM_ECC_CTL3 writer