Module psoc6_01_pac::gpio::prt::cfg

Port configuration register

Structs

CFG_SPEC
Port configuration register

DRIVE_MODE0_R
Field DRIVE_MODE0 reader - The GPIO drive mode for IO pin 0. Resistive pull-up and pull-down is selected in the drive mode. Note: when initializing IO’s that are connected to a live bus (such as I2C), make sure the peripheral and HSIOM (HSIOM_PRT_SELx) is properly configured before turning the IO on here to avoid producing glitches on the bus. Note: that peripherals other than GPIO & UDB/DSI directly control both the output and output-enable of the output buffer (peripherals can drive strong 0 or strong 1 in any mode except OFF=‘0’). Note: D_OUT, D_OUT_EN are pins of GPIO cell.

DRIVE_MODE0_W
Field DRIVE_MODE0 writer - The GPIO drive mode for IO pin 0. Resistive pull-up and pull-down is selected in the drive mode. Note: when initializing IO’s that are connected to a live bus (such as I2C), make sure the peripheral and HSIOM (HSIOM_PRT_SELx) is properly configured before turning the IO on here to avoid producing glitches on the bus. Note: that peripherals other than GPIO & UDB/DSI directly control both the output and output-enable of the output buffer (peripherals can drive strong 0 or strong 1 in any mode except OFF=‘0’). Note: D_OUT, D_OUT_EN are pins of GPIO cell.

DRIVE_MODE1_R
Field DRIVE_MODE1 reader - The GPIO drive mode for IO pin 1

DRIVE_MODE1_W
Field DRIVE_MODE1 writer - The GPIO drive mode for IO pin 1

DRIVE_MODE2_R
Field DRIVE_MODE2 reader - The GPIO drive mode for IO pin 2

DRIVE_MODE2_W
Field DRIVE_MODE2 writer - The GPIO drive mode for IO pin 2

DRIVE_MODE3_R
Field DRIVE_MODE3 reader - The GPIO drive mode for IO pin 3

DRIVE_MODE3_W
Field DRIVE_MODE3 writer - The GPIO drive mode for IO pin 3

DRIVE_MODE4_R
Field DRIVE_MODE4 reader - The GPIO drive mode for IO pin4

DRIVE_MODE4_W
Field DRIVE_MODE4 writer - The GPIO drive mode for IO pin4

DRIVE_MODE5_R
Field DRIVE_MODE5 reader - The GPIO drive mode for IO pin 5

DRIVE_MODE5_W
Field DRIVE_MODE5 writer - The GPIO drive mode for IO pin 5

DRIVE_MODE6_R
Field DRIVE_MODE6 reader - The GPIO drive mode for IO pin 6

DRIVE_MODE6_W
Field DRIVE_MODE6 writer - The GPIO drive mode for IO pin 6

DRIVE_MODE7_R
Field DRIVE_MODE7 reader - The GPIO drive mode for IO pin 7

DRIVE_MODE7_W
Field DRIVE_MODE7 writer - The GPIO drive mode for IO pin 7

IN_EN0_R
Field IN_EN0 reader - Enables the input buffer for IO pin 0. This bit should be cleared when analog signals are present on the pin to avoid crowbar currents. The output buffer can be used to drive analog signals high or low without issue. ‘0’: Input buffer disabled ‘1’: Input buffer enabled

IN_EN0_W
Field IN_EN0 writer - Enables the input buffer for IO pin 0. This bit should be cleared when analog signals are present on the pin to avoid crowbar currents. The output buffer can be used to drive analog signals high or low without issue. ‘0’: Input buffer disabled ‘1’: Input buffer enabled

IN_EN1_R
Field IN_EN1 reader - Enables the input buffer for IO pin 1

IN_EN1_W
Field IN_EN1 writer - Enables the input buffer for IO pin 1

IN_EN2_R
Field IN_EN2 reader - Enables the input buffer for IO pin 2

IN_EN2_W
Field IN_EN2 writer - Enables the input buffer for IO pin 2

IN_EN3_R
Field IN_EN3 reader - Enables the input buffer for IO pin 3

IN_EN3_W
Field IN_EN3 writer - Enables the input buffer for IO pin 3

IN_EN4_R
Field IN_EN4 reader - Enables the input buffer for IO pin 4

IN_EN4_W
Field IN_EN4 writer - Enables the input buffer for IO pin 4

IN_EN5_R
Field IN_EN5 reader - Enables the input buffer for IO pin 5

IN_EN5_W
Field IN_EN5 writer - Enables the input buffer for IO pin 5

IN_EN6_R
Field IN_EN6 reader - Enables the input buffer for IO pin 6

IN_EN6_W
Field IN_EN6 writer - Enables the input buffer for IO pin 6

IN_EN7_R
Field IN_EN7 reader - Enables the input buffer for IO pin 7

IN_EN7_W
Field IN_EN7 writer - Enables the input buffer for IO pin 7

R
Register CFG reader

W
Register CFG writer

Enums

DRIVE_MODE0_A
The GPIO drive mode for IO pin 0. Resistive pull-up and pull-down is selected in the drive mode. Note: when initializing IO’s that are connected to a live bus (such as I2C), make sure the peripheral and HSIOM (HSIOM_PRT_SELx) is properly configured before turning the IO on here to avoid producing glitches on the bus. Note: that peripherals other than GPIO & UDB/DSI directly control both the output and output-enable of the output buffer (peripherals can drive strong 0 or strong 1 in any mode except OFF=‘0’). Note: D_OUT, D_OUT_EN are pins of GPIO cell.