Enum psoc6_01_pac::gpio::prt::cfg::DRIVE_MODE0_A

#[repr(u8)]
pub enum DRIVE_MODE0_A {
    HIGHZ,
    RSVD,
    PULLUP,
    PULLDOWN,
    OD_DRIVESLOW,
    OD_DRIVESHIGH,
    STRONG,
    PULLUP_DOWN,
}

The GPIO drive mode for IO pin 0. Resistive pull-up and pull-down is selected in the drive mode. Note: when initializing IO’s that are connected to a live bus (such as I2C), make sure the peripheral and HSIOM (HSIOM_PRT_SELx) is properly configured before turning the IO on here to avoid producing glitches on the bus. Note: that peripherals other than GPIO & UDB/DSI directly control both the output and output-enable of the output buffer (peripherals can drive strong 0 or strong 1 in any mode except OFF=‘0’). Note: D_OUT, D_OUT_EN are pins of GPIO cell.

Value on reset: 0

Variants

HIGHZ

0: Output buffer is off creating a high impedance input D_OUT = ‘0’: High Impedance D_OUT = ‘1’: High Impedance

RSVD

1: N/A

PULLUP

2: Resistive pull up

For GPIO & UDB/DSI peripherals: When D_OUT_EN = 1: D_OUT = ‘0’: Strong pull down D_OUT = ‘1’: Weak/resistive pull up When D_OUT_EN = 0: D_OUT = ‘0’: High impedance D_OUT = ‘1’: High impedance

For peripherals other than GPIO & UDB/DSI: When D_OUT_EN = 1: D_OUT = ‘0’: Strong pull down D_OUT = ‘1’: Strong pull up When D_OUT_EN = 0: D_OUT = ‘0’: Weak/resistive pull up D_OUT = ‘1’: Weak/resistive pull up

PULLDOWN

3: Resistive pull down

For GPIO & UDB/DSI peripherals: When D_OUT_EN = 1: D_OUT = ‘0’: Weak/resistive pull down D_OUT = ‘1’: Strong pull up When D_OUT_EN = 0: D_OUT = ‘0’: High impedance D_OUT = ‘1’: High impedance

For peripherals other than GPIO & UDB/DSI: When D_OUT_EN = 1: D_OUT = ‘0’: Strong pull down D_OUT = ‘1’: Strong pull up When D_OUT_EN = 0: D_OUT = ‘0’: Weak/resistive pull down D_OUT = ‘1’: Weak/resistive pull down

OD_DRIVESLOW

4: Open drain, drives low

For GPIO & UDB/DSI peripherals: When D_OUT_EN = 1: D_OUT = ‘0’: Strong pull down D_OUT = ‘1’: High Impedance When D_OUT_EN = 0: D_OUT = ‘0’: High impedance D_OUT = ‘1’: High impedance

For peripherals other than GPIO & UDB/DSI: When D_OUT_EN = 1: D_OUT = ‘0’: Strong pull down D_OUT = ‘1’: Strong pull up When D_OUT_EN = 0: D_OUT = ‘0’: High Impedance D_OUT = ‘1’: High Impedance

OD_DRIVESHIGH

5: Open drain, drives high

For GPIO & UDB/DSI peripherals: When D_OUT_EN = 1: D_OUT = ‘0’: High Impedance D_OUT = ‘1’: Strong pull up When D_OUT_EN = 0: D_OUT = ‘0’: High impedance D_OUT = ‘1’: High impedance

For peripherals other than GPIO & UDB/DSI: When D_OUT_EN = 1: D_OUT = ‘0’: Strong pull down D_OUT = ‘1’: Strong pull up When D_OUT_EN = 0: D_OUT = ‘0’: High Impedance D_OUT = ‘1’: High Impedance

STRONG

6: Strong D_OUTput buffer

For GPIO & UDB/DSI peripherals: When D_OUT_EN = 1: D_OUT = ‘0’: Strong pull down D_OUT = ‘1’: Strong pull up When D_OUT_EN = 0: D_OUT = ‘0’: High impedance D_OUT = ‘1’: High impedance

For peripherals other than GPIO & UDB/DSI: When D_OUT_EN = 1: D_OUT = ‘0’: Strong pull down D_OUT = ‘1’: Strong pull up When D_OUT_EN = 0: D_OUT = ‘0’: High Impedance D_OUT = ‘1’: High Impedance

PULLUP_DOWN

7: Pull up or pull down

For GPIO & UDB/DSI peripherals: When D_OUT_EN = ‘0’: GPIO_DSI_OUT = ‘0’: Weak/resistive pull down GPIO_DSI_OUT = ‘1’: Weak/resistive pull up where ‘GPIO_DSI_OUT’ is a function of PORT_SEL, OUT & DSI_DATA_OUT.

For peripherals other than GPIO & UDB/DSI: When D_OUT_EN = 1: D_OUT = ‘0’: Strong pull down D_OUT = ‘1’: Strong pull up When D_OUT_EN = 0: D_OUT = ‘0’: Weak/resistive pull down D_OUT = ‘1’: Weak/resistive pull up