Module psoc6_01_pac::gpio::prt::cfg_sio

Port SIO configuration register

Structs

CFG_SIO_SPEC
Port SIO configuration register

IBUF_SEL01_R
Field IBUF_SEL01 reader - Selects the input buffer mode: 0: Singled ended input buffer 1: Differential input buffer

IBUF_SEL01_W
Field IBUF_SEL01 writer - Selects the input buffer mode: 0: Singled ended input buffer 1: Differential input buffer

IBUF_SEL23_R
Field IBUF_SEL23 reader - See corresponding definition for IO pins 0 and 1

IBUF_SEL23_W
Field IBUF_SEL23 writer - See corresponding definition for IO pins 0 and 1

IBUF_SEL45_R
Field IBUF_SEL45 reader - See corresponding definition for IO pins 0 and 1

IBUF_SEL45_W
Field IBUF_SEL45 writer - See corresponding definition for IO pins 0 and 1

IBUF_SEL67_R
Field IBUF_SEL67 reader - See corresponding definition for IO pins 0 and 1

IBUF_SEL67_W
Field IBUF_SEL67 writer - See corresponding definition for IO pins 0 and 1

R
Register CFG_SIO reader

VOH_SEL01_R
Field VOH_SEL01 reader - Selects the regulated Voh output level and trip point of the input buffer for a specific SIO pin pair. Voh depends on the selected reference voltage (VREF_SEL). ‘0’: Voh = 1reference; e.g. reference at 1.2V -> Voh = 1.2V ‘1’: Voh = 1.25reference; e.g. reference at 1.2V -> Voh = 1.5V ‘2’: Voh = 1.49reference; e.g. reference at 1.2V -> Voh = ~1.8V ‘3’: Voh = 1.67reference; e.g. reference at 1.2V -> Voh = 2V ‘4’: Voh = 2.08reference; e.g. reference at 1.2V -> Voh = 2.5V ‘5’: Voh = 2.5reference; e.g. reference at 1.2V -> Voh = 3V ‘6’: Voh = 2.78reference; e.g. reference at 1.2V -> Voh = ~3.3V ‘7’: Voh = 4.16reference; e.g. reference at 1.2V -> Voh = 5.0V Note: The upper value on Voh is limited to Vddio - 400mV

VOH_SEL01_W
Field VOH_SEL01 writer - Selects the regulated Voh output level and trip point of the input buffer for a specific SIO pin pair. Voh depends on the selected reference voltage (VREF_SEL). ‘0’: Voh = 1reference; e.g. reference at 1.2V -> Voh = 1.2V ‘1’: Voh = 1.25reference; e.g. reference at 1.2V -> Voh = 1.5V ‘2’: Voh = 1.49reference; e.g. reference at 1.2V -> Voh = ~1.8V ‘3’: Voh = 1.67reference; e.g. reference at 1.2V -> Voh = 2V ‘4’: Voh = 2.08reference; e.g. reference at 1.2V -> Voh = 2.5V ‘5’: Voh = 2.5reference; e.g. reference at 1.2V -> Voh = 3V ‘6’: Voh = 2.78reference; e.g. reference at 1.2V -> Voh = ~3.3V ‘7’: Voh = 4.16reference; e.g. reference at 1.2V -> Voh = 5.0V Note: The upper value on Voh is limited to Vddio - 400mV

VOH_SEL23_R
Field VOH_SEL23 reader - See corresponding definition for IO pins 0 and 1

VOH_SEL23_W
Field VOH_SEL23 writer - See corresponding definition for IO pins 0 and 1

VOH_SEL45_R
Field VOH_SEL45 reader - See corresponding definition for IO pins 0 and 1

VOH_SEL45_W
Field VOH_SEL45 writer - See corresponding definition for IO pins 0 and 1

VOH_SEL67_R
Field VOH_SEL67 reader - See corresponding definition for IO pins 0 and 1

VOH_SEL67_W
Field VOH_SEL67 writer - See corresponding definition for IO pins 0 and 1

VREF_SEL01_R
Field VREF_SEL01 reader - Selects reference voltage (Vref) trip-point of the input buffer: ‘0’: Trip-point reference from pin_ref ‘1’: Trip-point reference of SRSS internal reference Vref (1.2 V) ‘2’: Trip-point reference of AMUXBUS_A ‘3’: Trip-point reference of AMUXBUS_B

VREF_SEL01_W
Field VREF_SEL01 writer - Selects reference voltage (Vref) trip-point of the input buffer: ‘0’: Trip-point reference from pin_ref ‘1’: Trip-point reference of SRSS internal reference Vref (1.2 V) ‘2’: Trip-point reference of AMUXBUS_A ‘3’: Trip-point reference of AMUXBUS_B

VREF_SEL23_R
Field VREF_SEL23 reader - See corresponding definition for IO pins 0 and 1

VREF_SEL23_W
Field VREF_SEL23 writer - See corresponding definition for IO pins 0 and 1

VREF_SEL45_R
Field VREF_SEL45 reader - See corresponding definition for IO pins 0 and 1

VREF_SEL45_W
Field VREF_SEL45 writer - See corresponding definition for IO pins 0 and 1

VREF_SEL67_R
Field VREF_SEL67 reader - See corresponding definition for IO pins 0 and 1

VREF_SEL67_W
Field VREF_SEL67 writer - See corresponding definition for IO pins 0 and 1

VREG_EN01_R
Field VREG_EN01 reader - Selects the output buffer mode: ‘0’: Unregulated output buffer ‘1’: Regulated output buffer The regulated output mode is selected ONLY if the CFG.DRIVE_MODE bits are set to the strong pull up (Z_1 = ‘5’) mode. If the CFG.DRIVE_MODE bits are set to any other mode the regulated output buffer will be disabled and the standard CMOS output buffer is used.

VREG_EN01_W
Field VREG_EN01 writer - Selects the output buffer mode: ‘0’: Unregulated output buffer ‘1’: Regulated output buffer The regulated output mode is selected ONLY if the CFG.DRIVE_MODE bits are set to the strong pull up (Z_1 = ‘5’) mode. If the CFG.DRIVE_MODE bits are set to any other mode the regulated output buffer will be disabled and the standard CMOS output buffer is used.

VREG_EN23_R
Field VREG_EN23 reader - See corresponding definition for IO pins 0 and 1

VREG_EN23_W
Field VREG_EN23 writer - See corresponding definition for IO pins 0 and 1

VREG_EN45_R
Field VREG_EN45 reader - See corresponding definition for IO pins 0 and 1

VREG_EN45_W
Field VREG_EN45 writer - See corresponding definition for IO pins 0 and 1

VREG_EN67_R
Field VREG_EN67 reader - See corresponding definition for IO pins 0 and 1

VREG_EN67_W
Field VREG_EN67 writer - See corresponding definition for IO pins 0 and 1

VTRIP_SEL01_R
Field VTRIP_SEL01 reader - Selects the input buffer trip-point in single ended input buffer mode (IBUF_SEL = ‘0’): ‘0’: Input buffer functions as a CMOS input buffer. ‘1’: Input buffer functions as a TTL input buffer. In differential input buffer mode (IBUF_SEL = ‘1’) ‘0’: Trip-point is 0.5Vddio or 0.5Voh (depends on VREF_SEL/VOH_SEL) ‘1’: Trip-point is 0.4Vddio or 1.0Vref (depends on VREF_SEL)

VTRIP_SEL01_W
Field VTRIP_SEL01 writer - Selects the input buffer trip-point in single ended input buffer mode (IBUF_SEL = ‘0’): ‘0’: Input buffer functions as a CMOS input buffer. ‘1’: Input buffer functions as a TTL input buffer. In differential input buffer mode (IBUF_SEL = ‘1’) ‘0’: Trip-point is 0.5Vddio or 0.5Voh (depends on VREF_SEL/VOH_SEL) ‘1’: Trip-point is 0.4Vddio or 1.0Vref (depends on VREF_SEL)

VTRIP_SEL23_R
Field VTRIP_SEL23 reader - See corresponding definition for IO pins 0 and 1

VTRIP_SEL23_W
Field VTRIP_SEL23 writer - See corresponding definition for IO pins 0 and 1

VTRIP_SEL45_R
Field VTRIP_SEL45 reader - See corresponding definition for IO pins 0 and 1

VTRIP_SEL45_W
Field VTRIP_SEL45 writer - See corresponding definition for IO pins 0 and 1

VTRIP_SEL67_R
Field VTRIP_SEL67 reader - See corresponding definition for IO pins 0 and 1

VTRIP_SEL67_W
Field VTRIP_SEL67 writer - See corresponding definition for IO pins 0 and 1

W
Register CFG_SIO writer