Struct psoc6_01_pac::gpio::prt::cfg_sio::R

pub struct R(_);

Register CFG_SIO reader

Implementations

impl R

pub fn vreg_en01(&self) -> VREG_EN01_R

Bit 0 - Selects the output buffer mode: ‘0’: Unregulated output buffer ‘1’: Regulated output buffer The regulated output mode is selected ONLY if the CFG.DRIVE_MODE bits are set to the strong pull up (Z_1 = ‘5’) mode. If the CFG.DRIVE_MODE bits are set to any other mode the regulated output buffer will be disabled and the standard CMOS output buffer is used.

pub fn ibuf_sel01(&self) -> IBUF_SEL01_R

Bit 1 - Selects the input buffer mode: 0: Singled ended input buffer 1: Differential input buffer

pub fn vtrip_sel01(&self) -> VTRIP_SEL01_R

Bit 2 - Selects the input buffer trip-point in single ended input buffer mode (IBUF_SEL = ‘0’): ‘0’: Input buffer functions as a CMOS input buffer. ‘1’: Input buffer functions as a TTL input buffer. In differential input buffer mode (IBUF_SEL = ‘1’) ‘0’: Trip-point is 0.5Vddio or 0.5Voh (depends on VREF_SEL/VOH_SEL) ‘1’: Trip-point is 0.4Vddio or 1.0Vref (depends on VREF_SEL)

pub fn vref_sel01(&self) -> VREF_SEL01_R

Bits 3:4 - Selects reference voltage (Vref) trip-point of the input buffer: ‘0’: Trip-point reference from pin_ref ‘1’: Trip-point reference of SRSS internal reference Vref (1.2 V) ‘2’: Trip-point reference of AMUXBUS_A ‘3’: Trip-point reference of AMUXBUS_B

pub fn voh_sel01(&self) -> VOH_SEL01_R

Bits 5:7 - Selects the regulated Voh output level and trip point of the input buffer for a specific SIO pin pair. Voh depends on the selected reference voltage (VREF_SEL). ‘0’: Voh = 1reference; e.g. reference at 1.2V -> Voh = 1.2V ‘1’: Voh = 1.25reference; e.g. reference at 1.2V -> Voh = 1.5V ‘2’: Voh = 1.49reference; e.g. reference at 1.2V -> Voh = ~1.8V ‘3’: Voh = 1.67reference; e.g. reference at 1.2V -> Voh = 2V ‘4’: Voh = 2.08reference; e.g. reference at 1.2V -> Voh = 2.5V ‘5’: Voh = 2.5reference; e.g. reference at 1.2V -> Voh = 3V ‘6’: Voh = 2.78reference; e.g. reference at 1.2V -> Voh = ~3.3V ‘7’: Voh = 4.16reference; e.g. reference at 1.2V -> Voh = 5.0V Note: The upper value on Voh is limited to Vddio - 400mV

pub fn vreg_en23(&self) -> VREG_EN23_R

Bit 8 - See corresponding definition for IO pins 0 and 1

pub fn ibuf_sel23(&self) -> IBUF_SEL23_R

Bit 9 - See corresponding definition for IO pins 0 and 1

pub fn vtrip_sel23(&self) -> VTRIP_SEL23_R

Bit 10 - See corresponding definition for IO pins 0 and 1

pub fn vref_sel23(&self) -> VREF_SEL23_R

Bits 11:12 - See corresponding definition for IO pins 0 and 1

pub fn voh_sel23(&self) -> VOH_SEL23_R

Bits 13:15 - See corresponding definition for IO pins 0 and 1

pub fn vreg_en45(&self) -> VREG_EN45_R

Bit 16 - See corresponding definition for IO pins 0 and 1

pub fn ibuf_sel45(&self) -> IBUF_SEL45_R

Bit 17 - See corresponding definition for IO pins 0 and 1

pub fn vtrip_sel45(&self) -> VTRIP_SEL45_R

Bit 18 - See corresponding definition for IO pins 0 and 1

pub fn vref_sel45(&self) -> VREF_SEL45_R

Bits 19:20 - See corresponding definition for IO pins 0 and 1

pub fn voh_sel45(&self) -> VOH_SEL45_R

Bits 21:23 - See corresponding definition for IO pins 0 and 1

pub fn vreg_en67(&self) -> VREG_EN67_R

Bit 24 - See corresponding definition for IO pins 0 and 1

pub fn ibuf_sel67(&self) -> IBUF_SEL67_R

Bit 25 - See corresponding definition for IO pins 0 and 1

pub fn vtrip_sel67(&self) -> VTRIP_SEL67_R

Bit 26 - See corresponding definition for IO pins 0 and 1

pub fn vref_sel67(&self) -> VREF_SEL67_R

Bits 27:28 - See corresponding definition for IO pins 0 and 1

pub fn voh_sel67(&self) -> VOH_SEL67_R

Bits 29:31 - See corresponding definition for IO pins 0 and 1