Module psoc6_01_pac::gpio::prt::intr_mask

Port interrupt mask register

Structs

EDGE0_R
Field EDGE0 reader - Masks edge interrupt on IO pin 0 ‘0’: Pin interrupt forwarding disabled ‘1’: Pin interrupt forwarding enabled

EDGE0_W
Field EDGE0 writer - Masks edge interrupt on IO pin 0 ‘0’: Pin interrupt forwarding disabled ‘1’: Pin interrupt forwarding enabled

EDGE1_R
Field EDGE1 reader - Masks edge interrupt on IO pin 1

EDGE1_W
Field EDGE1 writer - Masks edge interrupt on IO pin 1

EDGE2_R
Field EDGE2 reader - Masks edge interrupt on IO pin 2

EDGE2_W
Field EDGE2 writer - Masks edge interrupt on IO pin 2

EDGE3_R
Field EDGE3 reader - Masks edge interrupt on IO pin 3

EDGE3_W
Field EDGE3 writer - Masks edge interrupt on IO pin 3

EDGE4_R
Field EDGE4 reader - Masks edge interrupt on IO pin 4

EDGE4_W
Field EDGE4 writer - Masks edge interrupt on IO pin 4

EDGE5_R
Field EDGE5 reader - Masks edge interrupt on IO pin 5

EDGE5_W
Field EDGE5 writer - Masks edge interrupt on IO pin 5

EDGE6_R
Field EDGE6 reader - Masks edge interrupt on IO pin 6

EDGE6_W
Field EDGE6 writer - Masks edge interrupt on IO pin 6

EDGE7_R
Field EDGE7 reader - Masks edge interrupt on IO pin 7

EDGE7_W
Field EDGE7 writer - Masks edge interrupt on IO pin 7

FLT_EDGE_R
Field FLT_EDGE reader - Masks edge interrupt on filtered pin selected by INTR_CFG.FLT_SEL

FLT_EDGE_W
Field FLT_EDGE writer - Masks edge interrupt on filtered pin selected by INTR_CFG.FLT_SEL

INTR_MASK_SPEC
Port interrupt mask register

R
Register INTR_MASK reader

W
Register INTR_MASK writer