Struct psoc6_01_pac::hsiom::prt::port_sel0::IO0_SEL_R

pub struct IO0_SEL_R(_);

Field IO0_SEL reader - Selects connection for IO pin 0 route.

Implementations

impl IO0_SEL_R

pub fn variant(&self) -> IO0_SEL_A

Get enumerated values variant

pub fn is_gpio(&self) -> bool

Checks if the value of the field is GPIO

pub fn is_gpio_dsi(&self) -> bool

Checks if the value of the field is GPIO_DSI

pub fn is_dsi_dsi(&self) -> bool

Checks if the value of the field is DSI_DSI

pub fn is_dsi_gpio(&self) -> bool

Checks if the value of the field is DSI_GPIO

pub fn is_amuxa(&self) -> bool

Checks if the value of the field is AMUXA

pub fn is_amuxb(&self) -> bool

Checks if the value of the field is AMUXB

pub fn is_amuxa_dsi(&self) -> bool

Checks if the value of the field is AMUXA_DSI

pub fn is_amuxb_dsi(&self) -> bool

Checks if the value of the field is AMUXB_DSI

pub fn is_act_0(&self) -> bool

Checks if the value of the field is ACT_0

pub fn is_act_1(&self) -> bool

Checks if the value of the field is ACT_1

pub fn is_act_2(&self) -> bool

Checks if the value of the field is ACT_2

pub fn is_act_3(&self) -> bool

Checks if the value of the field is ACT_3

pub fn is_ds_0(&self) -> bool

Checks if the value of the field is DS_0

pub fn is_ds_1(&self) -> bool

Checks if the value of the field is DS_1

pub fn is_ds_2(&self) -> bool

Checks if the value of the field is DS_2

pub fn is_ds_3(&self) -> bool

Checks if the value of the field is DS_3

pub fn is_act_4(&self) -> bool

Checks if the value of the field is ACT_4

pub fn is_act_5(&self) -> bool

Checks if the value of the field is ACT_5

pub fn is_act_6(&self) -> bool

Checks if the value of the field is ACT_6

pub fn is_act_7(&self) -> bool

Checks if the value of the field is ACT_7

pub fn is_act_8(&self) -> bool

Checks if the value of the field is ACT_8

pub fn is_act_9(&self) -> bool

Checks if the value of the field is ACT_9

pub fn is_act_10(&self) -> bool

Checks if the value of the field is ACT_10

pub fn is_act_11(&self) -> bool

Checks if the value of the field is ACT_11

pub fn is_act_12(&self) -> bool

Checks if the value of the field is ACT_12

pub fn is_act_13(&self) -> bool

Checks if the value of the field is ACT_13

pub fn is_act_14(&self) -> bool

Checks if the value of the field is ACT_14

pub fn is_act_15(&self) -> bool

Checks if the value of the field is ACT_15

pub fn is_ds_4(&self) -> bool

Checks if the value of the field is DS_4

pub fn is_ds_5(&self) -> bool

Checks if the value of the field is DS_5

pub fn is_ds_6(&self) -> bool

Checks if the value of the field is DS_6

pub fn is_ds_7(&self) -> bool

Checks if the value of the field is DS_7