Module psoc6_01_pac::i2s0::tx_ctl

Transmitter control

Structs

B_CLOCK_INV_R
Field B_CLOCK_INV reader - Serial data transmission is advanced by 0.5 SCK cycles. This bit is valid only in TX slave mode. When set to ‘1’, the serial data will be transmitted 0.5 SCK cycles earlier than when set to ‘0’. 1) TX_CTL.SCKI_POL=0 and TX_CTL.B_CLOCK_INV=0: Serial data will be transmitted off the SCK falling edge 2) TX_CTL.SCKI_POL=0 and TX_CTL.B_CLOCK_INV=1: Serial data will be transmitted off the SCK rising edge that is 0.5 SCK cycles before the SCK falling edge in 1) 3) TX_CTL.SCKI_POL=1 and TX_CTL.B_CLOCK_INV=0: Serial data will be transmitted off the SCK rising edge 4) TX_CTL.SCKI_POL=1 and TX_CTL.B_CLOCK_INV=1: Serial data will be transmitted off the SCK falling edge that is 0.5 SCK cycles before the SCK rising edge in 3) (Note that this is only the appearance w.r.t. SCK edge, the actual timing is generated by an internal clock that runs 8x the SCK frequency). The word sync (TX_WS) signal is not affected by this bit setting. Note: When Master mode, must be ‘0’. (Note: This bit is connected to AR38U12.TX_CFG.TX_BCLKINV)

B_CLOCK_INV_W
Field B_CLOCK_INV writer - Serial data transmission is advanced by 0.5 SCK cycles. This bit is valid only in TX slave mode. When set to ‘1’, the serial data will be transmitted 0.5 SCK cycles earlier than when set to ‘0’. 1) TX_CTL.SCKI_POL=0 and TX_CTL.B_CLOCK_INV=0: Serial data will be transmitted off the SCK falling edge 2) TX_CTL.SCKI_POL=0 and TX_CTL.B_CLOCK_INV=1: Serial data will be transmitted off the SCK rising edge that is 0.5 SCK cycles before the SCK falling edge in 1) 3) TX_CTL.SCKI_POL=1 and TX_CTL.B_CLOCK_INV=0: Serial data will be transmitted off the SCK rising edge 4) TX_CTL.SCKI_POL=1 and TX_CTL.B_CLOCK_INV=1: Serial data will be transmitted off the SCK falling edge that is 0.5 SCK cycles before the SCK rising edge in 3) (Note that this is only the appearance w.r.t. SCK edge, the actual timing is generated by an internal clock that runs 8x the SCK frequency). The word sync (TX_WS) signal is not affected by this bit setting. Note: When Master mode, must be ‘0’. (Note: This bit is connected to AR38U12.TX_CFG.TX_BCLKINV)

CH_LEN_R
Field CH_LEN reader - Channel length in number of bits: Note: - When this field is configured to ‘6’ or ‘7’, the length is set to 32-bit (same as ‘5’). - When TDM mode, must be 32-bit length to this field. (Note: These bits are connected to AR38U12.TX_CFG.TX_CHLEN)

CH_LEN_W
Field CH_LEN writer - Channel length in number of bits: Note: - When this field is configured to ‘6’ or ‘7’, the length is set to 32-bit (same as ‘5’). - When TDM mode, must be 32-bit length to this field. (Note: These bits are connected to AR38U12.TX_CFG.TX_CHLEN)

CH_NR_R
Field CH_NR reader - Specifies number of channels per frame: Note: only ‘2channels’ is supported during Left Justfied or I2S mode. Hence software must set ‘1’ to this field in the modes. (Note: These bits are connected to AR38U12.TX_CFG.TX_CHSET)

CH_NR_W
Field CH_NR writer - Specifies number of channels per frame: Note: only ‘2channels’ is supported during Left Justfied or I2S mode. Hence software must set ‘1’ to this field in the modes. (Note: These bits are connected to AR38U12.TX_CFG.TX_CHSET)

I2S_MODE_R
Field I2S_MODE reader - Select I2S, left-justified or TDM: (Note: These bits are connected to AR38U12.TX_CFG.TX_I2S_MODE)

I2S_MODE_W
Field I2S_MODE writer - Select I2S, left-justified or TDM: (Note: These bits are connected to AR38U12.TX_CFG.TX_I2S_MODE)

MS_R
Field MS reader - Set interface in master or slave mode: (Note: This bit is connected to AR38U12.TX_CFG.TX_MS)

MS_W
Field MS writer - Set interface in master or slave mode: (Note: This bit is connected to AR38U12.TX_CFG.TX_MS)

OVHDATA_R
Field OVHDATA reader - Set overhead value: ‘0’: Set to ‘0’ ‘1’: Set to ‘1’ (Note: This bit is connected to AR38U12.TX_CFG.TX_OVHDATA)

OVHDATA_W
Field OVHDATA writer - Set overhead value: ‘0’: Set to ‘0’ ‘1’: Set to ‘1’ (Note: This bit is connected to AR38U12.TX_CFG.TX_OVHDATA)

R
Register TX_CTL reader

SCKI_POL_R
Field SCKI_POL reader - TX slave bit clock polarity. When this bit is 1, the incoming tx_sck signal is inverted before it is received by the I2S transceiver core. This bit does not affect the internal serial data transmission timing. The word sync (TX_WS) signal is not affected by this bit setting. See TX_CTL.B_CLOCK_INV for more details.

SCKI_POL_W
Field SCKI_POL writer - TX slave bit clock polarity. When this bit is 1, the incoming tx_sck signal is inverted before it is received by the I2S transceiver core. This bit does not affect the internal serial data transmission timing. The word sync (TX_WS) signal is not affected by this bit setting. See TX_CTL.B_CLOCK_INV for more details.

SCKO_POL_R
Field SCKO_POL reader - TX master bit clock polarity. When this bit is 1, the outgoing tx_sck signal is inverted after it has been transmitted from the I2S transceiver core. This bit does not affect the internal serial data transmission timing. The word sync (TX_WS) signal is not affected by this bit setting. ‘0’: When transmitter is in master mode, serial data is transmitted from the falling bit clock edge ‘1’: When transmitter is in master mode, serial data is transmitted from the rising bit clock edge

SCKO_POL_W
Field SCKO_POL writer - TX master bit clock polarity. When this bit is 1, the outgoing tx_sck signal is inverted after it has been transmitted from the I2S transceiver core. This bit does not affect the internal serial data transmission timing. The word sync (TX_WS) signal is not affected by this bit setting. ‘0’: When transmitter is in master mode, serial data is transmitted from the falling bit clock edge ‘1’: When transmitter is in master mode, serial data is transmitted from the rising bit clock edge

TX_CTL_SPEC
Transmitter control

W
Register TX_CTL writer

WD_EN_R
Field WD_EN reader - Set watchdog for ‘tx_ws_in’: ‘0’: Disabled. ‘1’: Enabled.

WD_EN_W
Field WD_EN writer - Set watchdog for ‘tx_ws_in’: ‘0’: Disabled. ‘1’: Enabled.

WORD_LEN_R
Field WORD_LEN reader - Word length in number of bits: Note: - When this field is configured to ‘6’ or ‘7’, the length is set to 32-bit (same as ‘5’). - Don’t configure this field as beyond Channel length. (Note: These bits are connected to AR38U12.TX_CFG.TX_IWL)

WORD_LEN_W
Field WORD_LEN writer - Word length in number of bits: Note: - When this field is configured to ‘6’ or ‘7’, the length is set to 32-bit (same as ‘5’). - Don’t configure this field as beyond Channel length. (Note: These bits are connected to AR38U12.TX_CFG.TX_IWL)

WS_PULSE_R
Field WS_PULSE reader - Set WS pulse width in TDM mode: (Note: This bit is connected to AR38U12.TX_CFG.TX_WS_PULSE) Note: When not TDM mode, must be ‘1’.

WS_PULSE_W
Field WS_PULSE writer - Set WS pulse width in TDM mode: (Note: This bit is connected to AR38U12.TX_CFG.TX_WS_PULSE) Note: When not TDM mode, must be ‘1’.

Enums

B_CLOCK_INV_A
Serial data transmission is advanced by 0.5 SCK cycles. This bit is valid only in TX slave mode. When set to ‘1’, the serial data will be transmitted 0.5 SCK cycles earlier than when set to ‘0’. 1) TX_CTL.SCKI_POL=0 and TX_CTL.B_CLOCK_INV=0: Serial data will be transmitted off the SCK falling edge 2) TX_CTL.SCKI_POL=0 and TX_CTL.B_CLOCK_INV=1: Serial data will be transmitted off the SCK rising edge that is 0.5 SCK cycles before the SCK falling edge in 1) 3) TX_CTL.SCKI_POL=1 and TX_CTL.B_CLOCK_INV=0: Serial data will be transmitted off the SCK rising edge 4) TX_CTL.SCKI_POL=1 and TX_CTL.B_CLOCK_INV=1: Serial data will be transmitted off the SCK falling edge that is 0.5 SCK cycles before the SCK rising edge in 3) (Note that this is only the appearance w.r.t. SCK edge, the actual timing is generated by an internal clock that runs 8x the SCK frequency). The word sync (TX_WS) signal is not affected by this bit setting. Note: When Master mode, must be ‘0’. (Note: This bit is connected to AR38U12.TX_CFG.TX_BCLKINV)

CH_LEN_A
Channel length in number of bits: Note: - When this field is configured to ‘6’ or ‘7’, the length is set to 32-bit (same as ‘5’). - When TDM mode, must be 32-bit length to this field. (Note: These bits are connected to AR38U12.TX_CFG.TX_CHLEN)

CH_NR_A
Specifies number of channels per frame: Note: only ‘2channels’ is supported during Left Justfied or I2S mode. Hence software must set ‘1’ to this field in the modes. (Note: These bits are connected to AR38U12.TX_CFG.TX_CHSET)

I2S_MODE_A
Select I2S, left-justified or TDM: (Note: These bits are connected to AR38U12.TX_CFG.TX_I2S_MODE)

MS_A
Set interface in master or slave mode: (Note: This bit is connected to AR38U12.TX_CFG.TX_MS)

WORD_LEN_A
Word length in number of bits: Note: - When this field is configured to ‘6’ or ‘7’, the length is set to 32-bit (same as ‘5’). - Don’t configure this field as beyond Channel length. (Note: These bits are connected to AR38U12.TX_CFG.TX_IWL)

WS_PULSE_A
Set WS pulse width in TDM mode: (Note: This bit is connected to AR38U12.TX_CFG.TX_WS_PULSE) Note: When not TDM mode, must be ‘1’.