Module psoc6_01_pac::ipc::struct_::lock_status

IPC lock status

Structs

ACQUIRED_R
Field ACQUIRED reader - Specifies if the lock is acquired. This field is set to ‘1’, if a ACQUIRE read transfer successfully acquires the lock (the ACQUIRE read transfer returns ACQUIRE.SUCCESS as ‘1’). If zero, P, NS, PC, and MS are not valid.

LOCK_STATUS_SPEC
IPC lock status

MS_R
Field MS reader - This field specifies the bus master identifier that successfully acquired the lock.

NS_R
Field NS reader - This field specifies the secure/non-secure access control: ‘0’: secure. ‘1’: non-secure.

PC_R
Field PC reader - This field specifies the protection context that successfully acquired the lock.

P_R
Field P reader - This field specifies the user/privileged access control: ‘0’: user mode. ‘1’: privileged mode.

R
Register LOCK_STATUS reader