Module psoc6_01_pac::lcd0::control

LCD Configuration Register

Structs

BIAS_R
Field BIAS reader - PWM bias selection

BIAS_W
Field BIAS writer - PWM bias selection

COM_NUM_R
Field COM_NUM reader - The number of COM connections minus 2. So: 0: 2 COM’s 1: 3 COM’s … 13: 15 COM’s 14: 16 COM’s 15: undefined

COM_NUM_W
Field COM_NUM writer - The number of COM connections minus 2. So: 0: 2 COM’s 1: 3 COM’s … 13: 15 COM’s 14: 16 COM’s 15: undefined

CONTROL_SPEC
LCD Configuration Register

HS_EN_R
Field HS_EN reader - High speed (HS) generator enable 1: enable 0: disable

HS_EN_W
Field HS_EN writer - High speed (HS) generator enable 1: enable 0: disable

LCD_MODE_R
Field LCD_MODE reader - HS/LS Mode selection

LCD_MODE_W
Field LCD_MODE writer - HS/LS Mode selection

LS_EN_R
Field LS_EN reader - Low speed (LS) generator enable 1: enable 0: disable

LS_EN_STAT_R
Field LS_EN_STAT reader - LS enable status bit. This bit is a copy of LS_EN that is synchronized to the low speed clock domain and back to the system clock domain. Firmware can use this bit to observe whether LS_EN has taken effect in the low speed clock domain. Firmware should never change the configuration for the LS generator without ensuring this bit is 0. The following procedure should be followed to disable the LS generator: 1. If LS_EN=0 we are done. Exit the procedure. 2. Check that LS_EN_STAT=1. If not, wait until it is. This will catch the case of a recent enable (LS_EN=1) that has not taken effect yet. 3. Set LS_EN=0. 4. Wait until LS_EN_STAT=0.

LS_EN_W
Field LS_EN writer - Low speed (LS) generator enable 1: enable 0: disable

OP_MODE_R
Field OP_MODE reader - Driving mode configuration

OP_MODE_W
Field OP_MODE writer - Driving mode configuration

R
Register CONTROL reader

TYPE_R
Field TYPE reader - LCD driving waveform type configuration.

TYPE_W
Field TYPE writer - LCD driving waveform type configuration.

W
Register CONTROL writer

Enums

BIAS_A
PWM bias selection

LCD_MODE_A
HS/LS Mode selection

OP_MODE_A
Driving mode configuration

TYPE_A
LCD driving waveform type configuration.