Struct psoc6_01_pac::lcd0::control::R

pub struct R(_);

Register CONTROL reader

Implementations

impl R

pub fn ls_en(&self) -> LS_EN_R

Bit 0 - Low speed (LS) generator enable 1: enable 0: disable

pub fn hs_en(&self) -> HS_EN_R

Bit 1 - High speed (HS) generator enable 1: enable 0: disable

pub fn lcd_mode(&self) -> LCD_MODE_R

Bit 2 - HS/LS Mode selection

pub fn type_(&self) -> TYPE_R

Bit 3 - LCD driving waveform type configuration.

pub fn op_mode(&self) -> OP_MODE_R

Bit 4 - Driving mode configuration

pub fn bias(&self) -> BIAS_R

Bits 5:6 - PWM bias selection

pub fn com_num(&self) -> COM_NUM_R

Bits 8:11 - The number of COM connections minus 2. So: 0: 2 COM’s 1: 3 COM’s … 13: 15 COM’s 14: 16 COM’s 15: undefined

pub fn ls_en_stat(&self) -> LS_EN_STAT_R

Bit 31 - LS enable status bit. This bit is a copy of LS_EN that is synchronized to the low speed clock domain and back to the system clock domain. Firmware can use this bit to observe whether LS_EN has taken effect in the low speed clock domain. Firmware should never change the configuration for the LS generator without ensuring this bit is 0. The following procedure should be followed to disable the LS generator: 1. If LS_EN=0 we are done. Exit the procedure. 2. Check that LS_EN_STAT=1. If not, wait until it is. This will catch the case of a recent enable (LS_EN=1) that has not taken effect yet. 3. Set LS_EN=0. 4. Wait until LS_EN_STAT=0.