Module psoc6_01_pac::lpcomp::cmp1_ctrl

Comparator 1 control Register

Structs

CMP1_CTRL_SPEC
Comparator 1 control Register

DSI_BYPASS1_R
Field DSI_BYPASS1 reader - Asynchronous: bypass comparator output synchronization for DSI output: 0=synchronize (level or pulse), 1=bypass (output async). Note that in DeepSleep mode this bit needs to be set to observe the DSI output on the dedicated pin.

DSI_BYPASS1_W
Field DSI_BYPASS1 writer - Asynchronous: bypass comparator output synchronization for DSI output: 0=synchronize (level or pulse), 1=bypass (output async). Note that in DeepSleep mode this bit needs to be set to observe the DSI output on the dedicated pin.

DSI_LEVEL1_R
Field DSI_LEVEL1 reader - Synchronous comparator DSI (trigger) output : 0=pulse, 1=level

DSI_LEVEL1_W
Field DSI_LEVEL1 writer - Synchronous comparator DSI (trigger) output : 0=pulse, 1=level

HYST1_R
Field HYST1 reader - Add 30mV hysteresis to the comparator 0= Disable Hysteresis 1= Enable Hysteresis

HYST1_W
Field HYST1 writer - Add 30mV hysteresis to the comparator 0= Disable Hysteresis 1= Enable Hysteresis

INTTYPE1_R
Field INTTYPE1 reader - Sets which edge will trigger an IRQ

INTTYPE1_W
Field INTTYPE1 writer - Sets which edge will trigger an IRQ

MODE1_R
Field MODE1 reader - Operating mode for the comparator

MODE1_W
Field MODE1 writer - Operating mode for the comparator

R
Register CMP1_CTRL reader

W
Register CMP1_CTRL writer

Enums

INTTYPE1_A
Sets which edge will trigger an IRQ

MODE1_A
Operating mode for the comparator