Module psoc6_01_pac::pass::aref::aref_ctrl

global AREF control

Structs

AREF_BIAS_SCALE_R
Field AREF_BIAS_SCALE reader - BIAS Current Control for all AREF Amplifiers. (These are risk mitigation bits that should not be touched by the customer: the impact on IDDA/noise/startup still needs to be characterized) 0: 125nA (reduced bias: reduction in total AREF IDDA, higher noise and longer startup times) 1: 250nA (‘default’ setting to meet bandgap performance (noise/startup) and IDDA specifications) 2: 375nA (increased bias: increase in total AREF IDDA, lower noise and shorter startup times) 3: 500nA (further increased bias: increase in total AREF IDDA, lower noise and shorter startup times)

AREF_BIAS_SCALE_W
Field AREF_BIAS_SCALE writer - BIAS Current Control for all AREF Amplifiers. (These are risk mitigation bits that should not be touched by the customer: the impact on IDDA/noise/startup still needs to be characterized) 0: 125nA (reduced bias: reduction in total AREF IDDA, higher noise and longer startup times) 1: 250nA (‘default’ setting to meet bandgap performance (noise/startup) and IDDA specifications) 2: 375nA (increased bias: increase in total AREF IDDA, lower noise and shorter startup times) 3: 500nA (further increased bias: increase in total AREF IDDA, lower noise and shorter startup times)

AREF_CTRL_SPEC
global AREF control

AREF_MODE_R
Field AREF_MODE reader - Control bit to trade off AREF settling and noise performance

AREF_MODE_W
Field AREF_MODE writer - Control bit to trade off AREF settling and noise performance

AREF_RMB_R
Field AREF_RMB reader - AREF control signals (RMB). Bit 0: Manual VBG startup circuit enable 0: normal VBG startup circuit operation 1: VBG startup circuit is forced ‘always on’ Bit 1: Manual disable of IPTAT2 DAC 0: normal IPTAT2 DAC operation 1: PTAT2 DAC is disabled while VBG startup is active Bit 2: Manual enable of VBG offset correction DAC 0: normal VBG offset correction DAC operation 1: VBG offset correction DAC is enabled while VBG startup is active

AREF_RMB_W
Field AREF_RMB writer - AREF control signals (RMB). Bit 0: Manual VBG startup circuit enable 0: normal VBG startup circuit operation 1: VBG startup circuit is forced ‘always on’ Bit 1: Manual disable of IPTAT2 DAC 0: normal IPTAT2 DAC operation 1: PTAT2 DAC is disabled while VBG startup is active Bit 2: Manual enable of VBG offset correction DAC 0: normal VBG offset correction DAC operation 1: VBG offset correction DAC is enabled while VBG startup is active

CLOCK_PUMP_PERI_SEL_R
Field CLOCK_PUMP_PERI_SEL reader - CTBm charge pump clock source select. This field has nothing to do with the AREF. 0: Use the dedicated pump clock from SRSS (default) 1: Use one of the CLK_PERI dividers

CLOCK_PUMP_PERI_SEL_W
Field CLOCK_PUMP_PERI_SEL writer - CTBm charge pump clock source select. This field has nothing to do with the AREF. 0: Use the dedicated pump clock from SRSS (default) 1: Use one of the CLK_PERI dividers

CTB_IPTAT_REDIRECT_R
Field CTB_IPTAT_REDIRECT reader - Re-direct the CTB IPTAT output current. This can be used to reduce amplifier bias glitches during power mode transitions (for PSoC4A/B DSAB backwards compatibility). 0: Opamp.IPTAT = AREF.IPTAT and Opamp.IZTAT= AREF.IZTAT 1: Opamp.IPTAT = HiZ and Opamp.IZTAT= AREF.IPTAT *Note that in Deep Sleep, the AREF IZTAT and/or IPTAT currents can be disabled and therefore the corresponding Opamp.IZTAT/IPTAT will be HiZ.

CTB_IPTAT_REDIRECT_W
Field CTB_IPTAT_REDIRECT writer - Re-direct the CTB IPTAT output current. This can be used to reduce amplifier bias glitches during power mode transitions (for PSoC4A/B DSAB backwards compatibility). 0: Opamp.IPTAT = AREF.IPTAT and Opamp.IZTAT= AREF.IZTAT 1: Opamp.IPTAT = HiZ and Opamp.IZTAT= AREF.IPTAT *Note that in Deep Sleep, the AREF IZTAT and/or IPTAT currents can be disabled and therefore the corresponding Opamp.IZTAT/IPTAT will be HiZ.

CTB_IPTAT_SCALE_R
Field CTB_IPTAT_SCALE reader - CTB IPTAT current scaler. This bit must be set in order to operate the CTB amplifiers in the lowest power mode. This bit is chip-wide (controls all CTB amplifiers). 0: 1uA 1: 100nA

CTB_IPTAT_SCALE_W
Field CTB_IPTAT_SCALE writer - CTB IPTAT current scaler. This bit must be set in order to operate the CTB amplifiers in the lowest power mode. This bit is chip-wide (controls all CTB amplifiers). 0: 1uA 1: 100nA

DEEPSLEEP_MODE_R
Field DEEPSLEEP_MODE reader - AREF DeepSleep Operation Modes (only applies if DEEPSLEEP_ON = 1)

DEEPSLEEP_MODE_W
Field DEEPSLEEP_MODE writer - AREF DeepSleep Operation Modes (only applies if DEEPSLEEP_ON = 1)

DEEPSLEEP_ON_R
Field DEEPSLEEP_ON reader - - 0: AREF IP disabled/off during DeepSleep power mode - 1: AREF IP remains enabled during DeepSleep power mode (if ENABLED=1)

DEEPSLEEP_ON_W
Field DEEPSLEEP_ON writer - - 0: AREF IP disabled/off during DeepSleep power mode - 1: AREF IP remains enabled during DeepSleep power mode (if ENABLED=1)

ENABLED_R
Field ENABLED reader - Disable AREF

ENABLED_W
Field ENABLED writer - Disable AREF

IZTAT_SEL_R
Field IZTAT_SEL reader - iztat current select control

IZTAT_SEL_W
Field IZTAT_SEL writer - iztat current select control

R
Register AREF_CTRL reader

VREF_SEL_R
Field VREF_SEL reader - bandgap voltage select control

VREF_SEL_W
Field VREF_SEL writer - bandgap voltage select control

W
Register AREF_CTRL writer

Enums

AREF_MODE_A
Control bit to trade off AREF settling and noise performance

DEEPSLEEP_MODE_A
AREF DeepSleep Operation Modes (only applies if DEEPSLEEP_ON = 1)

IZTAT_SEL_A
iztat current select control

VREF_SEL_A
bandgap voltage select control