Struct psoc6_01_pac::pass::aref::aref_ctrl::R

pub struct R(_);

Register AREF_CTRL reader

Implementations

impl R

pub fn aref_mode(&self) -> AREF_MODE_R

Bit 0 - Control bit to trade off AREF settling and noise performance

pub fn aref_bias_scale(&self) -> AREF_BIAS_SCALE_R

Bits 2:3 - BIAS Current Control for all AREF Amplifiers. (These are risk mitigation bits that should not be touched by the customer: the impact on IDDA/noise/startup still needs to be characterized) 0: 125nA (reduced bias: reduction in total AREF IDDA, higher noise and longer startup times) 1: 250nA (‘default’ setting to meet bandgap performance (noise/startup) and IDDA specifications) 2: 375nA (increased bias: increase in total AREF IDDA, lower noise and shorter startup times) 3: 500nA (further increased bias: increase in total AREF IDDA, lower noise and shorter startup times)

pub fn aref_rmb(&self) -> AREF_RMB_R

Bits 4:6 - AREF control signals (RMB). Bit 0: Manual VBG startup circuit enable 0: normal VBG startup circuit operation 1: VBG startup circuit is forced ‘always on’ Bit 1: Manual disable of IPTAT2 DAC 0: normal IPTAT2 DAC operation 1: PTAT2 DAC is disabled while VBG startup is active Bit 2: Manual enable of VBG offset correction DAC 0: normal VBG offset correction DAC operation 1: VBG offset correction DAC is enabled while VBG startup is active

pub fn ctb_iptat_scale(&self) -> CTB_IPTAT_SCALE_R

Bit 7 - CTB IPTAT current scaler. This bit must be set in order to operate the CTB amplifiers in the lowest power mode. This bit is chip-wide (controls all CTB amplifiers). 0: 1uA 1: 100nA

pub fn ctb_iptat_redirect(&self) -> CTB_IPTAT_REDIRECT_R

Bits 8:15 - Re-direct the CTB IPTAT output current. This can be used to reduce amplifier bias glitches during power mode transitions (for PSoC4A/B DSAB backwards compatibility). 0: Opamp.IPTAT = AREF.IPTAT and Opamp.IZTAT= AREF.IZTAT 1: Opamp.IPTAT = HiZ and Opamp.IZTAT= AREF.IPTAT *Note that in Deep Sleep, the AREF IZTAT and/or IPTAT currents can be disabled and therefore the corresponding Opamp.IZTAT/IPTAT will be HiZ.

pub fn iztat_sel(&self) -> IZTAT_SEL_R

Bit 16 - iztat current select control

pub fn clock_pump_peri_sel(&self) -> CLOCK_PUMP_PERI_SEL_R

Bit 19 - CTBm charge pump clock source select. This field has nothing to do with the AREF. 0: Use the dedicated pump clock from SRSS (default) 1: Use one of the CLK_PERI dividers

pub fn vref_sel(&self) -> VREF_SEL_R

Bits 20:21 - bandgap voltage select control

pub fn deepsleep_mode(&self) -> DEEPSLEEP_MODE_R

Bits 28:29 - AREF DeepSleep Operation Modes (only applies if DEEPSLEEP_ON = 1)

pub fn deepsleep_on(&self) -> DEEPSLEEP_ON_R

Bit 30 - - 0: AREF IP disabled/off during DeepSleep power mode - 1: AREF IP remains enabled during DeepSleep power mode (if ENABLED=1)

pub fn enabled(&self) -> ENABLED_R

Bit 31 - Disable AREF