Module psoc6_01_pac::pdm0::ctl

Control

Structs

CTL_SPEC
Control

ENABLED_R
Field ENABLED reader - Enables the PDM component: ‘0’: Disabled. ‘1’: Enabled.

ENABLED_W
Field ENABLED writer - Enables the PDM component: ‘0’: Disabled. ‘1’: Enabled.

PGA_L_R
Field PGA_L reader - Left channel PGA gain: +1.5dB/step, -12dB ~ +10.5dB ‘0’: -12 dB ‘1’: -10.5 dB … ‘15’: +10.5 dB (Note: These bits are connected to AR36U12.PDM_CORE_CFG.PGA_L)

PGA_L_W
Field PGA_L writer - Left channel PGA gain: +1.5dB/step, -12dB ~ +10.5dB ‘0’: -12 dB ‘1’: -10.5 dB … ‘15’: +10.5 dB (Note: These bits are connected to AR36U12.PDM_CORE_CFG.PGA_L)

PGA_R_R
Field PGA_R reader - Right channel PGA gain: +1.5dB/step, -12dB ~ +10.5dB ‘0’: -12 dB ‘1’: -10.5 dB … ‘15’ +10.5 dB (Note: These bits are connected to AR36U12.PDM_CORE_CFG.PGA_R)

PGA_R_W
Field PGA_R writer - Right channel PGA gain: +1.5dB/step, -12dB ~ +10.5dB ‘0’: -12 dB ‘1’: -10.5 dB … ‘15’ +10.5 dB (Note: These bits are connected to AR36U12.PDM_CORE_CFG.PGA_R)

R
Register CTL reader

SOFT_MUTE_R
Field SOFT_MUTE reader - Soft mute function to mute the volume smoothly ‘0’: Disabled. ‘1’: Enabled. (Note: This bit is connected to AR36U12.PDM_CORE_CFG.SOFT_MUTE)

SOFT_MUTE_W
Field SOFT_MUTE writer - Soft mute function to mute the volume smoothly ‘0’: Disabled. ‘1’: Enabled. (Note: This bit is connected to AR36U12.PDM_CORE_CFG.SOFT_MUTE)

STEP_SEL_R
Field STEP_SEL reader - Set fine gain step for smooth PGA or Soft-Mute attenuation transition. ‘0’: 0.13dB ‘1’: 0.26dB (Note: This bit is connected to AR36U12.PDM_CORE2_CFG.SEL_STEP)

STEP_SEL_W
Field STEP_SEL writer - Set fine gain step for smooth PGA or Soft-Mute attenuation transition. ‘0’: 0.13dB ‘1’: 0.26dB (Note: This bit is connected to AR36U12.PDM_CORE2_CFG.SEL_STEP)

W
Register CTL writer