Module psoc6_01_pac::pdm0::mode_ctl

Mode control

Structs

CKO_DELAY_R
Field CKO_DELAY reader - Phase difference from the rising edge of internal sampler clock (CLK_IS) to that of PDM_CKO clock: (Note: These bits are connected to AR36U12.PDM_CORE2_CFG.PDMCKO_DLY)

CKO_DELAY_W
Field CKO_DELAY writer - Phase difference from the rising edge of internal sampler clock (CLK_IS) to that of PDM_CKO clock: (Note: These bits are connected to AR36U12.PDM_CORE2_CFG.PDMCKO_DLY)

HPF_EN_N_R
Field HPF_EN_N reader - Enable high pass filter (active low) ‘1’: Disabled. ‘0’: Enabled. (Note: This bit is connected to AR36U12.PDM_CORE_CFG.ADCHPD)

HPF_EN_N_W
Field HPF_EN_N writer - Enable high pass filter (active low) ‘1’: Disabled. ‘0’: Enabled. (Note: This bit is connected to AR36U12.PDM_CORE_CFG.ADCHPD)

HPF_GAIN_R
Field HPF_GAIN reader - Adjust high pass filter coefficients. H(Z) = (1 - Z-1 ) / [1 - (1- 2 -HPF_GAIN) Z-1 ] (Note: These bits are connected to AR36U12.PDM_CORE_CFG.HPGAIN)

HPF_GAIN_W
Field HPF_GAIN writer - Adjust high pass filter coefficients. H(Z) = (1 - Z-1 ) / [1 - (1- 2 -HPF_GAIN) Z-1 ] (Note: These bits are connected to AR36U12.PDM_CORE_CFG.HPGAIN)

MODE_CTL_SPEC
Mode control

PCM_CH_SET_R
Field PCM_CH_SET reader - Specifies PCM output channels as mono or stereo: (Note: These bits are connected to AR36U12.PDM_CORE2_CFG.PCM_CHSET)

PCM_CH_SET_W
Field PCM_CH_SET writer - Specifies PCM output channels as mono or stereo: (Note: These bits are connected to AR36U12.PDM_CORE2_CFG.PCM_CHSET)

R
Register MODE_CTL reader

SWAP_LR_R
Field SWAP_LR reader - Input data L/R channel swap: ‘1’: Right/Left channel recording swap ‘0’: No Swap (Note: This bit is connected to AR36U12.PDM_CORE_CFG.LRSWAP)

SWAP_LR_W
Field SWAP_LR writer - Input data L/R channel swap: ‘1’: Right/Left channel recording swap ‘0’: No Swap (Note: This bit is connected to AR36U12.PDM_CORE_CFG.LRSWAP)

S_CYCLES_R
Field S_CYCLES reader - Set time step for gain change during PGA or soft mute operation in number of 1/a sampling rate. (Note: These bits are connected to AR36U12.PDM_CORE_CFG.S_CYCLES)

S_CYCLES_W
Field S_CYCLES writer - Set time step for gain change during PGA or soft mute operation in number of 1/a sampling rate. (Note: These bits are connected to AR36U12.PDM_CORE_CFG.S_CYCLES)

W
Register MODE_CTL writer

Enums

CKO_DELAY_A
Phase difference from the rising edge of internal sampler clock (CLK_IS) to that of PDM_CKO clock: (Note: These bits are connected to AR36U12.PDM_CORE2_CFG.PDMCKO_DLY)

PCM_CH_SET_A
Specifies PCM output channels as mono or stereo: (Note: These bits are connected to AR36U12.PDM_CORE2_CFG.PCM_CHSET)

S_CYCLES_A
Set time step for gain change during PGA or soft mute operation in number of 1/a sampling rate. (Note: These bits are connected to AR36U12.PDM_CORE_CFG.S_CYCLES)