Module psoc6_01_pac::peri::gr::sl_ctl

Slave control

Structs

DISABLED_0_R
Field DISABLED_0 reader - Peripheral group, slave 0 permanent disable. Setting this bit to 1 has the same effect as setting ENABLED_0 to 0. However, once set to 1, this bit cannot be changed back to 0 anymore.

DISABLED_0_W
Field DISABLED_0 writer - Peripheral group, slave 0 permanent disable. Setting this bit to 1 has the same effect as setting ENABLED_0 to 0. However, once set to 1, this bit cannot be changed back to 0 anymore.

DISABLED_1_R
Field DISABLED_1 reader - N/A

DISABLED_1_W
Field DISABLED_1 writer - N/A

DISABLED_2_R
Field DISABLED_2 reader - N/A

DISABLED_2_W
Field DISABLED_2 writer - N/A

DISABLED_3_R
Field DISABLED_3 reader - N/A

DISABLED_3_W
Field DISABLED_3 writer - N/A

DISABLED_4_R
Field DISABLED_4 reader - N/A

DISABLED_4_W
Field DISABLED_4 writer - N/A

DISABLED_5_R
Field DISABLED_5 reader - N/A

DISABLED_5_W
Field DISABLED_5 writer - N/A

DISABLED_6_R
Field DISABLED_6 reader - N/A

DISABLED_6_W
Field DISABLED_6 writer - N/A

DISABLED_7_R
Field DISABLED_7 reader - N/A

DISABLED_7_W
Field DISABLED_7 writer - N/A

DISABLED_8_R
Field DISABLED_8 reader - N/A

DISABLED_8_W
Field DISABLED_8 writer - N/A

DISABLED_9_R
Field DISABLED_9 reader - N/A

DISABLED_9_W
Field DISABLED_9 writer - N/A

DISABLED_10_R
Field DISABLED_10 reader - N/A

DISABLED_10_W
Field DISABLED_10 writer - N/A

DISABLED_11_R
Field DISABLED_11 reader - N/A

DISABLED_11_W
Field DISABLED_11 writer - N/A

DISABLED_12_R
Field DISABLED_12 reader - N/A

DISABLED_12_W
Field DISABLED_12 writer - N/A

DISABLED_13_R
Field DISABLED_13 reader - N/A

DISABLED_13_W
Field DISABLED_13 writer - N/A

DISABLED_14_R
Field DISABLED_14 reader - N/A

DISABLED_14_W
Field DISABLED_14 writer - N/A

DISABLED_15_R
Field DISABLED_15 reader - N/A

DISABLED_15_W
Field DISABLED_15 writer - N/A

ENABLED_0_R
Field ENABLED_0 reader - Peripheral group, slave 0 enable. If the slave is disabled, its clock is gated off (constant ‘0’) and its resets are activated. Note: For peripheral group 0 (the peripheral interconnect MMIO registers), this field is a constant ‘1’ (SW: R): the slave can NOT be disabled.

ENABLED_0_W
Field ENABLED_0 writer - Peripheral group, slave 0 enable. If the slave is disabled, its clock is gated off (constant ‘0’) and its resets are activated. Note: For peripheral group 0 (the peripheral interconnect MMIO registers), this field is a constant ‘1’ (SW: R): the slave can NOT be disabled.

ENABLED_1_R
Field ENABLED_1 reader - Peripheral group, slave 1 enable. If the slave is disabled, its clock is gated off (constant ‘0’) and its resets are activated. Note: For peripheral group 0 (the peripheral interconnect, master interface MMIO registers), this field is a constant ‘1’ (SW: R): the slave can NOT be disabled.

ENABLED_1_W
Field ENABLED_1 writer - Peripheral group, slave 1 enable. If the slave is disabled, its clock is gated off (constant ‘0’) and its resets are activated. Note: For peripheral group 0 (the peripheral interconnect, master interface MMIO registers), this field is a constant ‘1’ (SW: R): the slave can NOT be disabled.

ENABLED_2_R
Field ENABLED_2 reader - N/A

ENABLED_2_W
Field ENABLED_2 writer - N/A

ENABLED_3_R
Field ENABLED_3 reader - N/A

ENABLED_3_W
Field ENABLED_3 writer - N/A

ENABLED_4_R
Field ENABLED_4 reader - N/A

ENABLED_4_W
Field ENABLED_4 writer - N/A

ENABLED_5_R
Field ENABLED_5 reader - N/A

ENABLED_5_W
Field ENABLED_5 writer - N/A

ENABLED_6_R
Field ENABLED_6 reader - N/A

ENABLED_6_W
Field ENABLED_6 writer - N/A

ENABLED_7_R
Field ENABLED_7 reader - N/A

ENABLED_7_W
Field ENABLED_7 writer - N/A

ENABLED_8_R
Field ENABLED_8 reader - N/A

ENABLED_8_W
Field ENABLED_8 writer - N/A

ENABLED_9_R
Field ENABLED_9 reader - N/A

ENABLED_9_W
Field ENABLED_9 writer - N/A

ENABLED_10_R
Field ENABLED_10 reader - N/A

ENABLED_10_W
Field ENABLED_10 writer - N/A

ENABLED_11_R
Field ENABLED_11 reader - N/A

ENABLED_11_W
Field ENABLED_11 writer - N/A

ENABLED_12_R
Field ENABLED_12 reader - N/A

ENABLED_12_W
Field ENABLED_12 writer - N/A

ENABLED_13_R
Field ENABLED_13 reader - N/A

ENABLED_13_W
Field ENABLED_13 writer - N/A

ENABLED_14_R
Field ENABLED_14 reader - N/A

ENABLED_14_W
Field ENABLED_14 writer - N/A

ENABLED_15_R
Field ENABLED_15 reader - N/A

ENABLED_15_W
Field ENABLED_15 writer - N/A

R
Register SL_CTL reader

SL_CTL_SPEC
Slave control

W
Register SL_CTL writer