Struct psoc6_01_pac::peri::gr::sl_ctl::R

pub struct R(_);

Register SL_CTL reader

Implementations

impl R

pub fn enabled_0(&self) -> ENABLED_0_R

Bit 0 - Peripheral group, slave 0 enable. If the slave is disabled, its clock is gated off (constant ‘0’) and its resets are activated. Note: For peripheral group 0 (the peripheral interconnect MMIO registers), this field is a constant ‘1’ (SW: R): the slave can NOT be disabled.

pub fn enabled_1(&self) -> ENABLED_1_R

Bit 1 - Peripheral group, slave 1 enable. If the slave is disabled, its clock is gated off (constant ‘0’) and its resets are activated. Note: For peripheral group 0 (the peripheral interconnect, master interface MMIO registers), this field is a constant ‘1’ (SW: R): the slave can NOT be disabled.

pub fn enabled_2(&self) -> ENABLED_2_R

Bit 2 - N/A

pub fn enabled_3(&self) -> ENABLED_3_R

Bit 3 - N/A

pub fn enabled_4(&self) -> ENABLED_4_R

Bit 4 - N/A

pub fn enabled_5(&self) -> ENABLED_5_R

Bit 5 - N/A

pub fn enabled_6(&self) -> ENABLED_6_R

Bit 6 - N/A

pub fn enabled_7(&self) -> ENABLED_7_R

Bit 7 - N/A

pub fn enabled_8(&self) -> ENABLED_8_R

Bit 8 - N/A

pub fn enabled_9(&self) -> ENABLED_9_R

Bit 9 - N/A

pub fn enabled_10(&self) -> ENABLED_10_R

Bit 10 - N/A

pub fn enabled_11(&self) -> ENABLED_11_R

Bit 11 - N/A

pub fn enabled_12(&self) -> ENABLED_12_R

Bit 12 - N/A

pub fn enabled_13(&self) -> ENABLED_13_R

Bit 13 - N/A

pub fn enabled_14(&self) -> ENABLED_14_R

Bit 14 - N/A

pub fn enabled_15(&self) -> ENABLED_15_R

Bit 15 - N/A

pub fn disabled_0(&self) -> DISABLED_0_R

Bit 16 - Peripheral group, slave 0 permanent disable. Setting this bit to 1 has the same effect as setting ENABLED_0 to 0. However, once set to 1, this bit cannot be changed back to 0 anymore.

pub fn disabled_1(&self) -> DISABLED_1_R

Bit 17 - N/A

pub fn disabled_2(&self) -> DISABLED_2_R

Bit 18 - N/A

pub fn disabled_3(&self) -> DISABLED_3_R

Bit 19 - N/A

pub fn disabled_4(&self) -> DISABLED_4_R

Bit 20 - N/A

pub fn disabled_5(&self) -> DISABLED_5_R

Bit 21 - N/A

pub fn disabled_6(&self) -> DISABLED_6_R

Bit 22 - N/A

pub fn disabled_7(&self) -> DISABLED_7_R

Bit 23 - N/A

pub fn disabled_8(&self) -> DISABLED_8_R

Bit 24 - N/A

pub fn disabled_9(&self) -> DISABLED_9_R

Bit 25 - N/A

pub fn disabled_10(&self) -> DISABLED_10_R

Bit 26 - N/A

pub fn disabled_11(&self) -> DISABLED_11_R

Bit 27 - N/A

pub fn disabled_12(&self) -> DISABLED_12_R

Bit 28 - N/A

pub fn disabled_13(&self) -> DISABLED_13_R

Bit 29 - N/A

pub fn disabled_14(&self) -> DISABLED_14_R

Bit 30 - N/A

pub fn disabled_15(&self) -> DISABLED_15_R

Bit 31 - N/A