Module psoc6_01_pac::peri_ms::ppu_pr::ms_size

Master region, size

Structs

MS_SIZE_SPEC
Master region, size

R
Register MS_SIZE reader

REGION_SIZE_R
Field REGION_SIZE reader - This field specifies the size of the master region: ‘5’: 64 B region The master region includes the SL_ADDR, SL_SIZE, SL_ATT0, …, SL_ATT3, MS_ADDR, MS_SIZE, MS_ATT0, …, MS_ATT3 registers. Therefore, the access privileges for all these registers is determined by MS_ATT0, …, MS_ATT3.

VALID_R
Field VALID reader - Master region enable: ‘1’: Enabled.