Module psoc6_01_pac::peri_ms::ppu_pr::sl_size

Slave region, size

Structs

R
Register SL_SIZE reader

REGION_SIZE_R
Field REGION_SIZE reader - This field specifies the size of the slave region: ‘0’: Undefined. ‘1’: 4 B region (this is the smallest region size). ‘2’: 8 B region ‘3’: 16 B region ‘4’: 32 B region ‘5’: 64 B region ‘6’: 128 B region ‘7’: 256 B region ‘8’: 512 B region ‘9’: 1 KB region ‘10’: 2 KB region ‘11’: 4 KB region ‘12’: 8 KB region ‘13’: 16 KB region ‘14’: 32 KB region ‘15’: 64 KB region ‘16’: 128 KB region ‘17’: 256 KB region ‘18’: 512 KB region ‘19’: 1 MB region ‘20’: 2 MB region ‘21’: 4 MB region ‘22’: 8 MB region ‘23’: 16 MB region ‘24’: 32 MB region ‘25’: 64 MB region ‘26’: 128 MB region ‘27’: 256 MB region ‘28’: 512 MB region ‘29’: 1 GB region ‘30’: 2 GB region ‘31’: 4 GB region

REGION_SIZE_W
Field REGION_SIZE writer - This field specifies the size of the slave region: ‘0’: Undefined. ‘1’: 4 B region (this is the smallest region size). ‘2’: 8 B region ‘3’: 16 B region ‘4’: 32 B region ‘5’: 64 B region ‘6’: 128 B region ‘7’: 256 B region ‘8’: 512 B region ‘9’: 1 KB region ‘10’: 2 KB region ‘11’: 4 KB region ‘12’: 8 KB region ‘13’: 16 KB region ‘14’: 32 KB region ‘15’: 64 KB region ‘16’: 128 KB region ‘17’: 256 KB region ‘18’: 512 KB region ‘19’: 1 MB region ‘20’: 2 MB region ‘21’: 4 MB region ‘22’: 8 MB region ‘23’: 16 MB region ‘24’: 32 MB region ‘25’: 64 MB region ‘26’: 128 MB region ‘27’: 256 MB region ‘28’: 512 MB region ‘29’: 1 GB region ‘30’: 2 GB region ‘31’: 4 GB region

SL_SIZE_SPEC
Slave region, size

VALID_R
Field VALID reader - Slave region enable: ‘0’: Disabled. A disabled region will never result in a match on the transfer address. ‘1’: Enabled.

VALID_W
Field VALID writer - Slave region enable: ‘0’: Disabled. A disabled region will never result in a match on the transfer address. ‘1’: Enabled.

W
Register SL_SIZE writer