Module psoc6_01_pac::prot::mpu::mpu_struct::att

MPU region attrributes

Structs

ATT_SPEC
MPU region attrributes

ENABLED_R
Field ENABLED reader - Region enable: ‘0’: Disabled. A disabled region will never result in a match on the bus transfer address. ‘1’: Enabled. Note: a disabled address region performs logic gating to reduce dynamic power consumption.

ENABLED_W
Field ENABLED writer - Region enable: ‘0’: Disabled. A disabled region will never result in a match on the bus transfer address. ‘1’: Enabled. Note: a disabled address region performs logic gating to reduce dynamic power consumption.

NS_R
Field NS reader - Non-secure: ‘0’: Secure (secure accesses allowed, non-secure access NOT allowed). ‘1’: Non-secure (both secure and non-secure accesses allowed).

NS_W
Field NS writer - Non-secure: ‘0’: Secure (secure accesses allowed, non-secure access NOT allowed). ‘1’: Non-secure (both secure and non-secure accesses allowed).

PR_R
Field PR reader - Privileged read enable: ‘0’: Disabled (privileged, read accesses are NOT allowed). ‘1’: Enabled (privileged, read accesses are allowed).

PR_W
Field PR writer - Privileged read enable: ‘0’: Disabled (privileged, read accesses are NOT allowed). ‘1’: Enabled (privileged, read accesses are allowed).

PW_R
Field PW reader - Privileged write enable: ‘0’: Disabled (privileged, write accesses are NOT allowed). ‘1’: Enabled (privileged, write accesses are allowed).

PW_W
Field PW writer - Privileged write enable: ‘0’: Disabled (privileged, write accesses are NOT allowed). ‘1’: Enabled (privileged, write accesses are allowed).

PX_R
Field PX reader - Privileged execute enable: ‘0’: Disabled (privileged, execute accesses are NOT allowed). ‘1’: Enabled (privileged, execute accesses are allowed).

PX_W
Field PX writer - Privileged execute enable: ‘0’: Disabled (privileged, execute accesses are NOT allowed). ‘1’: Enabled (privileged, execute accesses are allowed).

R
Register ATT reader

REGION_SIZE_R
Field REGION_SIZE reader - This field specifies the region size: ‘0’-‘6’: Undefined. ‘7’: 256 B region ‘8’: 512 B region ‘9’: 1 KB region ‘10’: 2 KB region ‘11’: 4 KB region ‘12’: 8 KB region ‘13’: 16 KB region ‘14’: 32 KB region ‘15’: 64 KB region ‘16’: 128 KB region ‘17’: 256 KB region ‘18’: 512 KB region ‘19’: 1 MB region ‘20’: 2 MB region ‘21’: 4 MB region ‘22’: 8 MB region ‘23’: 16 MB region ‘24’: 32 MB region ‘25’: 64 MB region ‘26’: 128 MB region ‘27’: 256 MB region ‘28’: 512 MB region ‘39’: 1 GB region ‘30’: 2 GB region ‘31’: 4 GB region

REGION_SIZE_W
Field REGION_SIZE writer - This field specifies the region size: ‘0’-‘6’: Undefined. ‘7’: 256 B region ‘8’: 512 B region ‘9’: 1 KB region ‘10’: 2 KB region ‘11’: 4 KB region ‘12’: 8 KB region ‘13’: 16 KB region ‘14’: 32 KB region ‘15’: 64 KB region ‘16’: 128 KB region ‘17’: 256 KB region ‘18’: 512 KB region ‘19’: 1 MB region ‘20’: 2 MB region ‘21’: 4 MB region ‘22’: 8 MB region ‘23’: 16 MB region ‘24’: 32 MB region ‘25’: 64 MB region ‘26’: 128 MB region ‘27’: 256 MB region ‘28’: 512 MB region ‘39’: 1 GB region ‘30’: 2 GB region ‘31’: 4 GB region

UR_R
Field UR reader - User read enable: ‘0’: Disabled (user, read accesses are NOT allowed). ‘1’: Enabled (user, read accesses are allowed).

UR_W
Field UR writer - User read enable: ‘0’: Disabled (user, read accesses are NOT allowed). ‘1’: Enabled (user, read accesses are allowed).

UW_R
Field UW reader - User write enable: ‘0’: Disabled (user, write accesses are NOT allowed). ‘1’: Enabled (user, write accesses are allowed).

UW_W
Field UW writer - User write enable: ‘0’: Disabled (user, write accesses are NOT allowed). ‘1’: Enabled (user, write accesses are allowed).

UX_R
Field UX reader - User execute enable: ‘0’: Disabled (user, execute accesses are NOT allowed). ‘1’: Enabled (user, execute accesses are allowed).

UX_W
Field UX writer - User execute enable: ‘0’: Disabled (user, execute accesses are NOT allowed). ‘1’: Enabled (user, execute accesses are allowed).

W
Register ATT writer