Module psoc6_01_pac::sar::chan_config

Channel configuration register.

Structs

AVG_EN_R
Field AVG_EN reader - Averaging enable for this channel. If set the AVG_CNT and AVG_SHIFT settings are used for sampling the addressed pin(s)

AVG_EN_W
Field AVG_EN writer - Averaging enable for this channel. If set the AVG_CNT and AVG_SHIFT settings are used for sampling the addressed pin(s)

CHAN_CONFIG_SPEC
Channel configuration register.

DIFFERENTIAL_EN_R
Field DIFFERENTIAL_EN reader - Differential enable for this channel. If NEG_ADDR_EN=0 and this bit is 1 then POS_PIN_ADDR[0] is ignored and considered to be 0, i.e. POS_PIN_ADDR points to the even pin of a pin pair. In that case the even pin of the pair is connected to Vplus and the odd pin of the pair is connected to Vminus. POS_PORT_ADDR is used to identify the port that contains the pins. - 0: The voltage on the addressed pin is measured (Single-ended) and the resulting value is stored in the corresponding data register. - 1: The differential voltage on the addressed pin pair is measured and the resulting value is stored in the corresponding data register. (if NEG_ADDR_EN=0 then POS_PIN_ADDR[0] is ignored).

DIFFERENTIAL_EN_W
Field DIFFERENTIAL_EN writer - Differential enable for this channel. If NEG_ADDR_EN=0 and this bit is 1 then POS_PIN_ADDR[0] is ignored and considered to be 0, i.e. POS_PIN_ADDR points to the even pin of a pin pair. In that case the even pin of the pair is connected to Vplus and the odd pin of the pair is connected to Vminus. POS_PORT_ADDR is used to identify the port that contains the pins. - 0: The voltage on the addressed pin is measured (Single-ended) and the resulting value is stored in the corresponding data register. - 1: The differential voltage on the addressed pin pair is measured and the resulting value is stored in the corresponding data register. (if NEG_ADDR_EN=0 then POS_PIN_ADDR[0] is ignored).

DSI_OUT_EN_R
Field DSI_OUT_EN reader - DSI data output enable for this channel. - 0: the conversion result for this channel is only stored in the channel data register and the corresponding CHAN_DATA_VALID bit is set. - 1: the conversion result for this channel is stored in the channel data register and the corresponding CHAN_DATA_VALID bit is set. The same data (same formatting), together with the channel number, is sent out on the DSI communication channel for processing in UDBs.

DSI_OUT_EN_W
Field DSI_OUT_EN writer - DSI data output enable for this channel. - 0: the conversion result for this channel is only stored in the channel data register and the corresponding CHAN_DATA_VALID bit is set. - 1: the conversion result for this channel is stored in the channel data register and the corresponding CHAN_DATA_VALID bit is set. The same data (same formatting), together with the channel number, is sent out on the DSI communication channel for processing in UDBs.

NEG_ADDR_EN_R
Field NEG_ADDR_EN reader - 1 - The NEG_PIN_ADDR and NEG_PORT_ADDR determines what drives the Vminus pin. This is a variation of differential mode with no even-odd pair limitation 0 - The NEG_SEL determines what drives the Vminus pin.

NEG_ADDR_EN_W
Field NEG_ADDR_EN writer - 1 - The NEG_PIN_ADDR and NEG_PORT_ADDR determines what drives the Vminus pin. This is a variation of differential mode with no even-odd pair limitation 0 - The NEG_SEL determines what drives the Vminus pin.

NEG_PIN_ADDR_R
Field NEG_PIN_ADDR reader - Address of the neg pin to be sampled by this channel.

NEG_PIN_ADDR_W
Field NEG_PIN_ADDR writer - Address of the neg pin to be sampled by this channel.

NEG_PORT_ADDR_R
Field NEG_PORT_ADDR reader - Address of the neg port that contains the pin to be sampled by this channel.

NEG_PORT_ADDR_W
Field NEG_PORT_ADDR writer - Address of the neg port that contains the pin to be sampled by this channel.

POS_PIN_ADDR_R
Field POS_PIN_ADDR reader - Address of the pin to be sampled by this channel (connected to Vplus)

POS_PIN_ADDR_W
Field POS_PIN_ADDR writer - Address of the pin to be sampled by this channel (connected to Vplus)

POS_PORT_ADDR_R
Field POS_PORT_ADDR reader - Address of the port that contains the pin to be sampled by this channel (connected to Vplus)

POS_PORT_ADDR_W
Field POS_PORT_ADDR writer - Address of the port that contains the pin to be sampled by this channel (connected to Vplus)

R
Register CHAN_CONFIG[%s] reader

SAMPLE_TIME_SEL_R
Field SAMPLE_TIME_SEL reader - Sample time select: select which of the 4 global sample times to use for this channel

SAMPLE_TIME_SEL_W
Field SAMPLE_TIME_SEL writer - Sample time select: select which of the 4 global sample times to use for this channel

W
Register CHAN_CONFIG[%s] writer

Enums

NEG_PORT_ADDR_A
Address of the neg port that contains the pin to be sampled by this channel.

POS_PORT_ADDR_A
Address of the port that contains the pin to be sampled by this channel (connected to Vplus)