Module psoc6_01_pac::sar::ctrl

Analog control register.

Structs

BOOSTPUMP_EN_R
Field BOOSTPUMP_EN reader - deprecated

BOOSTPUMP_EN_W
Field BOOSTPUMP_EN writer - deprecated

COMP_DLY_R
Field COMP_DLY reader - Set the comparator latch delay in accordance with SAR conversion rate

COMP_DLY_W
Field COMP_DLY writer - Set the comparator latch delay in accordance with SAR conversion rate

COMP_PWR_R
Field COMP_PWR reader - Comparator power mode.

COMP_PWR_W
Field COMP_PWR writer - Comparator power mode.

CTRL_SPEC
Analog control register.

DEEPSLEEP_ON_R
Field DEEPSLEEP_ON reader - - 0: SARMUX IP disabled off during DeepSleep power mode - 1: SARMUX IP remains enabled during DeepSleep power mode (if ENABLED=1)

DEEPSLEEP_ON_W
Field DEEPSLEEP_ON writer - - 0: SARMUX IP disabled off during DeepSleep power mode - 1: SARMUX IP remains enabled during DeepSleep power mode (if ENABLED=1)

DSI_MODE_R
Field DSI_MODE reader - SAR sequencer takes configuration from DSI signals (note this also has the same effect as SWITCH_DISABLE==1) - 0: Normal mode, SAR sequencer operates according to CHAN_EN enables and CHAN_CONFIG channel configurations - 1: CHAN_EN, INJ_START_EN and channel configurations in CHAN_CONFIG and INJ_CHAN_CONFIG are ignored

DSI_MODE_W
Field DSI_MODE writer - SAR sequencer takes configuration from DSI signals (note this also has the same effect as SWITCH_DISABLE==1) - 0: Normal mode, SAR sequencer operates according to CHAN_EN enables and CHAN_CONFIG channel configurations - 1: CHAN_EN, INJ_START_EN and channel configurations in CHAN_CONFIG and INJ_CHAN_CONFIG are ignored

DSI_SYNC_CONFIG_R
Field DSI_SYNC_CONFIG reader - - 0: bypass clock domain synchronization of the DSI config signals. - 1: synchronize the DSI config signals to peripheral clock domain.

DSI_SYNC_CONFIG_W
Field DSI_SYNC_CONFIG writer - - 0: bypass clock domain synchronization of the DSI config signals. - 1: synchronize the DSI config signals to peripheral clock domain.

ENABLED_R
Field ENABLED reader - - 0: SAR IP disabled (put analog in power down and stop clocks), also can clear FW_TRIGGER and INJ_START_EN (if not tailgating) on write. - 1: SAR IP enabled.

ENABLED_W
Field ENABLED writer - - 0: SAR IP disabled (put analog in power down and stop clocks), also can clear FW_TRIGGER and INJ_START_EN (if not tailgating) on write. - 1: SAR IP enabled.

NEG_SEL_R
Field NEG_SEL reader - SARADC internal NEG selection for Single ended conversion

NEG_SEL_W
Field NEG_SEL writer - SARADC internal NEG selection for Single ended conversion

PWR_CTRL_VREF_R
Field PWR_CTRL_VREF reader - VREF buffer low power mode.

PWR_CTRL_VREF_W
Field PWR_CTRL_VREF writer - VREF buffer low power mode.

R
Register CTRL reader

REFBUF_EN_R
Field REFBUF_EN reader - For normal ADC operation this bit must be set, for all reference choices - internal, external or vdda based reference. Setting this bit is critical to proper function of switches inside SARREF block.

REFBUF_EN_W
Field REFBUF_EN writer - For normal ADC operation this bit must be set, for all reference choices - internal, external or vdda based reference. Setting this bit is critical to proper function of switches inside SARREF block.

SAR_HW_CTRL_NEGVREF_R
Field SAR_HW_CTRL_NEGVREF reader - Hardware control: 0=only firmware control, 1=hardware control masked by firmware setting for VREF to NEG switch.

SAR_HW_CTRL_NEGVREF_W
Field SAR_HW_CTRL_NEGVREF writer - Hardware control: 0=only firmware control, 1=hardware control masked by firmware setting for VREF to NEG switch.

SPARE_R
Field SPARE reader - Spare controls, not yet designated, for late changes done with an ECO

SPARE_W
Field SPARE writer - Spare controls, not yet designated, for late changes done with an ECO

SWITCH_DISABLE_R
Field SWITCH_DISABLE reader - Disable SAR sequencer from enabling routing switches (note DSI and firmware can always close switches independent of this control) - 0: Normal mode, SAR sequencer changes switches according to pin address in channel configurations - 1: Switches disabled, SAR sequencer does not enable any switches, it is the responsibility of the firmware or UDBs (through DSI) to set the switches to route the signal to be converted through the SARMUX

SWITCH_DISABLE_W
Field SWITCH_DISABLE writer - Disable SAR sequencer from enabling routing switches (note DSI and firmware can always close switches independent of this control) - 0: Normal mode, SAR sequencer changes switches according to pin address in channel configurations - 1: Switches disabled, SAR sequencer does not enable any switches, it is the responsibility of the firmware or UDBs (through DSI) to set the switches to route the signal to be converted through the SARMUX

VREF_BYP_CAP_EN_R
Field VREF_BYP_CAP_EN reader - VREF bypass cap enable for when VREF buffer is on

VREF_BYP_CAP_EN_W
Field VREF_BYP_CAP_EN writer - VREF bypass cap enable for when VREF buffer is on

VREF_SEL_R
Field VREF_SEL reader - SARADC internal VREF selection.

VREF_SEL_W
Field VREF_SEL writer - SARADC internal VREF selection.

W
Register CTRL writer

Enums

COMP_DLY_A
Set the comparator latch delay in accordance with SAR conversion rate

COMP_PWR_A
Comparator power mode.

NEG_SEL_A
SARADC internal NEG selection for Single ended conversion

PWR_CTRL_VREF_A
VREF buffer low power mode.

VREF_SEL_A
SARADC internal VREF selection.