Module psoc6_01_pac::sar::inj_chan_config

Injection channel configuration register.

Structs

INJ_AVG_EN_R
Field INJ_AVG_EN reader - Averaging enable for this channel. If set the AVG_CNT and AVG_SHIFT settings are used for sampling the addressed pin(s)

INJ_AVG_EN_W
Field INJ_AVG_EN writer - Averaging enable for this channel. If set the AVG_CNT and AVG_SHIFT settings are used for sampling the addressed pin(s)

INJ_CHAN_CONFIG_SPEC
Injection channel configuration register.

INJ_DIFFERENTIAL_EN_R
Field INJ_DIFFERENTIAL_EN reader - Differential enable for this channel. - 0: The voltage on the addressed pin is measured (Single-ended) and the resulting value is stored in the corresponding data register. - 1: The differential voltage on the addressed pin pair is measured and the resulting value is stored in the corresponding data register. (INJ_PIN_ADDR[0] is ignored).

INJ_DIFFERENTIAL_EN_W
Field INJ_DIFFERENTIAL_EN writer - Differential enable for this channel. - 0: The voltage on the addressed pin is measured (Single-ended) and the resulting value is stored in the corresponding data register. - 1: The differential voltage on the addressed pin pair is measured and the resulting value is stored in the corresponding data register. (INJ_PIN_ADDR[0] is ignored).

INJ_PIN_ADDR_R
Field INJ_PIN_ADDR reader - Address of the pin to be sampled by this injection channel. If differential is enabled then INJ_PIN_ADDR[0] is ignored and considered to be 0, i.e. INJ_PIN_ADDR points to the even pin of a pin pair.

INJ_PIN_ADDR_W
Field INJ_PIN_ADDR writer - Address of the pin to be sampled by this injection channel. If differential is enabled then INJ_PIN_ADDR[0] is ignored and considered to be 0, i.e. INJ_PIN_ADDR points to the even pin of a pin pair.

INJ_PORT_ADDR_R
Field INJ_PORT_ADDR reader - Address of the port that contains the pin to be sampled by this channel.

INJ_PORT_ADDR_W
Field INJ_PORT_ADDR writer - Address of the port that contains the pin to be sampled by this channel.

INJ_SAMPLE_TIME_SEL_R
Field INJ_SAMPLE_TIME_SEL reader - Injection sample time select: select which of the 4 global sample times to use for this channel

INJ_SAMPLE_TIME_SEL_W
Field INJ_SAMPLE_TIME_SEL writer - Injection sample time select: select which of the 4 global sample times to use for this channel

INJ_START_EN_R
Field INJ_START_EN reader - Set by firmware to enable the injection channel. If INJ_TAILGATING is not set this bit also functions as trigger for this channel. Cleared by hardware after this channel has been sampled (i.e. this channel is always one shot even if CONTINUOUS is set). Also cleared if the SAR is disabled.

INJ_START_EN_W
Field INJ_START_EN writer - Set by firmware to enable the injection channel. If INJ_TAILGATING is not set this bit also functions as trigger for this channel. Cleared by hardware after this channel has been sampled (i.e. this channel is always one shot even if CONTINUOUS is set). Also cleared if the SAR is disabled.

INJ_TAILGATING_R
Field INJ_TAILGATING reader - Injection channel tailgating. - 0: no tailgating for this channel, SAR is immediately triggered when the INJ_START_EN bit is set if the SAR is not busy. If the SAR is busy, the INJ channel addressed pin is sampled at the end of the current scan. - 1: injection channel tailgating. The addressed pin is sampled after the next trigger and after all enabled channels have been scanned.

INJ_TAILGATING_W
Field INJ_TAILGATING writer - Injection channel tailgating. - 0: no tailgating for this channel, SAR is immediately triggered when the INJ_START_EN bit is set if the SAR is not busy. If the SAR is busy, the INJ channel addressed pin is sampled at the end of the current scan. - 1: injection channel tailgating. The addressed pin is sampled after the next trigger and after all enabled channels have been scanned.

R
Register INJ_CHAN_CONFIG reader

W
Register INJ_CHAN_CONFIG writer

Enums

INJ_PORT_ADDR_A
Address of the port that contains the pin to be sampled by this channel.