Module psoc6_01_pac::sar::inj_result

Injection channel result register

Structs

INJ_COLLISION_INTR_MIR_R
Field INJ_COLLISION_INTR_MIR reader - mirror bit of corresponding bit in SAR_INTR register

INJ_EOC_INTR_MIR_R
Field INJ_EOC_INTR_MIR reader - mirror bit of corresponding bit in SAR_INTR register

INJ_NEWVALUE_R
Field INJ_NEWVALUE reader - The data in this register received a new value (only relevant for UAB, this bit shows the value of the UAB valid bit)

INJ_RANGE_INTR_MIR_R
Field INJ_RANGE_INTR_MIR reader - mirror bit of corresponding bit in SAR_INTR register

INJ_RESULT_R
Field INJ_RESULT reader - SAR conversion result of the channel.

INJ_RESULT_SPEC
Injection channel result register

INJ_SATURATE_INTR_MIR_R
Field INJ_SATURATE_INTR_MIR reader - mirror bit of corresponding bit in SAR_INTR register

R
Register INJ_RESULT reader