Module psoc6_01_pac::sar::intr

Interrupt request register.

Structs

DSI_COLLISION_INTR_R
Field DSI_COLLISION_INTR reader - DSI Collision Interrupt: hardware sets this interrupt when the DSI trigger signal is asserted while the SAR is BUSY. Raising this interrupt is delayed to when the scan caused by the DSI trigger has been completed, i.e. not when the preceding scan with which this trigger collided is completed. When this interrupt is set it implies that the channels were sampled later than was intended (jitter). Write with ‘1’ to clear bit.

DSI_COLLISION_INTR_W
Field DSI_COLLISION_INTR writer - DSI Collision Interrupt: hardware sets this interrupt when the DSI trigger signal is asserted while the SAR is BUSY. Raising this interrupt is delayed to when the scan caused by the DSI trigger has been completed, i.e. not when the preceding scan with which this trigger collided is completed. When this interrupt is set it implies that the channels were sampled later than was intended (jitter). Write with ‘1’ to clear bit.

EOS_INTR_R
Field EOS_INTR reader - End Of Scan Interrupt: hardware sets this interrupt after completing a scan of all the enabled channels. Write with ‘1’ to clear bit.

EOS_INTR_W
Field EOS_INTR writer - End Of Scan Interrupt: hardware sets this interrupt after completing a scan of all the enabled channels. Write with ‘1’ to clear bit.

FW_COLLISION_INTR_R
Field FW_COLLISION_INTR reader - Firmware Collision Interrupt: hardware sets this interrupt when FW_TRIGGER is asserted while the SAR is BUSY. Raising this interrupt is delayed to when the scan caused by the FW_TRIGGER has been completed, i.e. not when the preceding scan with which this trigger collided is completed. When this interrupt is set it implies that the channels were sampled later than was intended (jitter). Write with ‘1’ to clear bit.

FW_COLLISION_INTR_W
Field FW_COLLISION_INTR writer - Firmware Collision Interrupt: hardware sets this interrupt when FW_TRIGGER is asserted while the SAR is BUSY. Raising this interrupt is delayed to when the scan caused by the FW_TRIGGER has been completed, i.e. not when the preceding scan with which this trigger collided is completed. When this interrupt is set it implies that the channels were sampled later than was intended (jitter). Write with ‘1’ to clear bit.

INJ_COLLISION_INTR_R
Field INJ_COLLISION_INTR reader - Injection Collision Interrupt: hardware sets this interrupt when the injection trigger signal is asserted (INJ_START_EN==1 && INJ_TAILGATING==0) while the SAR is BUSY. Raising this interrupt is delayed to when the sampling of the injection channel has been completed, i.e. not when the preceding scan with which this trigger collided is completed. When this interrupt is set it implies that the injection channel was sampled later than was intended. Write with ‘1’ to clear bit.

INJ_COLLISION_INTR_W
Field INJ_COLLISION_INTR writer - Injection Collision Interrupt: hardware sets this interrupt when the injection trigger signal is asserted (INJ_START_EN==1 && INJ_TAILGATING==0) while the SAR is BUSY. Raising this interrupt is delayed to when the sampling of the injection channel has been completed, i.e. not when the preceding scan with which this trigger collided is completed. When this interrupt is set it implies that the injection channel was sampled later than was intended. Write with ‘1’ to clear bit.

INJ_EOC_INTR_R
Field INJ_EOC_INTR reader - Injection End of Conversion Interrupt: hardware sets this interrupt after completing the conversion for the injection channel (irrespective of if tailgating was used). Write with ‘1’ to clear bit.

INJ_EOC_INTR_W
Field INJ_EOC_INTR writer - Injection End of Conversion Interrupt: hardware sets this interrupt after completing the conversion for the injection channel (irrespective of if tailgating was used). Write with ‘1’ to clear bit.

INJ_RANGE_INTR_R
Field INJ_RANGE_INTR reader - Injection Range detect Interrupt: hardware sets this interrupt if the injection conversion result (after averaging) met the condition specified by the SAR_RANGE registers. Write with ‘1’ to clear bit.

INJ_RANGE_INTR_W
Field INJ_RANGE_INTR writer - Injection Range detect Interrupt: hardware sets this interrupt if the injection conversion result (after averaging) met the condition specified by the SAR_RANGE registers. Write with ‘1’ to clear bit.

INJ_SATURATE_INTR_R
Field INJ_SATURATE_INTR reader - Injection Saturation Interrupt: hardware sets this interrupt if an injection conversion result (before averaging) is either 0x000 or 0xFFF, this is an indication that the ADC likely saturated. Write with ‘1’ to clear bit.

INJ_SATURATE_INTR_W
Field INJ_SATURATE_INTR writer - Injection Saturation Interrupt: hardware sets this interrupt if an injection conversion result (before averaging) is either 0x000 or 0xFFF, this is an indication that the ADC likely saturated. Write with ‘1’ to clear bit.

INTR_SPEC
Interrupt request register.

OVERFLOW_INTR_R
Field OVERFLOW_INTR reader - Overflow Interrupt: hardware sets this interrupt when it sets a new EOS_INTR while that bit was not yet cleared by the firmware. Write with ‘1’ to clear bit.

OVERFLOW_INTR_W
Field OVERFLOW_INTR writer - Overflow Interrupt: hardware sets this interrupt when it sets a new EOS_INTR while that bit was not yet cleared by the firmware. Write with ‘1’ to clear bit.

R
Register INTR reader

W
Register INTR writer