Struct psoc6_01_pac::sar::sample_ctrl::W

pub struct W(_);

Register SAMPLE_CTRL writer

Implementations

impl W

pub fn left_align(&mut self) -> LEFT_ALIGN_W<'_>

Bit 1 - Left align data in data[15:0], default data is right aligned in data[11:0], with sign extension to 16 bits if the channel is differential.

pub fn single_ended_signed(&mut self) -> SINGLE_ENDED_SIGNED_W<'_>

Bit 2 - Output data from a single ended conversion as a signed value If AVG_MODE = 1 (Interleaved averaging), then SINGLE_ENDED_SIGNED must be configured identically to DIFFERENTIAL_SIGNED.

pub fn differential_signed(&mut self) -> DIFFERENTIAL_SIGNED_W<'_>

Bit 3 - Output data from a differential conversion as a signed value when DIFFERENTIAL_EN or NEG_ADDR_EN is set to 1 If AVG_MODE = 1 (Interleaved averaging), then DIFFERENTIAL_SIGNED must be configured identically to SINGLE_ENDED_SIGNED.

pub fn avg_cnt(&mut self) -> AVG_CNT_W<'_>

Bits 4:6 - Averaging Count for channels that have averaging enabled (AVG_EN). A channel will be sampled (1<<(AVG_CNT+1)) = [2..256] times. - In ACCUNDUMP mode (1st order accumulate and dump filter) a channel will be sampled back to back, the average result is calculated and stored and then the next enabled channel is sampled. If shifting is not enabled (AVG_SHIFT=0) then the result is forced to shift right so that is fits in 16 bits, so right shift is done by max(0,AVG_CNT-3). - In INTERLEAVED mode one sample is taken per triggered scan, only in the scan where the final averaging count is reached a valid average is calculated and stored in the RESULT register (by definition the same scan for all the channels that have averaging enabled). In all other scans the RESULT register for averaged channels will have an invalid result and the intermediate accumulated value is stored in the 16-bit WORK register. In this mode make sure that the averaging count is low enough to ensure that the intermediate value does not exceed 16-bits otherwise the MSBs will be lost. So for a 12-bit resolution the averaging count should be set to 16 or less (AVG_CNT=<3).

pub fn avg_shift(&mut self) -> AVG_SHIFT_W<'_>

Bit 7 - Averaging shifting: after averaging the result is shifted right to fit in 12 bits.

pub fn avg_mode(&mut self) -> AVG_MODE_W<'_>

Bit 8 - Averaging mode, in DSI mode this bit is ignored and only AccuNDump mode is available.

pub fn continuous(&mut self) -> CONTINUOUS_W<'_>

Bit 16 - - 0: Wait for next FW_TRIGGER (one shot) or hardware trigger (e.g. from TPWM for periodic triggering) before scanning enabled channels. - 1: Continuously scan enabled channels, ignore triggers.

pub fn dsi_trigger_en(&mut self) -> DSI_TRIGGER_EN_W<'_>

Bit 17 - - 0: firmware trigger only: disable hardware trigger tr_sar_in. - 1: enable hardware trigger tr_sar_in (e.g. from TCPWM, GPIO or UDB).

pub fn dsi_trigger_level(&mut self) -> DSI_TRIGGER_LEVEL_W<'_>

Bit 18 - - 0: trigger signal is a pulse input, a positive edge detected on the trigger signal triggers a new scan. - 1: trigger signal is a level input, as long as the trigger signal remains high the SAR will do continuous scans.

pub fn dsi_sync_trigger(&mut self) -> DSI_SYNC_TRIGGER_W<'_>

Bit 19 - - 0: bypass clock domain synchronization of the trigger signal. - 1: synchronize the trigger signal to the SAR clock domain, if needed an edge detect is done in the peripheral clock domain.

pub fn uab_scan_mode(&mut self) -> UAB_SCAN_MODE_W<'_>

Bit 22 - Select whether UABs are scheduled or unscheduled. When no UAB is scanned this selection is ignored.

pub fn repeat_invalid(&mut self) -> REPEAT_INVALID_W<'_>

Bit 23 - For unscheduled UAB_SCAN_MODE only, do the following if an invalid sample is received: - 0: use the last known valid sample for that channel and clear the NEWVALUE flag - 1: repeat the conversions until a valid sample is received (caveat: could be never if the UAB valid window is incorrectly schedule w.r.t. SAR sampling)

pub fn valid_sel(&mut self) -> VALID_SEL_W<'_>

Bits 24:26 - Static UAB Valid select 0=UAB0 half 0 Valid output 1=UAB0 half 1 Valid output 2=UAB1 half 0 Valid output 3=UAB1 half 1 Valid output 4=UAB2 half 0 Valid output 5=UAB2 half 1 Valid output 6=UAB3 half 0 Valid output 7=UAB3 half 1 Valid output

pub fn valid_sel_en(&mut self) -> VALID_SEL_EN_W<'_>

Bit 27 - Enable static UAB Valid selection (override Hardware)

pub fn valid_ignore(&mut self) -> VALID_IGNORE_W<'_>

Bit 28 - Ignore UAB valid signal, including the dynamic/Hardware from AROUTE and the static Valid selection from the VALID_SEL fields above

pub fn trigger_out_en(&mut self) -> TRIGGER_OUT_EN_W<'_>

Bit 30 - SAR output trigger enable (used for UAB synchronization). To ensure multiple UABs starting at the same trigger it is recommended to use this bit to temporarily disable the trigger output until all those UABs are set to run (UAB.SRAM_CTRL.RUN=1).

pub fn eos_dsi_out_en(&mut self) -> EOS_DSI_OUT_EN_W<'_>

Bit 31 - Enable to output EOS_INTR to DSI. When enabled each time EOS_INTR is set by the hardware also a trigger pulse is send on the tr_sar_out signal.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.