Module psoc6_01_pac::scb0::intr_rx_mask

Receiver interrupt mask

Structs

BAUD_DETECT_R
Field BAUD_DETECT reader - Mask bit for corresponding bit in interrupt request register.

BAUD_DETECT_W
Field BAUD_DETECT writer - Mask bit for corresponding bit in interrupt request register.

BLOCKED_R
Field BLOCKED reader - Mask bit for corresponding bit in interrupt request register.

BLOCKED_W
Field BLOCKED writer - Mask bit for corresponding bit in interrupt request register.

BREAK_DETECT_R
Field BREAK_DETECT reader - Mask bit for corresponding bit in interrupt request register.

BREAK_DETECT_W
Field BREAK_DETECT writer - Mask bit for corresponding bit in interrupt request register.

FRAME_ERROR_R
Field FRAME_ERROR reader - Mask bit for corresponding bit in interrupt request register.

FRAME_ERROR_W
Field FRAME_ERROR writer - Mask bit for corresponding bit in interrupt request register.

FULL_R
Field FULL reader - Mask bit for corresponding bit in interrupt request register.

FULL_W
Field FULL writer - Mask bit for corresponding bit in interrupt request register.

INTR_RX_MASK_SPEC
Receiver interrupt mask

NOT_EMPTY_R
Field NOT_EMPTY reader - Mask bit for corresponding bit in interrupt request register.

NOT_EMPTY_W
Field NOT_EMPTY writer - Mask bit for corresponding bit in interrupt request register.

OVERFLOW_R
Field OVERFLOW reader - Mask bit for corresponding bit in interrupt request register.

OVERFLOW_W
Field OVERFLOW writer - Mask bit for corresponding bit in interrupt request register.

PARITY_ERROR_R
Field PARITY_ERROR reader - Mask bit for corresponding bit in interrupt request register.

PARITY_ERROR_W
Field PARITY_ERROR writer - Mask bit for corresponding bit in interrupt request register.

R
Register INTR_RX_MASK reader

TRIGGER_R
Field TRIGGER reader - Mask bit for corresponding bit in interrupt request register.

TRIGGER_W
Field TRIGGER writer - Mask bit for corresponding bit in interrupt request register.

UNDERFLOW_R
Field UNDERFLOW reader - Mask bit for corresponding bit in interrupt request register.

UNDERFLOW_W
Field UNDERFLOW writer - Mask bit for corresponding bit in interrupt request register.

W
Register INTR_RX_MASK writer