Module psoc6_01_pac::scb0::intr_s_mask

Slave interrupt mask

Structs

I2C_ACK_R
Field I2C_ACK reader - Mask bit for corresponding bit in interrupt request register.

I2C_ACK_W
Field I2C_ACK writer - Mask bit for corresponding bit in interrupt request register.

I2C_ADDR_MATCH_R
Field I2C_ADDR_MATCH reader - Mask bit for corresponding bit in interrupt request register.

I2C_ADDR_MATCH_W
Field I2C_ADDR_MATCH writer - Mask bit for corresponding bit in interrupt request register.

I2C_ARB_LOST_R
Field I2C_ARB_LOST reader - Mask bit for corresponding bit in interrupt request register.

I2C_ARB_LOST_W
Field I2C_ARB_LOST writer - Mask bit for corresponding bit in interrupt request register.

I2C_BUS_ERROR_R
Field I2C_BUS_ERROR reader - Mask bit for corresponding bit in interrupt request register.

I2C_BUS_ERROR_W
Field I2C_BUS_ERROR writer - Mask bit for corresponding bit in interrupt request register.

I2C_GENERAL_R
Field I2C_GENERAL reader - Mask bit for corresponding bit in interrupt request register.

I2C_GENERAL_W
Field I2C_GENERAL writer - Mask bit for corresponding bit in interrupt request register.

I2C_NACK_R
Field I2C_NACK reader - Mask bit for corresponding bit in interrupt request register.

I2C_NACK_W
Field I2C_NACK writer - Mask bit for corresponding bit in interrupt request register.

I2C_START_R
Field I2C_START reader - Mask bit for corresponding bit in interrupt request register.

I2C_START_W
Field I2C_START writer - Mask bit for corresponding bit in interrupt request register.

I2C_STOP_R
Field I2C_STOP reader - Mask bit for corresponding bit in interrupt request register.

I2C_STOP_W
Field I2C_STOP writer - Mask bit for corresponding bit in interrupt request register.

I2C_WRITE_STOP_R
Field I2C_WRITE_STOP reader - Mask bit for corresponding bit in interrupt request register.

I2C_WRITE_STOP_W
Field I2C_WRITE_STOP writer - Mask bit for corresponding bit in interrupt request register.

INTR_S_MASK_SPEC
Slave interrupt mask

R
Register INTR_S_MASK reader

SPI_BUS_ERROR_R
Field SPI_BUS_ERROR reader - Mask bit for corresponding bit in interrupt request register.

SPI_BUS_ERROR_W
Field SPI_BUS_ERROR writer - Mask bit for corresponding bit in interrupt request register.

SPI_EZ_STOP_R
Field SPI_EZ_STOP reader - Mask bit for corresponding bit in interrupt request register.

SPI_EZ_STOP_W
Field SPI_EZ_STOP writer - Mask bit for corresponding bit in interrupt request register.

SPI_EZ_WRITE_STOP_R
Field SPI_EZ_WRITE_STOP reader - Mask bit for corresponding bit in interrupt request register.

SPI_EZ_WRITE_STOP_W
Field SPI_EZ_WRITE_STOP writer - Mask bit for corresponding bit in interrupt request register.

W
Register INTR_S_MASK writer