Module psoc6_01_pac::scb0::spi_ctrl

SPI control

Structs

CPHA_R
Field CPHA reader - Indicates the clock phase. This field, together with the CPOL field, indicates when MOSI data is driven and MISO data is captured: - Motorola mode 0. CPOL is ‘0’, CPHA is ‘0’: MOSI is driven on a falling edge of SCLK. MISO is captured on a rising edge of SCLK. - Motorola mode 1. CPOL is ‘0’, CPHA is ‘1’: MOSI is driven on a rising edge of SCLK. MISO is captured on a falling edge of SCLK. - Motorola mode 2. CPOL is ‘1’, CPHA is ‘0’: MOSI is driven on a rising edge of SCLK. MISO is captured on a falling edge of SCLK. - Motorola mode 3. CPOL is ‘1’, CPHA is ‘1’: MOSI is driven on a falling edge of SCLK. MISO is captured on a rising edge of SCLK. In SPI Motorola submode, all four CPOL/CPHA modes are valid. in SPI NS submode, only CPOL=0 CPHA=0 mode is valid. in SPI TI submode, only CPOL=0 CPHA=1 mode is valid.

CPHA_W
Field CPHA writer - Indicates the clock phase. This field, together with the CPOL field, indicates when MOSI data is driven and MISO data is captured: - Motorola mode 0. CPOL is ‘0’, CPHA is ‘0’: MOSI is driven on a falling edge of SCLK. MISO is captured on a rising edge of SCLK. - Motorola mode 1. CPOL is ‘0’, CPHA is ‘1’: MOSI is driven on a rising edge of SCLK. MISO is captured on a falling edge of SCLK. - Motorola mode 2. CPOL is ‘1’, CPHA is ‘0’: MOSI is driven on a rising edge of SCLK. MISO is captured on a falling edge of SCLK. - Motorola mode 3. CPOL is ‘1’, CPHA is ‘1’: MOSI is driven on a falling edge of SCLK. MISO is captured on a rising edge of SCLK. In SPI Motorola submode, all four CPOL/CPHA modes are valid. in SPI NS submode, only CPOL=0 CPHA=0 mode is valid. in SPI TI submode, only CPOL=0 CPHA=1 mode is valid.

CPOL_R
Field CPOL reader - Indicates the clock polarity. This field, together with the CPHA field, indicates when MOSI data is driven and MISO data is captured: - CPOL is ‘0’: SCLK is ‘0’ when not transmitting data. - CPOL is ‘1’: SCLK is ‘1’ when not transmitting data.

CPOL_W
Field CPOL writer - Indicates the clock polarity. This field, together with the CPHA field, indicates when MOSI data is driven and MISO data is captured: - CPOL is ‘0’: SCLK is ‘0’ when not transmitting data. - CPOL is ‘1’: SCLK is ‘1’ when not transmitting data.

LATE_MISO_SAMPLE_R
Field LATE_MISO_SAMPLE reader - Changes the SCLK edge on which MISO is captured. Only used in master mode. When ‘0’, the default applies (for Motorola as determined by CPOL and CPHA, for Texas Instruments on the falling edge of SCLK and for National Semiconductors on the rising edge of SCLK). When ‘1’, the alternate clock edge is used (which comes half a SPI SCLK period later). Late sampling addresses the round trip delay associated with transmitting SCLK from the master to the slave and transmitting MISO from the slave to the master.

LATE_MISO_SAMPLE_W
Field LATE_MISO_SAMPLE writer - Changes the SCLK edge on which MISO is captured. Only used in master mode. When ‘0’, the default applies (for Motorola as determined by CPOL and CPHA, for Texas Instruments on the falling edge of SCLK and for National Semiconductors on the rising edge of SCLK). When ‘1’, the alternate clock edge is used (which comes half a SPI SCLK period later). Late sampling addresses the round trip delay associated with transmitting SCLK from the master to the slave and transmitting MISO from the slave to the master.

LOOPBACK_R
Field LOOPBACK reader - Local loopback control (does NOT affect the information on the pins). Only used in master mode. Not used in National Semiconductors submode. ‘0’: the SPI master MISO line ‘spi_miso_in’ is connected to the SPI MISO pin. ‘1’: the SPI master MISO line ‘spi_miso_in’ is connected to the SPI master MOSI line ‘spi_mosi_out’. In other words, in loopback mode the SPI master receives on MISO what it transmits on MOSI.

LOOPBACK_W
Field LOOPBACK writer - Local loopback control (does NOT affect the information on the pins). Only used in master mode. Not used in National Semiconductors submode. ‘0’: the SPI master MISO line ‘spi_miso_in’ is connected to the SPI MISO pin. ‘1’: the SPI master MISO line ‘spi_miso_in’ is connected to the SPI master MOSI line ‘spi_mosi_out’. In other words, in loopback mode the SPI master receives on MISO what it transmits on MOSI.

MASTER_MODE_R
Field MASTER_MODE reader - Master (‘1’) or slave (‘0’) mode. In master mode, transmission will commence on availability of data frames in the TX FIFO. In slave mode, when selected and there is no data frame in the TX FIFO, the slave will transmit all ’1’s. In both master and slave modes, received data frames will be lost if the RX FIFO is full.

MASTER_MODE_W
Field MASTER_MODE writer - Master (‘1’) or slave (‘0’) mode. In master mode, transmission will commence on availability of data frames in the TX FIFO. In slave mode, when selected and there is no data frame in the TX FIFO, the slave will transmit all ’1’s. In both master and slave modes, received data frames will be lost if the RX FIFO is full.

MODE_R
Field MODE reader - N/A

MODE_W
Field MODE writer - N/A

R
Register SPI_CTRL reader

SCLK_CONTINUOUS_R
Field SCLK_CONTINUOUS reader - Only applicable in master mode. ‘0’: SCLK is generated, when the SPI master is enabled and data is transmitted. ‘1’: SCLK is generated, when the SPI master is enabled. This mode is useful for slave devices that use SCLK for functional operation other than just SPI functionality.

SCLK_CONTINUOUS_W
Field SCLK_CONTINUOUS writer - Only applicable in master mode. ‘0’: SCLK is generated, when the SPI master is enabled and data is transmitted. ‘1’: SCLK is generated, when the SPI master is enabled. This mode is useful for slave devices that use SCLK for functional operation other than just SPI functionality.

SELECT_PRECEDE_R
Field SELECT_PRECEDE reader - Only used in SPI Texas Instruments’ submode. When ‘1’, the data frame start indication is a pulse on the SELECT line that precedes the transfer of the first data frame bit. When ‘0’, the data frame start indication is a pulse on the SELECT line that coincides with the transfer of the first data frame bit.

SELECT_PRECEDE_W
Field SELECT_PRECEDE writer - Only used in SPI Texas Instruments’ submode. When ‘1’, the data frame start indication is a pulse on the SELECT line that precedes the transfer of the first data frame bit. When ‘0’, the data frame start indication is a pulse on the SELECT line that coincides with the transfer of the first data frame bit.

SPI_CTRL_SPEC
SPI control

SSEL_CONTINUOUS_R
Field SSEL_CONTINUOUS reader - Continuous SPI data transfers enabled (‘1’) or not (‘0’). This field is used in master mode. In slave mode, both continuous and non-continuous SPI data transfers are supported independent of this field. When continuous transfers are enabled individual data frame transfers are not necessarily separated by slave deselection (as indicated by the level or pulse on the SELECT line): if the TX FIFO has multiple data frames, data frames are send out without slave deselection. When continuous transfers are not enabled individual data frame transfers are always separated by slave deselection: independent of the availability of TX FIFO data frames.

SSEL_CONTINUOUS_W
Field SSEL_CONTINUOUS writer - Continuous SPI data transfers enabled (‘1’) or not (‘0’). This field is used in master mode. In slave mode, both continuous and non-continuous SPI data transfers are supported independent of this field. When continuous transfers are enabled individual data frame transfers are not necessarily separated by slave deselection (as indicated by the level or pulse on the SELECT line): if the TX FIFO has multiple data frames, data frames are send out without slave deselection. When continuous transfers are not enabled individual data frame transfers are always separated by slave deselection: independent of the availability of TX FIFO data frames.

SSEL_POLARITY0_R
Field SSEL_POLARITY0 reader - Slave select polarity. SSEL_POLARITY0 applies to the outgoing SPI slave select signal 0 (master mode) and to the incoming SPI slave select signal (slave mode). For Motorola and National Semiconductors submodes: ‘0’: slave select is low/‘0’ active. ‘1’: slave select is high/‘1’ active. For Texas Instruments submode: ‘0’: high/‘1’ active precede/coincide pulse. ‘1’: low/‘0’ active precede/coincide pulse.

SSEL_POLARITY0_W
Field SSEL_POLARITY0 writer - Slave select polarity. SSEL_POLARITY0 applies to the outgoing SPI slave select signal 0 (master mode) and to the incoming SPI slave select signal (slave mode). For Motorola and National Semiconductors submodes: ‘0’: slave select is low/‘0’ active. ‘1’: slave select is high/‘1’ active. For Texas Instruments submode: ‘0’: high/‘1’ active precede/coincide pulse. ‘1’: low/‘0’ active precede/coincide pulse.

SSEL_POLARITY1_R
Field SSEL_POLARITY1 reader - Slave select polarity.

SSEL_POLARITY1_W
Field SSEL_POLARITY1 writer - Slave select polarity.

SSEL_POLARITY2_R
Field SSEL_POLARITY2 reader - Slave select polarity.

SSEL_POLARITY2_W
Field SSEL_POLARITY2 writer - Slave select polarity.

SSEL_POLARITY3_R
Field SSEL_POLARITY3 reader - Slave select polarity.

SSEL_POLARITY3_W
Field SSEL_POLARITY3 writer - Slave select polarity.

SSEL_R
Field SSEL reader - Selects one of the four incoming/outgoing SPI slave select signals: - 0: Slave 0, SSEL[0]. - 1: Slave 1, SSEL[1]. - 2: Slave 2, SSEL[2]. - 3: Slave 3, SSEL[3]. The IP should be disabled when changes are made to this field.

SSEL_W
Field SSEL writer - Selects one of the four incoming/outgoing SPI slave select signals: - 0: Slave 0, SSEL[0]. - 1: Slave 1, SSEL[1]. - 2: Slave 2, SSEL[2]. - 3: Slave 3, SSEL[3]. The IP should be disabled when changes are made to this field.

W
Register SPI_CTRL writer

Enums

MODE_A
N/A