Module psoc6_01_pac::scb0::uart_flow_ctrl

UART flow control

Structs

CTS_ENABLED_R
Field CTS_ENABLED reader - Enable use of CTS input signal ‘uart_cts_in’ by the UART transmitter: ‘0’: Disabled. The UART transmitter ignores ‘uart_cts_in’, and transmits when a data frame is available for transmission in the TX FIFO or the TX shift register. ‘1’: Enabled. The UART transmitter uses ‘uart_cts_in’ to qualify the transmission of data. It transmits when ‘uart_cts_in’ is active and a data frame is available for transmission in the TX FIFO or the TX shift register. If UART_CTRL.LOOPBACK is ‘1’, ‘uart_cts_in’ is connected to ‘uart_rts_out’ in the IP (both signals are subjected to signal polarity changes as indicated by RTS_POLARITY and CTS_POLARITY).

CTS_ENABLED_W
Field CTS_ENABLED writer - Enable use of CTS input signal ‘uart_cts_in’ by the UART transmitter: ‘0’: Disabled. The UART transmitter ignores ‘uart_cts_in’, and transmits when a data frame is available for transmission in the TX FIFO or the TX shift register. ‘1’: Enabled. The UART transmitter uses ‘uart_cts_in’ to qualify the transmission of data. It transmits when ‘uart_cts_in’ is active and a data frame is available for transmission in the TX FIFO or the TX shift register. If UART_CTRL.LOOPBACK is ‘1’, ‘uart_cts_in’ is connected to ‘uart_rts_out’ in the IP (both signals are subjected to signal polarity changes as indicated by RTS_POLARITY and CTS_POLARITY).

CTS_POLARITY_R
Field CTS_POLARITY reader - Polarity of the CTS input signal ‘uart_cts_in’: ‘0’: CTS is low/‘0’ active; ‘uart_cts_in’ is ‘0’ when active and ‘uart_cts_in’ is ‘1’ when inactive. ‘1’: CTS is high/‘1’ active; ‘uart_cts_in’ is ‘1’ when active and ‘uart_cts_in’ is ‘0’ when inactive.

CTS_POLARITY_W
Field CTS_POLARITY writer - Polarity of the CTS input signal ‘uart_cts_in’: ‘0’: CTS is low/‘0’ active; ‘uart_cts_in’ is ‘0’ when active and ‘uart_cts_in’ is ‘1’ when inactive. ‘1’: CTS is high/‘1’ active; ‘uart_cts_in’ is ‘1’ when active and ‘uart_cts_in’ is ‘0’ when inactive.

R
Register UART_FLOW_CTRL reader

RTS_POLARITY_R
Field RTS_POLARITY reader - Polarity of the RTS output signal ‘uart_rts_out’: ‘0’: RTS is low/‘0’ active; ‘uart_rts_out’ is ‘0’ when active and ‘uart_rts_out’ is ‘1’ when inactive. ‘1’: RTS is high/‘1’ active; ‘uart_rts_out’ is ‘1’ when active and ‘uart_rts_out’ is ‘0’ when inactive. During IP reset (Hibernate system power mode), ‘uart_rts_out’ is ‘1’. This represents an inactive state assuming a low/‘0’ active polarity.

RTS_POLARITY_W
Field RTS_POLARITY writer - Polarity of the RTS output signal ‘uart_rts_out’: ‘0’: RTS is low/‘0’ active; ‘uart_rts_out’ is ‘0’ when active and ‘uart_rts_out’ is ‘1’ when inactive. ‘1’: RTS is high/‘1’ active; ‘uart_rts_out’ is ‘1’ when active and ‘uart_rts_out’ is ‘0’ when inactive. During IP reset (Hibernate system power mode), ‘uart_rts_out’ is ‘1’. This represents an inactive state assuming a low/‘0’ active polarity.

TRIGGER_LEVEL_R
Field TRIGGER_LEVEL reader - Trigger level. When the receiver FIFO has less entries than the amount of this field, a Ready To Send (RTS) output signal ‘uart_rts_out’ is activated. By setting this field to ‘0’, flow control is effectively SW disabled (may be useful for debug purposes).

TRIGGER_LEVEL_W
Field TRIGGER_LEVEL writer - Trigger level. When the receiver FIFO has less entries than the amount of this field, a Ready To Send (RTS) output signal ‘uart_rts_out’ is activated. By setting this field to ‘0’, flow control is effectively SW disabled (may be useful for debug purposes).

UART_FLOW_CTRL_SPEC
UART flow control

W
Register UART_FLOW_CTRL writer