Module psoc6_01_pac::scb5::cmd_resp_status

Command/response status

Structs

CMD_RESP_EC_BUSY_R
Field CMD_RESP_EC_BUSY reader - Indicates whether the CURR_RD_ADDR and CURR_WR_ADDR fields in this register are reliable (when CMD_RESP_EC_BUSY is ‘0’) or not reliable (when CMD_RESP_EC_BUSY is ‘1’). Note: - When there is no ongoing bus transfer, CMD_RESP_EC_BUSY is ‘0’ (reliable). - When there is a ongoing bus transfer, CMD_RESP_EC_BUSY is ‘0’ (reliable), when the CURR_RD_ADDR and CURR_WR_ADDR are not being updated by the HW. - When there is a ongoing bus transfer, CMD_RESP_EC_BUSY is ‘1’ (not reliable), when the CURR_RD_ADDR or CURR_WR_ADDR are being updated by the HW. Note that this update lasts one I2C clock cycle, or two SPI clock cycles.

CMD_RESP_EC_BUS_BUSY_R
Field CMD_RESP_EC_BUS_BUSY reader - Indicates whether there is an ongoing bus transfer to the IP. ‘0’: no ongoing bus transfer. ‘1’: ongoing bus transfer. For SPI, the field is ‘1’ when the slave is selected. For I2C, the field is set to ‘1’ at a I2C START/RESTART. In case of an address match, the field is set to ‘0’ on a I2C STOP. In case of NO address match, the field is set to ‘0’ after the failing address match.

CMD_RESP_STATUS_SPEC
Command/response status

CURR_RD_ADDR_R
Field CURR_RD_ADDR reader - I2C/SPI read current address for CMD_RESP mode. HW increments the field after a read access to the memory buffer. However, when the last memory buffer address is reached, the address is NOT incremented (but remains at the maximum memory buffer address). The field is used to determine how many bytes have been read (# bytes = CURR_RD_ADDR - CMD_RESP_CTRL.BASE_RD_ADDR). This field is reliable during when there is no bus transfer. This field is potentially unreliable when there is a bus transfer bus transfer: when CMD_RESP_EC_BUSY is ‘0’, the field is reliable.

CURR_WR_ADDR_R
Field CURR_WR_ADDR reader - I2C/SPI write current address for CMD_RESP mode. HW increments the field after a read access to the memory buffer. However, when the last memory buffer address is reached, the address is NOT incremented (but remains at the maximum memory buffer address). The field is used to determine how many bytes have been written (# bytes = CURR_WR_ADDR - CMD_RESP_CTRL.BASE_WR_ADDR). This field is reliable during when there is no bus transfer. This field is potentially unreliable when there is a bus transfer bus transfer: when CMD_RESP_EC_BUSY is ‘0’, the field is reliable.

R
Register CMD_RESP_STATUS reader