Module psoc6_01_pac::scb5::ctrl

Generic control

Structs

ADDR_ACCEPT_R
Field ADDR_ACCEPT reader - Determines whether a received matching address is accepted in the RX FIFO (‘1’) or not (‘0’). In I2C mode, this field is used to allow the slave to put the received slave address or general call address in the RX FIFO. Note that a received matching address is put in the RX FIFO when ADDR_ACCEPT is ‘1’ for both I2C read and write transfers. In multi-processor UART receiver mode, this field is used to allow the receiver to put the received address in the RX FIFO. Note: non-matching addresses are never put in the RX FIFO.

ADDR_ACCEPT_W
Field ADDR_ACCEPT writer - Determines whether a received matching address is accepted in the RX FIFO (‘1’) or not (‘0’). In I2C mode, this field is used to allow the slave to put the received slave address or general call address in the RX FIFO. Note that a received matching address is put in the RX FIFO when ADDR_ACCEPT is ‘1’ for both I2C read and write transfers. In multi-processor UART receiver mode, this field is used to allow the receiver to put the received address in the RX FIFO. Note: non-matching addresses are never put in the RX FIFO.

BLOCK_R
Field BLOCK reader - Only used in externally clocked mode. If the externally clocked logic and the MMIO SW accesses to EZ memory coincide/collide, this bit determines whether a SW access should block and result in bus wait states (‘BLOCK is 1’) or not (BLOCK is ‘0’). IF BLOCK is ‘0’ and the accesses collide, MMIO read operations return 0xffff:ffff and MMIO write operations are ignored. Colliding accesses are registered as interrupt causes: field BLOCKED of MMIO registers INTR_TX and INTR_RX.

BLOCK_W
Field BLOCK writer - Only used in externally clocked mode. If the externally clocked logic and the MMIO SW accesses to EZ memory coincide/collide, this bit determines whether a SW access should block and result in bus wait states (‘BLOCK is 1’) or not (BLOCK is ‘0’). IF BLOCK is ‘0’ and the accesses collide, MMIO read operations return 0xffff:ffff and MMIO write operations are ignored. Colliding accesses are registered as interrupt causes: field BLOCKED of MMIO registers INTR_TX and INTR_RX.

BYTE_MODE_R
Field BYTE_MODE reader - Determines the number of bits per FIFO data element: ‘0’: 16-bit FIFO data elements. ‘1’: 8-bit FIFO data elements. This mode doubles the amount of FIFO entries, but TX_CTRL.DATA_WIDTH and RX_CTRL.DATA_WIDTH are restricted to [0, 7].

BYTE_MODE_W
Field BYTE_MODE writer - Determines the number of bits per FIFO data element: ‘0’: 16-bit FIFO data elements. ‘1’: 8-bit FIFO data elements. This mode doubles the amount of FIFO entries, but TX_CTRL.DATA_WIDTH and RX_CTRL.DATA_WIDTH are restricted to [0, 7].

CMD_RESP_MODE_R
Field CMD_RESP_MODE reader - Determines CMD_RESP mode of operation: ‘0’: CMD_RESP mode disabled. ‘1’: CMD_RESP mode enabled (also requires EC_AM_MODE and EC_OP_MODE to be set to ‘1’).

CMD_RESP_MODE_W
Field CMD_RESP_MODE writer - Determines CMD_RESP mode of operation: ‘0’: CMD_RESP mode disabled. ‘1’: CMD_RESP mode enabled (also requires EC_AM_MODE and EC_OP_MODE to be set to ‘1’).

CTRL_SPEC
Generic control

EC_AM_MODE_R
Field EC_AM_MODE reader - Internally clocked mode (‘0’) or externally clocked mode (‘1’) address matching (I2C) or selection (SPI). In internally clocked mode, the serial interface protocols run off the peripheral clock. In externally clocked mode, the serial interface protocols run off the clock as provided by the serial interface. Externally clocked mode is only used for synchronous serial interface protocols (SPI and I2C) in slave mode. In SPI mode, only Motorola submode (all Motorola modes: 0, 1, 2, 3) is supported. In UART mode this field should be ‘0’.

EC_AM_MODE_W
Field EC_AM_MODE writer - Internally clocked mode (‘0’) or externally clocked mode (‘1’) address matching (I2C) or selection (SPI). In internally clocked mode, the serial interface protocols run off the peripheral clock. In externally clocked mode, the serial interface protocols run off the clock as provided by the serial interface. Externally clocked mode is only used for synchronous serial interface protocols (SPI and I2C) in slave mode. In SPI mode, only Motorola submode (all Motorola modes: 0, 1, 2, 3) is supported. In UART mode this field should be ‘0’.

EC_OP_MODE_R
Field EC_OP_MODE reader - Internally clocked mode (‘0’) or externally clocked mode (‘1’) operation. In internally clocked mode, the serial interface protocols run off the peripheral clock. In externally clocked mode, the serial interface protocols run off the clock as provided by the serial interface. Externally clocked operation mode is only used for synchronous serial interface protocols (SPI and I2C) in slave mode AND EZ mode. In SPI mode, only Motorola submode (all Motorola modes: 0, 1, 2, 3) is supported. The maximum SPI slave, EZ mode bitrate is 48 Mbps (transmission and IO delays outside the IP will degrade the effective bitrate). In UART mode this field should be ‘0’.

EC_OP_MODE_W
Field EC_OP_MODE writer - Internally clocked mode (‘0’) or externally clocked mode (‘1’) operation. In internally clocked mode, the serial interface protocols run off the peripheral clock. In externally clocked mode, the serial interface protocols run off the clock as provided by the serial interface. Externally clocked operation mode is only used for synchronous serial interface protocols (SPI and I2C) in slave mode AND EZ mode. In SPI mode, only Motorola submode (all Motorola modes: 0, 1, 2, 3) is supported. The maximum SPI slave, EZ mode bitrate is 48 Mbps (transmission and IO delays outside the IP will degrade the effective bitrate). In UART mode this field should be ‘0’.

ENABLED_R
Field ENABLED reader - IP enabled (‘1’) or not (‘0’). The proper order in which to initialize the IP is as follows: - Program protocol specific information using SPI_CTRL, UART_CTRL (and UART_TX_CTRL and UART_RX_CTRL) or I2C_CTRL. This includes selection of a submode, master/slave functionality and transmitter/receiver functionality when applicable. - Program generic transmitter (TX_CTRL) and receiver (RX_CTRL) information. This includes enabling of the transmitter and receiver functionality. - Program transmitter FIFO (TX_FIFO_CTRL) and receiver FIFO (RX_FIFO_CTRL) information. - Program CTRL to enable IP, select the specific operation mode and oversampling factor. When the IP is enabled, no control information should be changed. Changes should be made AFTER disabling the IP, e.g. to modify the operation mode (from I2C to SPI) or to go from externally to internally clocked. The change takes effect after the IP is re-enabled. Note that disabling the IP will cause re-initialization of the design and associated state is lost (e.g. FIFO content).

ENABLED_W
Field ENABLED writer - IP enabled (‘1’) or not (‘0’). The proper order in which to initialize the IP is as follows: - Program protocol specific information using SPI_CTRL, UART_CTRL (and UART_TX_CTRL and UART_RX_CTRL) or I2C_CTRL. This includes selection of a submode, master/slave functionality and transmitter/receiver functionality when applicable. - Program generic transmitter (TX_CTRL) and receiver (RX_CTRL) information. This includes enabling of the transmitter and receiver functionality. - Program transmitter FIFO (TX_FIFO_CTRL) and receiver FIFO (RX_FIFO_CTRL) information. - Program CTRL to enable IP, select the specific operation mode and oversampling factor. When the IP is enabled, no control information should be changed. Changes should be made AFTER disabling the IP, e.g. to modify the operation mode (from I2C to SPI) or to go from externally to internally clocked. The change takes effect after the IP is re-enabled. Note that disabling the IP will cause re-initialization of the design and associated state is lost (e.g. FIFO content).

EZ_MODE_R
Field EZ_MODE reader - Non EZ mode (‘0’) or EZ mode (‘1’). In EZ mode, a meta protocol is applied to the serial interface protocol. This meta protocol adds meaning to the data frames transferred by the serial interface protocol: a data frame can represent a memory address, a write memory data element or a read memory data element. EZ mode is only used for synchronous serial interface protocols: SPI and I2C. In SPI mode, only Motorola submode (all Motorola modes: 0, 1, 2, 3) is supported and the transmitter should use continuous data frames; i.e. data frames mot separated by slave deselection. This mode is only applicable to slave functionality. In EZ mode, the slave can read from and write to an addressable memory structure of up to 256 bytes. In EZ mode, data frames should 8-bit in size and should be transmitted and received with the Most Significant Bit (MSB) first. In UART mode this field should be ‘0’.

EZ_MODE_W
Field EZ_MODE writer - Non EZ mode (‘0’) or EZ mode (‘1’). In EZ mode, a meta protocol is applied to the serial interface protocol. This meta protocol adds meaning to the data frames transferred by the serial interface protocol: a data frame can represent a memory address, a write memory data element or a read memory data element. EZ mode is only used for synchronous serial interface protocols: SPI and I2C. In SPI mode, only Motorola submode (all Motorola modes: 0, 1, 2, 3) is supported and the transmitter should use continuous data frames; i.e. data frames mot separated by slave deselection. This mode is only applicable to slave functionality. In EZ mode, the slave can read from and write to an addressable memory structure of up to 256 bytes. In EZ mode, data frames should 8-bit in size and should be transmitted and received with the Most Significant Bit (MSB) first. In UART mode this field should be ‘0’.

MODE_R
Field MODE reader - N/A

MODE_W
Field MODE writer - N/A

OVS_R
Field OVS reader - N/A

OVS_W
Field OVS writer - N/A

R
Register CTRL reader

W
Register CTRL writer

Enums

MODE_A
N/A