Struct psoc6_01_pac::scb5::i2c_ctrl::R

pub struct R(_);

Register I2C_CTRL reader

Implementations

impl R

pub fn high_phase_ovs(&self) -> HIGH_PHASE_OVS_R

Bits 0:3 - Serial I2C interface high phase oversampling factor. HIGH_PHASE_OVS + 1 peripheral clock periods constitute the high phase of a bit period. The valid range is [5, 15] with input signal median filtering and [4, 15] without input signal median filtering. The field is only used in master mode. In slave mode, the field is NOT used. However, there is a frequency requirement for the IP clock wrt. the regular interface (IF) high time to guarantee functional correct behavior. With input signal median filtering, the IF high time should be >= 6 IP clock cycles and <= 16 IP clock cycles. Without input signal median filtering, the IF high time should be >= 5 IP clock cycles and <= 16 IP clock cycles.

pub fn low_phase_ovs(&self) -> LOW_PHASE_OVS_R

Bits 4:7 - Serial I2C interface low phase oversampling factor. LOW_PHASE_OVS + 1 peripheral clock periods constitute the low phase of a bit period. The valid range is [7, 15] with input signal median filtering and [6, 15] without input signal median filtering. The field is only used in master mode. In slave mode, the field is NOT used. However, there is a frequency requirement for the IP clock wrt. the regular (no stretching) interface (IF) low time to guarantee functional correct behavior. With input signal median filtering, the IF low time should be >= 8 IP clock cycles and <= 16 IP clock cycles. Without input signal median filtering, the IF low time should be >= 7 IP clock cycles and <= 16 IP clock cycles.

pub fn m_ready_data_ack(&self) -> M_READY_DATA_ACK_R

Bit 8 - When ‘1’, a received data element by the master is immediately ACK’d when the receiver FIFO is not full.

pub fn m_not_ready_data_nack(&self) -> M_NOT_READY_DATA_NACK_R

Bit 9 - When ‘1’, a received data element byte the master is immediately NACK’d when the receiver FIFO is full. When ‘0’, clock stretching is used instead (till the receiver FIFO is no longer full).

pub fn s_general_ignore(&self) -> S_GENERAL_IGNORE_R

Bit 11 - When ‘1’, a received general call slave address is immediately NACK’d (no ACK or clock stretching) and treated as a non matching slave address. This is useful for slaves that do not need any data supplied within the general call structure.

pub fn s_ready_addr_ack(&self) -> S_READY_ADDR_ACK_R

Bit 12 - When ‘1’, a received (matching) slave address is immediately ACK’d when the receiver FIFO is not full. In EZ mode, this field should be set to ‘1’.

pub fn s_ready_data_ack(&self) -> S_READY_DATA_ACK_R

Bit 13 - When ‘1’, a received data element by the slave is immediately ACK’d when the receiver FIFO is not full. In EZ mode, this field should be set to ‘1’.

pub fn s_not_ready_addr_nack(&self) -> S_NOT_READY_ADDR_NACK_R

Bit 14 - For internally clocked logic (EC_AM is ‘0’ and EC_OP is ‘0’) on an address match or general call address (and S_GENERAL_IGNORE is ‘0’). Only used when: - EC_AM is ‘0’, EC_OP is ‘0’ and non EZ mode. Functionality is as follows: - 1: a received (matching) slave address is immediately NACK’d when the receiver FIFO is full. - 0: clock stretching is performed (till the receiver FIFO is no longer full). For externally clocked logic (EC_AM is ‘1’) on an address match or general call address (and S_GENERAL_IGNORE is ‘0’). Only used when (NOT used when EC_AM is ‘1’ and EC_OP is ‘1’ and address match and EZ mode): - EC_AM is ‘1’ and EC_OP is ‘0’. - EC_AM is ‘1’ and general call address match. - EC_AM is ‘1’ and non EZ mode. Functionality is as follows: - 1: a received (matching or general) slave address is always immediately NACK’d. There are two possibilities: 1). the internally clocked logic is enabled (we are in Active system power mode) and it handles the rest of the current transfer. In this case the I2C master will not observe the NACK. 2). the internally clocked logic is not enabled (we are in DeepSleep system power mode). In this case the I2C master will observe the NACK and may retry the transfer in the future (which gives the internally clocked logic the time to wake up from DeepSleep system power mode). - 0: clock stretching is performed (till the internally clocked logic takes over). The internally clocked logic will handle the ongoing transfer as soon as it is enabled.

pub fn s_not_ready_data_nack(&self) -> S_NOT_READY_DATA_NACK_R

Bit 15 - For internally clocked logic only. Only used when: - non EZ mode. Functionality is as follows: - 1: a received data element byte the slave is immediately NACK’d when the receiver FIFO is full. - 0: clock stretching is performed (till the receiver FIFO is no longer full).

pub fn loopback(&self) -> LOOPBACK_R

Bit 16 - Local loopback control (does NOT affect the information on the pins). Only applicable in master/slave mode. When ‘0’, the I2C SCL and SDA lines are connected to the I2C SCL and SDA pins. When ‘1’, I2C SCL and SDA lines are routed internally in the peripheral, and as a result unaffected by other I2C devices. This allows a SCB I2C peripheral to address itself.

pub fn slave_mode(&self) -> SLAVE_MODE_R

Bit 30 - Slave mode enabled (‘1’) or not (‘0’).

pub fn master_mode(&self) -> MASTER_MODE_R

Bit 31 - Master mode enabled (‘1’) or not (‘0’). Note that both master and slave modes can be enabled at the same time. This allows the IP to address itself.