Module psoc6_01_pac::scb5::intr_rx

Receiver interrupt request

Structs

BAUD_DETECT_R
Field BAUD_DETECT reader - LIN baudrate detection is completed. The receiver software uses the UART_RX_STATUS.BR_COUNTER value to set the right IP clock (from the programmable clock IP) to guarantee successful receipt of the first LIN data frame (Protected Identifier Field) after the synchronization byte. Set to ‘1’, when event is detected. Write with ‘1’ to clear bit.

BAUD_DETECT_W
Field BAUD_DETECT writer - LIN baudrate detection is completed. The receiver software uses the UART_RX_STATUS.BR_COUNTER value to set the right IP clock (from the programmable clock IP) to guarantee successful receipt of the first LIN data frame (Protected Identifier Field) after the synchronization byte. Set to ‘1’, when event is detected. Write with ‘1’ to clear bit.

BLOCKED_R
Field BLOCKED reader - AHB-Lite read transfer can not get access to the EZ memory (EZ_DATA accesses), due to an externally clocked EZ access. This may happen when STATUS.EC_BUSY is ‘1’.

BLOCKED_W
Field BLOCKED writer - AHB-Lite read transfer can not get access to the EZ memory (EZ_DATA accesses), due to an externally clocked EZ access. This may happen when STATUS.EC_BUSY is ‘1’.

BREAK_DETECT_R
Field BREAK_DETECT reader - Break detection is successful: the line is ‘0’ for UART_RX_CTRL.BREAK_WIDTH + 1 bit period. Can occur at any time to address unanticipated break fields; i.e. ‘break-in-data’ is supported. This feature is supported for the UART standard and LIN submodes. For the UART standard submodes, ongoing receipt of data frames is NOT affected; i.e. Firmware is expected to take the proper action. For the LIN submode, possible ongoing receipt of a data frame is stopped and the (partially) received data frame is dropped and baud rate detection is started. Set to ‘1’, when event is detected. Write with ‘1’ to clear bit.

BREAK_DETECT_W
Field BREAK_DETECT writer - Break detection is successful: the line is ‘0’ for UART_RX_CTRL.BREAK_WIDTH + 1 bit period. Can occur at any time to address unanticipated break fields; i.e. ‘break-in-data’ is supported. This feature is supported for the UART standard and LIN submodes. For the UART standard submodes, ongoing receipt of data frames is NOT affected; i.e. Firmware is expected to take the proper action. For the LIN submode, possible ongoing receipt of a data frame is stopped and the (partially) received data frame is dropped and baud rate detection is started. Set to ‘1’, when event is detected. Write with ‘1’ to clear bit.

FRAME_ERROR_R
Field FRAME_ERROR reader - Frame error in received data frame. Set to ‘1’, when event is detected. Write with ‘1’ to clear bit. This can be either a start or stop bit(s) error: Start bit error: after the detection of the beginning of a start bit period (RX line changes from ‘1’ to ‘0’), the middle of the start bit period is sampled erroneously (RX line is ‘1’). Note: a start bit error is detected BEFORE a data frame is received. Stop bit error: the RX line is sampled as ‘0’, but a ‘1’ was expected. Note: a stop bit error may result in failure to receive successive data frame(s). Note: a stop bit error is detected AFTER a data frame is received. A stop bit error is detected after a data frame is received, and the UART_RX_CTL.DROP_ON_FRAME_ERROR field specifies whether the received frame is dropped or send to the RX FIFO. If UART_RX_CTL.DROP_ON_FRAME_ERROR is ‘1’, the received data frame is dropped. If UART_RX_CTL.DROP_ON_FRAME_ERROR is ‘0’, the received data frame is send to the RX FIFO. Note that Firmware can only identify the erroneous data frame in the RX FIFO if it is fast enough to read the data frame before the hardware writes a next data frame into the RX FIFO; i.e. the RX FIFO does not have error flags to tag erroneous data frames.

FRAME_ERROR_W
Field FRAME_ERROR writer - Frame error in received data frame. Set to ‘1’, when event is detected. Write with ‘1’ to clear bit. This can be either a start or stop bit(s) error: Start bit error: after the detection of the beginning of a start bit period (RX line changes from ‘1’ to ‘0’), the middle of the start bit period is sampled erroneously (RX line is ‘1’). Note: a start bit error is detected BEFORE a data frame is received. Stop bit error: the RX line is sampled as ‘0’, but a ‘1’ was expected. Note: a stop bit error may result in failure to receive successive data frame(s). Note: a stop bit error is detected AFTER a data frame is received. A stop bit error is detected after a data frame is received, and the UART_RX_CTL.DROP_ON_FRAME_ERROR field specifies whether the received frame is dropped or send to the RX FIFO. If UART_RX_CTL.DROP_ON_FRAME_ERROR is ‘1’, the received data frame is dropped. If UART_RX_CTL.DROP_ON_FRAME_ERROR is ‘0’, the received data frame is send to the RX FIFO. Note that Firmware can only identify the erroneous data frame in the RX FIFO if it is fast enough to read the data frame before the hardware writes a next data frame into the RX FIFO; i.e. the RX FIFO does not have error flags to tag erroneous data frames.

FULL_R
Field FULL reader - RX FIFO is full. Note that received data frames are lost when the RX FIFO is full. Dependent on CTRL.BYTE_MODE: (FF_DATA_NR = EZ_DATA_NR/2) BYTE_MODE is ‘0’: # entries == FF_DATA_NR/2. BYTE_MODE is ‘1’: # entries == FF_DATA_NR. Only used in FIFO mode.

FULL_W
Field FULL writer - RX FIFO is full. Note that received data frames are lost when the RX FIFO is full. Dependent on CTRL.BYTE_MODE: (FF_DATA_NR = EZ_DATA_NR/2) BYTE_MODE is ‘0’: # entries == FF_DATA_NR/2. BYTE_MODE is ‘1’: # entries == FF_DATA_NR. Only used in FIFO mode.

INTR_RX_SPEC
Receiver interrupt request

NOT_EMPTY_R
Field NOT_EMPTY reader - RX FIFO is not empty. Only used in FIFO mode.

NOT_EMPTY_W
Field NOT_EMPTY writer - RX FIFO is not empty. Only used in FIFO mode.

OVERFLOW_R
Field OVERFLOW reader - Attempt to write to a full RX FIFO. Note: in I2C mode, the OVERFLOW is set when a data frame is received and the RX FIFO is full, independent of whether it is ACK’d or NACK’d. Only used in FIFO mode.

OVERFLOW_W
Field OVERFLOW writer - Attempt to write to a full RX FIFO. Note: in I2C mode, the OVERFLOW is set when a data frame is received and the RX FIFO is full, independent of whether it is ACK’d or NACK’d. Only used in FIFO mode.

PARITY_ERROR_R
Field PARITY_ERROR reader - Parity error in received data frame. Set to ‘1’, when event is detected. Write with ‘1’ to clear bit. If UART_RX_CTL.DROP_ON_PARITY_ERROR is ‘1’, the received frame is dropped. If UART_RX_CTL.DROP_ON_PARITY_ERROR is ‘0’, the received frame is send to the RX FIFO. In SmartCard submode, negatively acknowledged data frames generate a parity error. Note that Firmware can only identify the erroneous data frame in the RX FIFO if it is fast enough to read the data frame before the hardware writes a next data frame into the RX FIFO.

PARITY_ERROR_W
Field PARITY_ERROR writer - Parity error in received data frame. Set to ‘1’, when event is detected. Write with ‘1’ to clear bit. If UART_RX_CTL.DROP_ON_PARITY_ERROR is ‘1’, the received frame is dropped. If UART_RX_CTL.DROP_ON_PARITY_ERROR is ‘0’, the received frame is send to the RX FIFO. In SmartCard submode, negatively acknowledged data frames generate a parity error. Note that Firmware can only identify the erroneous data frame in the RX FIFO if it is fast enough to read the data frame before the hardware writes a next data frame into the RX FIFO.

R
Register INTR_RX reader

TRIGGER_R
Field TRIGGER reader - More entries in the RX FIFO than the value specified by RX_FIFO_CTRL.TRIGGER_LEVEL. Only used in FIFO mode.

TRIGGER_W
Field TRIGGER writer - More entries in the RX FIFO than the value specified by RX_FIFO_CTRL.TRIGGER_LEVEL. Only used in FIFO mode.

UNDERFLOW_R
Field UNDERFLOW reader - Attempt to read from an empty RX FIFO. Only used in FIFO mode.

UNDERFLOW_W
Field UNDERFLOW writer - Attempt to read from an empty RX FIFO. Only used in FIFO mode.

W
Register INTR_RX writer