Module psoc6_01_pac::scb5::rx_fifo_status

Receiver FIFO status

Structs

R
Register RX_FIFO_STATUS reader

RD_PTR_R
Field RD_PTR reader - FIFO read pointer: FIFO location from which a data frame is read.

RX_FIFO_STATUS_SPEC
Receiver FIFO status

SR_VALID_R
Field SR_VALID reader - Indicates whether the RX shift registers holds a (partial) valid data frame (‘1’) or not (‘0’). The shift register can be considered the bottom of the RX FIFO (the data frame is not included in the USED field of the RX FIFO). The shift register is a working register and holds the data frame that is currently being received (when the protocol state machine is receiving a data frame).

USED_R
Field USED reader - Amount of entries in the receiver FIFO. The value of this field ranges from 0 to FF_DATA_NR (EZ_DATA_NR/2).

WR_PTR_R
Field WR_PTR reader - FIFO write pointer: FIFO location at which a new data frame is written by the hardware.