Module psoc6_01_pac::sdhc0::core::adma_err_stat_r

ADMA Error Status Register

Structs

ADMA_ERR_STATES_R
Field ADMA_ERR_STATES reader - ADMA Error States These bits indicate the state of ADMA when an error occurs during ADMA data transfer. Values: - 0x0 (ST_STOP): Stop DMA - SYS_ADR register points to a location next to the error descriptor - 0x1 (ST_FDS): Fetch Descriptor - SYS_ADR register points to the error descriptor - 0x2 (UNUSED): Never set this state - 0x3 (ST_TFR): Transfer Data - SYS_ADR register points to a location next to the error descriptor

ADMA_ERR_STAT_R_SPEC
ADMA Error Status Register

ADMA_LEN_ERR_R
Field ADMA_LEN_ERR reader - ADMA Length Mismatch Error States This error occurs in the following instances: - While the Block Count Enable is being set, the total data length specified by the Descriptor table is different from that specified by the Block Count and Block Length - When the total data length cannot be divided by the block length Values: - 0x0 (NO_ERR): No Error - 0x1 (ERROR): Error

R
Register ADMA_ERR_STAT_R reader