Module psoc6_01_pac::sdhc0::core::auto_cmd_stat_r

Auto CMD Status Register

Structs

AUTO_CMD12_NOT_EXEC_R
Field AUTO_CMD12_NOT_EXEC reader - Auto CMD12 Not Executed If multiple memory block data transfer is not started due to a command error, this bit is not set because it is not necessary to issue an Auto CMD12. Setting this bit to 1 means that the Host Controller cannot issue Auto CMD12 to stop multiple memory block data transfer, due to some error. If this bit is set to 1, error status bits (D04-D01) is meaningless. This bit is set to 0 when Auto CMD Error is generated by Auto CMD23. Values: - 0x1 (TRUE): Not Executed - 0x0 (FALSE): Executed

AUTO_CMD_CRC_ERR_R
Field AUTO_CMD_CRC_ERR reader - Auto CMD CRC Error This bit is set when detecting a CRC error in the command response. Values: - 0x1 (TRUE): CRC Error Generated - 0x0 (FALSE): No Error

AUTO_CMD_EBIT_ERR_R
Field AUTO_CMD_EBIT_ERR reader - Auto CMD End Bit Error This bit is set when detecting that the end bit of command response is 0. Values: - 0x1 (TRUE): End Bit Error Generated - 0x0 (FALSE): No Error

AUTO_CMD_IDX_ERR_R
Field AUTO_CMD_IDX_ERR reader - Auto CMD Index Error This bit is set if the command index error occurs in response to a command. Values: - 0x1 (TRUE): Error - 0x0 (FALSE): No Error

AUTO_CMD_RESP_ERR_R
Field AUTO_CMD_RESP_ERR reader - Auto CMD Response Error This bit is set when Response Error Check Enable in the Transfer Mode register is set to 1 and an error is detected in R1 response of either Auto CMD12 or CMD13. This status is ignored if any bit between D00 to D04 is set to 1. Values: - 0x1 (TRUE): Error - 0x0 (FALSE): No Error

AUTO_CMD_STAT_R_SPEC
Auto CMD Status Register

AUTO_CMD_TOUT_ERR_R
Field AUTO_CMD_TOUT_ERR reader - Auto CMD Timeout Error This bit is set if no response is returned with 64 SDCLK cycles from the end bit of the command. If this bit is set to 1, error status bits (D04-D01) are meaningless. Values: - 0x1 (TRUE): Time out - 0x0 (FALSE): No Error

CMD_NOT_ISSUED_AUTO_CMD12_R
Field CMD_NOT_ISSUED_AUTO_CMD12 reader - Command Not Issued By Auto CMD12 Error If this bit is set to 1, CMD_wo_DAT is not executed due to an Auto CMD12 Error (D04-D01) in this register. This bit is set to 0 when Auto CMD Error is generated by Auto CMD23. Values: - 0x1 (TRUE): Not Issued - 0x0 (FALSE): No Error

R
Register AUTO_CMD_STAT_R reader