Module psoc6_01_pac::sdhc0::core::bgap_ctrl_r

Block Gap Control Register

Structs

BGAP_CTRL_R_SPEC
Block Gap Control Register

CONTINUE_REQ_R
Field CONTINUE_REQ reader - Continue Request This bit is used to restart the transaction, which was stopped using the Stop At Block Gap Request. The Host Controller automatically clears this bit when the transaction restarts. If stop at block gap request is set to 1, any write to this bit is ignored. Values: - 0x0 (NO_AFFECT): No Affect - 0x1 (RESTART): Restart

CONTINUE_REQ_W
Field CONTINUE_REQ writer - Continue Request This bit is used to restart the transaction, which was stopped using the Stop At Block Gap Request. The Host Controller automatically clears this bit when the transaction restarts. If stop at block gap request is set to 1, any write to this bit is ignored. Values: - 0x0 (NO_AFFECT): No Affect - 0x1 (RESTART): Restart

INT_AT_BGAP_R
Field INT_AT_BGAP reader - Interrupt At Block Gap This bit is valid only in the 4-bit mode of an SDIO card and is used to select a sample point in the interrupt cycle. Setting to 1 enables interrupt detection at the block gap for a multiple block transfer. Values: - 0x0 (DISABLE): Disabled - 0x1 (ENABLE): Enabled

INT_AT_BGAP_W
Field INT_AT_BGAP writer - Interrupt At Block Gap This bit is valid only in the 4-bit mode of an SDIO card and is used to select a sample point in the interrupt cycle. Setting to 1 enables interrupt detection at the block gap for a multiple block transfer. Values: - 0x0 (DISABLE): Disabled - 0x1 (ENABLE): Enabled

R
Register BGAP_CTRL_R reader

RD_WAIT_CTRL_R
Field RD_WAIT_CTRL reader - N/A

RD_WAIT_CTRL_W
Field RD_WAIT_CTRL writer - N/A

STOP_BG_REQ_R
Field STOP_BG_REQ reader - Stop At Block Gap Request This bit is used to stop executing read and write transactions at the next block gap for non-DMA, SDMA, and ADMA transfers. Values: - 0x0 (XFER): Transfer - 0x1 (STOP): Stop

STOP_BG_REQ_W
Field STOP_BG_REQ writer - Stop At Block Gap Request This bit is used to stop executing read and write transactions at the next block gap for non-DMA, SDMA, and ADMA transfers. Values: - 0x0 (XFER): Transfer - 0x1 (STOP): Stop

W
Register BGAP_CTRL_R writer