Module psoc6_01_pac::sdhc0::core::boot_ctrl_r

eMMC Boot Control register

Structs

BOOT_ACK_ENABLE_R
Field BOOT_ACK_ENABLE reader - Boot Acknowledge Enable When this bit set, SDHC checks for boot acknowledge start pattern of 0-1-0 during boot operation. This bit is applicable for both mandatory and alternate boot mode. Values: - 0x1 (TRUE): Boot Ack enable - 0x0 (FALSE): Boot Ack disable

BOOT_ACK_ENABLE_W
Field BOOT_ACK_ENABLE writer - Boot Acknowledge Enable When this bit set, SDHC checks for boot acknowledge start pattern of 0-1-0 during boot operation. This bit is applicable for both mandatory and alternate boot mode. Values: - 0x1 (TRUE): Boot Ack enable - 0x0 (FALSE): Boot Ack disable

BOOT_CTRL_R_SPEC
eMMC Boot Control register

BOOT_TOUT_CNT_R
Field BOOT_TOUT_CNT reader - N/A

BOOT_TOUT_CNT_W
Field BOOT_TOUT_CNT writer - N/A

MAN_BOOT_EN_R
Field MAN_BOOT_EN reader - Mandatory Boot Enable This bit is used to initiate the mandatory boot operation. The application sets this bit along with VALIDATE_BOOT bit. Writing 0 is ignored. The SDHC clears this bit after the boot transfer is completed or terminated. Values: - 0x1 (MAN_BOOT_EN): Mandatory boot enable - 0x0 (MAN_BOOT_DIS): Mandatory boot disable

MAN_BOOT_EN_W
Field MAN_BOOT_EN writer - Mandatory Boot Enable This bit is used to initiate the mandatory boot operation. The application sets this bit along with VALIDATE_BOOT bit. Writing 0 is ignored. The SDHC clears this bit after the boot transfer is completed or terminated. Values: - 0x1 (MAN_BOOT_EN): Mandatory boot enable - 0x0 (MAN_BOOT_DIS): Mandatory boot disable

R
Register BOOT_CTRL_R reader

VALIDATE_BOOT_W
Field VALIDATE_BOOT writer - Validate Mandatory Boot Enable bit This bit is used to validate the MAN_BOOT_EN bit. Values: - 0x1 (TRUE): Validate Mandatory boot enable bit - 0x0 (FALSE): Ignore Mandatory boot Enable bit

W
Register BOOT_CTRL_R writer