Module psoc6_01_pac::sdhc0::core::capabilities2_r

Capabilities Register - 32 to 63

Structs

ADMA3_SUPPORT_R
Field ADMA3_SUPPORT reader - ADMA3 Support This bit indicates whether the Host Controller is capable of using ADMA3. Values: - 0x0 (FALSE): ADMA3 not Supported - 0x1 (TRUE): ADMA3 Supported

CAPABILITIES2_R_SPEC
Capabilities Register - 32 to 63

CLK_MUL_R
Field CLK_MUL reader - Clock Multiplier These bits indicate the clock multiplier of the programmable clock generator. Setting these bits to 0 means that the Host Controller does not support a programmable clock generator. - 0x0: Clock Multiplier is not Supported - 0x1: Clock Multiplier M = 2 - 0x2: Clock Multiplier M = 3 - ……… - 0xFF: Clock Multiplier M = 256

DDR50_SUPPORT_R
Field DDR50_SUPPORT reader - DDR50 Support (UHS-I only) Values: - 0x0 (FALSE): DDR50 is not supported - 0x1 (TRUE): DDR50 is supported

DRV_TYPEA_R
Field DRV_TYPEA reader - Driver Type A Support (UHS-I only) This bit indicates support of Driver Type A for 1.8 Signaling. Values: - 0x0 (FALSE): Driver Type A is not supported - 0x1 (TRUE): Driver Type A is supported

DRV_TYPEC_R
Field DRV_TYPEC reader - Driver Type C Support (UHS-I only) This bit indicates support of Driver Type C for 1.8 Signaling. Values: - 0x0 (FALSE): Driver Type C is not supported - 0x1 (TRUE): Driver Type C is supported

DRV_TYPED_R
Field DRV_TYPED reader - Driver Type D Support (UHS-I only) This bit indicates support of Driver Type D for 1.8 Signaling. Values: - 0x0 (FALSE): Driver Type D is not supported - 0x1 (TRUE): Driver Type D is supported

R
Register CAPABILITIES2_R reader

RETUNE_CNT_R
Field RETUNE_CNT reader - N/A

RE_TUNING_MODES_R
Field RE_TUNING_MODES reader - N/A

SDR50_SUPPORT_R
Field SDR50_SUPPORT reader - SDR50 Support (UHS-I only) Thsi bit indicates that SDR50 is supported. The bit 13 (USE_TUNING_SDR50) indicates whether SDR50 requires tuning or not. Values: - 0x0 (FALSE): SDR50 is not supported - 0x1 (TRUE): SDR50 is supported

SDR104_SUPPORT_R
Field SDR104_SUPPORT reader - SDR104 Support (UHS-I only) This bit mentions that SDR104 requires tuning. Values: - 0x0 (FALSE): SDR104 is not supported - 0x1 (TRUE): SDR104 is supported (NOT ACTUALLY SUPPORTED)

UHS2_SUPPORT_R
Field UHS2_SUPPORT reader - UHS-II Support (UHS-II only) This bit indicates whether Host Controller supports UHS-II. Values: - 0x0 (FALSE): UHS-II is not supported - 0x1 (TRUE): UHS-II is supported

USE_TUNING_SDR50_R
Field USE_TUNING_SDR50 reader - Use Tuning for SDR50 (UHS-I only) Values: - 0x0 (ZERO): SDR50 does not require tuning - 0x1 (ONE): SDR50 requires tuning

VDD2_18V_SUPPORT_R
Field VDD2_18V_SUPPORT reader - 1.8V VDD2 Support This bit indicates support of VDD2 for the Host System. Values: - 0x0 (FALSE): 1.8V VDD2 is not Supported - 0x1 (TRUE): 1.8V VDD2 is Supported