Struct psoc6_01_pac::sdhc0::core::capabilities2_r::R

pub struct R(_);

Register CAPABILITIES2_R reader

Implementations

impl R

pub fn sdr50_support(&self) -> SDR50_SUPPORT_R

Bit 0 - SDR50 Support (UHS-I only) Thsi bit indicates that SDR50 is supported. The bit 13 (USE_TUNING_SDR50) indicates whether SDR50 requires tuning or not. Values: - 0x0 (FALSE): SDR50 is not supported - 0x1 (TRUE): SDR50 is supported

pub fn sdr104_support(&self) -> SDR104_SUPPORT_R

Bit 1 - SDR104 Support (UHS-I only) This bit mentions that SDR104 requires tuning. Values: - 0x0 (FALSE): SDR104 is not supported - 0x1 (TRUE): SDR104 is supported (NOT ACTUALLY SUPPORTED)

pub fn ddr50_support(&self) -> DDR50_SUPPORT_R

Bit 2 - DDR50 Support (UHS-I only) Values: - 0x0 (FALSE): DDR50 is not supported - 0x1 (TRUE): DDR50 is supported

pub fn uhs2_support(&self) -> UHS2_SUPPORT_R

Bit 3 - UHS-II Support (UHS-II only) This bit indicates whether Host Controller supports UHS-II. Values: - 0x0 (FALSE): UHS-II is not supported - 0x1 (TRUE): UHS-II is supported

pub fn drv_typea(&self) -> DRV_TYPEA_R

Bit 4 - Driver Type A Support (UHS-I only) This bit indicates support of Driver Type A for 1.8 Signaling. Values: - 0x0 (FALSE): Driver Type A is not supported - 0x1 (TRUE): Driver Type A is supported

pub fn drv_typec(&self) -> DRV_TYPEC_R

Bit 5 - Driver Type C Support (UHS-I only) This bit indicates support of Driver Type C for 1.8 Signaling. Values: - 0x0 (FALSE): Driver Type C is not supported - 0x1 (TRUE): Driver Type C is supported

pub fn drv_typed(&self) -> DRV_TYPED_R

Bit 6 - Driver Type D Support (UHS-I only) This bit indicates support of Driver Type D for 1.8 Signaling. Values: - 0x0 (FALSE): Driver Type D is not supported - 0x1 (TRUE): Driver Type D is supported

pub fn retune_cnt(&self) -> RETUNE_CNT_R

Bits 8:11 - N/A

pub fn use_tuning_sdr50(&self) -> USE_TUNING_SDR50_R

Bit 13 - Use Tuning for SDR50 (UHS-I only) Values: - 0x0 (ZERO): SDR50 does not require tuning - 0x1 (ONE): SDR50 requires tuning

pub fn re_tuning_modes(&self) -> RE_TUNING_MODES_R

Bits 14:15 - N/A

pub fn clk_mul(&self) -> CLK_MUL_R

Bits 16:23 - Clock Multiplier These bits indicate the clock multiplier of the programmable clock generator. Setting these bits to 0 means that the Host Controller does not support a programmable clock generator. - 0x0: Clock Multiplier is not Supported - 0x1: Clock Multiplier M = 2 - 0x2: Clock Multiplier M = 3 - ……… - 0xFF: Clock Multiplier M = 256

pub fn adma3_support(&self) -> ADMA3_SUPPORT_R

Bit 27 - ADMA3 Support This bit indicates whether the Host Controller is capable of using ADMA3. Values: - 0x0 (FALSE): ADMA3 not Supported - 0x1 (TRUE): ADMA3 Supported

pub fn vdd2_18v_support(&self) -> VDD2_18V_SUPPORT_R

Bit 28 - 1.8V VDD2 Support This bit indicates support of VDD2 for the Host System. Values: - 0x0 (FALSE): 1.8V VDD2 is not Supported - 0x1 (TRUE): 1.8V VDD2 is Supported