Struct psoc6_01_pac::sdhc0::core::clk_ctrl_r::W

pub struct W(_);

Register CLK_CTRL_R writer

Implementations

impl W

pub fn internal_clk_en(&mut self) -> INTERNAL_CLK_EN_W<'_>

Bit 0 - Internal Clock Enable This bit is set to 0 when the Host Driver is not using the Host Controller or the Host Controller awaits a wakeup interrupt. The Host Controller must stop its internal clock to enter a very low power state. Certain registers are not accessible when this bit is off. So, to be safe turn it on for any register access. Values: - 0x0 (FALSE): Stop - 0x1 (TRUE): Oscillate

pub fn sd_clk_en(&mut self) -> SD_CLK_EN_W<'_>

Bit 2 - SD/eMMC Clock Enable This bit stops the clk_card output when set to 0. The SDCLK Frequency Select bit can be changed when this bit is set to 0. Values: - 0x0 (FALSE): Disable providing clk_card - 0x1 (TRUE): Enable providing clk_card

pub fn pll_enable(&mut self) -> PLL_ENABLE_W<'_>

Bit 3 - PLL Enable This bit is used to activate the PLL (applicable when Host Version 4 Enable = 1). Values: - 0x0 (FALSE): PLL is in low power mode - 0x1 (TRUE): PLL is enabled

pub fn clk_gen_select(&mut self) -> CLK_GEN_SELECT_W<'_>

Bit 5 - Clock Generator Select This bit is used to select the clock generator mode in SDCLK Frequency Select. Values: - 0x0 (FALSE): Divided Clock Mode - 0x1 (TRUE): Programmable Clock Mode

pub fn upper_freq_sel(&mut self) -> UPPER_FREQ_SEL_W<'_>

Bits 6:7 - These bits specify the upper 2 bits of 10-bit SDCLK Frequency Select control.

pub fn freq_sel(&mut self) -> FREQ_SEL_W<'_>

Bits 8:15 - SDCLK Frequency Select These bits are used to select the frequency of the SDCLK signal. 10-bit Divided Clock Mode: - 0x3FF - 1/2046 Divided clock - ………. - N - 1/2N Divided Clock - ………. - 0x002 - 1/4 Divided Clock - 0x001 - 1/2 Divided Clock - 0x000 - Base clock (10MHz - 255 MHz)

pub unsafe fn bits(&mut self, bits: u16) -> &mut Self

Writes raw bits to the register.