Module psoc6_01_pac::sdhc0::core::error_int_stat_en_r

Error Interrupt Status Enable Register

Structs

ADMA_ERR_STAT_EN_R
Field ADMA_ERR_STAT_EN reader - ADMA Error Status Enable Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

ADMA_ERR_STAT_EN_W
Field ADMA_ERR_STAT_EN writer - ADMA Error Status Enable Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

AUTO_CMD_ERR_STAT_EN_R
Field AUTO_CMD_ERR_STAT_EN reader - Auto CMD Error Status Enable (SD/eMMC Mode only). Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

AUTO_CMD_ERR_STAT_EN_W
Field AUTO_CMD_ERR_STAT_EN writer - Auto CMD Error Status Enable (SD/eMMC Mode only). Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

BOOT_ACK_ERR_STAT_EN_R
Field BOOT_ACK_ERR_STAT_EN reader - Boot Acknowledgment Error (eMMC Mode only) Setting this bit to 1 enables setting of Boot Acknowledgment Error in Error Interrupt Status register (ERROR_INT_STAT_R). Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

BOOT_ACK_ERR_STAT_EN_W
Field BOOT_ACK_ERR_STAT_EN writer - Boot Acknowledgment Error (eMMC Mode only) Setting this bit to 1 enables setting of Boot Acknowledgment Error in Error Interrupt Status register (ERROR_INT_STAT_R). Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

CMD_CRC_ERR_STAT_EN_R
Field CMD_CRC_ERR_STAT_EN reader - ommand CRC Error Status Enable (SD/eMMC Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

CMD_CRC_ERR_STAT_EN_W
Field CMD_CRC_ERR_STAT_EN writer - ommand CRC Error Status Enable (SD/eMMC Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

CMD_END_BIT_ERR_STAT_EN_R
Field CMD_END_BIT_ERR_STAT_EN reader - Command End Bit Error Status Enable (SD/eMMC Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

CMD_END_BIT_ERR_STAT_EN_W
Field CMD_END_BIT_ERR_STAT_EN writer - Command End Bit Error Status Enable (SD/eMMC Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

CMD_IDX_ERR_STAT_EN_R
Field CMD_IDX_ERR_STAT_EN reader - Command Index Error Status Enable (SD/eMMC Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

CMD_IDX_ERR_STAT_EN_W
Field CMD_IDX_ERR_STAT_EN writer - Command Index Error Status Enable (SD/eMMC Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

CMD_TOUT_ERR_STAT_EN_R
Field CMD_TOUT_ERR_STAT_EN reader - Command Timeout Error Status Enable (SD/eMMC Mode only). Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

CMD_TOUT_ERR_STAT_EN_W
Field CMD_TOUT_ERR_STAT_EN writer - Command Timeout Error Status Enable (SD/eMMC Mode only). Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

CUR_LMT_ERR_STAT_EN_R
Field CUR_LMT_ERR_STAT_EN reader - Current Limit Error Status Enable Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

CUR_LMT_ERR_STAT_EN_W
Field CUR_LMT_ERR_STAT_EN writer - Current Limit Error Status Enable Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

DATA_CRC_ERR_STAT_EN_R
Field DATA_CRC_ERR_STAT_EN reader - Data CRC Error Status Enable (SD/eMMC Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

DATA_CRC_ERR_STAT_EN_W
Field DATA_CRC_ERR_STAT_EN writer - Data CRC Error Status Enable (SD/eMMC Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

DATA_END_BIT_ERR_STAT_EN_R
Field DATA_END_BIT_ERR_STAT_EN reader - Data End Bit Error Status Enable (SD/eMMC Mode only). Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

DATA_END_BIT_ERR_STAT_EN_W
Field DATA_END_BIT_ERR_STAT_EN writer - Data End Bit Error Status Enable (SD/eMMC Mode only). Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

DATA_TOUT_ERR_STAT_EN_R
Field DATA_TOUT_ERR_STAT_EN reader - Data Timeout Error Status Enable (SD/eMMC Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

DATA_TOUT_ERR_STAT_EN_W
Field DATA_TOUT_ERR_STAT_EN writer - Data Timeout Error Status Enable (SD/eMMC Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

ERROR_INT_STAT_EN_R_SPEC
Error Interrupt Status Enable Register

R
Register ERROR_INT_STAT_EN_R reader

RESP_ERR_STAT_EN_R
Field RESP_ERR_STAT_EN reader - Response Error Status Enable (SD Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

RESP_ERR_STAT_EN_W
Field RESP_ERR_STAT_EN writer - Response Error Status Enable (SD Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

TUNING_ERR_STAT_EN_R
Field TUNING_ERR_STAT_EN reader - Tuning Error Status Enable (UHS-I Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

TUNING_ERR_STAT_EN_W
Field TUNING_ERR_STAT_EN writer - Tuning Error Status Enable (UHS-I Mode only) Values: - 0x0 (FALSE): Masked - 0x1 (TRUE): Enabled

VENDOR_ERR_STAT_EN1_R
Field VENDOR_ERR_STAT_EN1 reader - N/A

VENDOR_ERR_STAT_EN1_W
Field VENDOR_ERR_STAT_EN1 writer - N/A

VENDOR_ERR_STAT_EN2_R
Field VENDOR_ERR_STAT_EN2 reader - N/A

VENDOR_ERR_STAT_EN2_W
Field VENDOR_ERR_STAT_EN2 writer - N/A

VENDOR_ERR_STAT_EN3_R
Field VENDOR_ERR_STAT_EN3 reader - N/A

VENDOR_ERR_STAT_EN3_W
Field VENDOR_ERR_STAT_EN3 writer - N/A

W
Register ERROR_INT_STAT_EN_R writer