Module psoc6_01_pac::sdhc0::core::gp_out_r

General Purpose Output register

Structs

CARD_CLOCK_IN_DLY_R
Field CARD_CLOCK_IN_DLY reader - Delay CARD_CLOCK input internally to optimally sample CMD/DAT; set according to interface mode: 00: SD Default Speed, SD SDR12, eMMC Legacy 01: SD SDR25, SD SDR50 10: SD High Speed, eMMC High Speed SDR 11: SD DDR50, eMMC DDR

CARD_CLOCK_IN_DLY_W
Field CARD_CLOCK_IN_DLY writer - Delay CARD_CLOCK input internally to optimally sample CMD/DAT; set according to interface mode: 00: SD Default Speed, SD SDR12, eMMC Legacy 01: SD SDR25, SD SDR50 10: SD High Speed, eMMC High Speed SDR 11: SD DDR50, eMMC DDR

CARD_CLOCK_OE_R
Field CARD_CLOCK_OE reader - Active high output enable for the card clock output (clk_card) which is gated by CLK_CTRL_R.SD_CLK_EN: 0: disable OE to the clk_card output 1: enable OE to the clk_card output

CARD_CLOCK_OE_W
Field CARD_CLOCK_OE writer - Active high output enable for the card clock output (clk_card) which is gated by CLK_CTRL_R.SD_CLK_EN: 0: disable OE to the clk_card output 1: enable OE to the clk_card output

CARD_CLOCK_OUT_DLY_R
Field CARD_CLOCK_OUT_DLY reader - N/A

CARD_CLOCK_OUT_DLY_W
Field CARD_CLOCK_OUT_DLY writer - N/A

CARD_DETECT_EN_R
Field CARD_DETECT_EN reader - 0: Force card_detect_n input to 0 1: Normal card_detect_n operation allowing card detection from a device pin

CARD_DETECT_EN_W
Field CARD_DETECT_EN writer - 0: Force card_detect_n input to 0 1: Normal card_detect_n operation allowing card detection from a device pin

CARD_IF_PWR_EN_OE_R
Field CARD_IF_PWR_EN_OE reader - Active high output enable for the card interface power enable output (card_if_pwr_en) controlled through PWR_CTRL_R.SD_BUS_PWR_VDD1: 0: disable OE to the card_if_pwr_en output 1: enable OE to the card_if_pwr_en output

CARD_IF_PWR_EN_OE_W
Field CARD_IF_PWR_EN_OE writer - Active high output enable for the card interface power enable output (card_if_pwr_en) controlled through PWR_CTRL_R.SD_BUS_PWR_VDD1: 0: disable OE to the card_if_pwr_en output 1: enable OE to the card_if_pwr_en output

CARD_MECH_WRITE_PROT_EN_R
Field CARD_MECH_WRITE_PROT_EN reader - card_mech_write_prot, despite its name, is an active low signal (per the SD Host Controller Standard spec it is officially called SDWP#). Consider that in the following: 0: Force card_mech_write_prot input to 0 internally; this forces write protection to be active 1: Allow card_mech_write_prot to work normally per the device’s pin state

CARD_MECH_WRITE_PROT_EN_W
Field CARD_MECH_WRITE_PROT_EN writer - card_mech_write_prot, despite its name, is an active low signal (per the SD Host Controller Standard spec it is officially called SDWP#). Consider that in the following: 0: Force card_mech_write_prot input to 0 internally; this forces write protection to be active 1: Allow card_mech_write_prot to work normally per the device’s pin state

GP_OUT_R_SPEC
General Purpose Output register

IO_VOLT_SEL_OE_R
Field IO_VOLT_SEL_OE reader - Active high output enable for the IO voltage selection signal (io_volt_sel) controlled through HOST_CTRL_2.SIGNALING_EN: 0: disable OE to the io_volt_sel output 1: enable OE to the io_volt_sel output

IO_VOLT_SEL_OE_W
Field IO_VOLT_SEL_OE writer - Active high output enable for the IO voltage selection signal (io_volt_sel) controlled through HOST_CTRL_2.SIGNALING_EN: 0: disable OE to the io_volt_sel output 1: enable OE to the io_volt_sel output

LED_CTRL_OE_R
Field LED_CTRL_OE reader - Active high output enable for the LED output signal (led_ctrl) controlled through HOST_CTRL1_R.LED_CTRL: 0: disable OE associated with the led_ctrl output 1: enable OE associated with the led_ctrl output

LED_CTRL_OE_W
Field LED_CTRL_OE writer - Active high output enable for the LED output signal (led_ctrl) controlled through HOST_CTRL1_R.LED_CTRL: 0: disable OE associated with the led_ctrl output 1: enable OE associated with the led_ctrl output

R
Register GP_OUT_R reader

W
Register GP_OUT_R writer