Struct psoc6_01_pac::sdhc0::core::host_ctrl2_r::R

pub struct R(_);

Register HOST_CTRL2_R reader

Implementations

impl R

pub fn uhs_mode_sel(&self) -> UHS_MODE_SEL_R

Bits 0:2 - N/A

pub fn signaling_en(&self) -> SIGNALING_EN_R

Bit 3 - 1.8V Signaling Enable This bit controls voltage regulator for I/O cell in SD UHS-I mode. Setting this bit from 0 to 1 starts changing the signal voltage from 3.3V to 1.8V. Host Controller clears this bit if switching to 1.8V signaling fails per protocol. The value is reflected on the io_volt_sel output which can then be used to change an external regulator to supply 1.8V instead of 3.3V on the VDDIO pin associated with the CLK/CMD/DAT signals. Note: This bit must be set for all UHS-I speed modes (SDR12/SDR25/SDR50/DDR50). Values: - 0x0 (V_3_3): 3.3V Signalling - 0x1 (V_1_8): 1.8V Signalling

pub fn drv_strength_sel(&self) -> DRV_STRENGTH_SEL_R

Bits 4:5 - Driver Strength Select These bits are used to select the Host Controller output driver in 1.8V signaling UHS-I/eMMC speed modes. The value is reflected on the io_drive_strength[1:0] output. - 0x0 (TYPEB): Driver TYPEB is selected - 0x1 (TYPEA): Driver TYPEA is selected - 0x2 (TYPEC): Driver TYPEC is selected - 0x3 (TYPED): Driver TYPED is selected

pub fn exec_tuning(&self) -> EXEC_TUNING_R

Bit 6 - N/A

pub fn sample_clk_sel(&self) -> SAMPLE_CLK_SEL_R

Bit 7 - N/A

pub fn uhs2_if_enable(&self) -> UHS2_IF_ENABLE_R

Bit 8 - N/A

pub fn adma2_len_mode(&self) -> ADMA2_LEN_MODE_R

Bit 10 - ADMA2 Length Mode This bit selects ADMA2 Length mode to be either 16-bit or 26-bit. Values: - 0x0 (FALSE): 16-bit Data Length Mode - 0x1 (TRUE): 26-bit Data Length Mode

pub fn cmd23_enable(&self) -> CMD23_ENABLE_R

Bit 11 - CMD23 Enable If the card supports CMD23, this bit is set to 1. This bit is used to select Auto CMD23 or Auto CMD12 for ADMA3 data transfer. Values: - 0x0 (FALSE): Auto CMD23 is disabled - 0x1 (TRUE): Auto CMD23 is enabled

pub fn host_ver4_enable(&self) -> HOST_VER4_ENABLE_R

Bit 12 - Host Version 4 Enable This bit selects either Version 3.00 compatible mode or Version 4 mode. Functions of following fields are modified for Host Version 4 mode: - SDMA Address: SDMA uses ADMA System Address (05Fh-058h) instead of SDMA System Address register (003h-000h) - ADMA2/ADMA3 selection: ADMA3 is selected by DMA select in Host Control 1 register - 32-bit Block Count: SDMA System Address register (003h-000h) is modified to 32-bit Block Count register Note: It is recommended not to program ADMA3 Integrated Descriptor Address registers and Command Queuing registers (if applicable) while operating in Host version less than 4 mode (Host Version 4 Enable = 0). Values: - 0x0 (FALSE): Version 3.00 compatible mode - 0x1 (TRUE): Version 4 mode

pub fn addressing(&self) -> ADDRESSING_R

Bit 13 - N/A

pub fn async_int_enable(&self) -> ASYNC_INT_ENABLE_R

Bit 14 - Asynchronous Interrupt Enable This bit can be set if a card supports asynchronous interrupts and Asynchronous Interrupt Support is set to 1 in the Capabilities register. Values: - 0x0 (FALSE): Disabled - 0x1 (TRUE): Enabled

pub fn preset_val_enable(&self) -> PRESET_VAL_ENABLE_R

Bit 15 - N/A