Module psoc6_01_pac::sdhc0::core::mbiu_ctrl_r

MBIU Control register

Structs

BURST_INCR4_EN_R
Field BURST_INCR4_EN reader - INCR4 Burst Controls generation of INCR4 transfers on Master interface. Values: - 0x0 (FALSE): AHB INCR4 burst type is not generated on Master I/F - 0x1 (TRUE): AHB INCR4 burst type can be generated on Master I/F

BURST_INCR4_EN_W
Field BURST_INCR4_EN writer - INCR4 Burst Controls generation of INCR4 transfers on Master interface. Values: - 0x0 (FALSE): AHB INCR4 burst type is not generated on Master I/F - 0x1 (TRUE): AHB INCR4 burst type can be generated on Master I/F

BURST_INCR8_EN_R
Field BURST_INCR8_EN reader - INCR8 Burst Controls generation of INCR8 transfers on Master interface. Values: - 0x0 (FALSE): AHB INCR8 burst type is not generated on Master I/F - 0x1 (TRUE): AHB INCR8 burst type can be generated on Master I/F

BURST_INCR8_EN_W
Field BURST_INCR8_EN writer - INCR8 Burst Controls generation of INCR8 transfers on Master interface. Values: - 0x0 (FALSE): AHB INCR8 burst type is not generated on Master I/F - 0x1 (TRUE): AHB INCR8 burst type can be generated on Master I/F

BURST_INCR16_EN_R
Field BURST_INCR16_EN reader - INCR16 Burst Controls generation of INCR16 transfers on Master interface. Values: - 0x0 (FALSE): AHB INCR16 burst type is not generated on Master I/F - 0x1 (TRUE): AHB INCR16 burst type can be generated on Master I/F

BURST_INCR16_EN_W
Field BURST_INCR16_EN writer - INCR16 Burst Controls generation of INCR16 transfers on Master interface. Values: - 0x0 (FALSE): AHB INCR16 burst type is not generated on Master I/F - 0x1 (TRUE): AHB INCR16 burst type can be generated on Master I/F

MBIU_CTRL_R_SPEC
MBIU Control register

R
Register MBIU_CTRL_R reader

UNDEFL_INCR_EN_R
Field UNDEFL_INCR_EN reader - Undefined INCR Burst Controls generation of undefined length INCR transfer on Master interface. Values: - 0x0 (FALSE): Undefined INCR type burst is the least preferred burst on AHB Master I/F - 0x1 (TRUE): Undefined INCR type burst is the most preferred burst on AHB Master I/F

UNDEFL_INCR_EN_W
Field UNDEFL_INCR_EN writer - Undefined INCR Burst Controls generation of undefined length INCR transfer on Master interface. Values: - 0x0 (FALSE): Undefined INCR type burst is the least preferred burst on AHB Master I/F - 0x1 (TRUE): Undefined INCR type burst is the most preferred burst on AHB Master I/F

W
Register MBIU_CTRL_R writer