Module psoc6_01_pac::sdhc0::core::wup_ctrl_r

Wakeup Control Register

Structs

R
Register WUP_CTRL_R reader

W
Register WUP_CTRL_R writer

WUP_CARD_INSERT_R
Field WUP_CARD_INSERT reader - Wakeup Event Enable on SD Card Insertion This bit enables wakeup event through Card Insertion assertion in the Normal Interrupt Status register. FN_WUS (Wake Up Support) in CIS does not affect this bit. Values: - 0x0 (DISABLED): Disable - 0x1 (ENABLED): Enable

WUP_CARD_INSERT_W
Field WUP_CARD_INSERT writer - Wakeup Event Enable on SD Card Insertion This bit enables wakeup event through Card Insertion assertion in the Normal Interrupt Status register. FN_WUS (Wake Up Support) in CIS does not affect this bit. Values: - 0x0 (DISABLED): Disable - 0x1 (ENABLED): Enable

WUP_CARD_INT_R
Field WUP_CARD_INT reader - Wakeup Event Enable on SDIO Card Interrupt (through DAT[1]). This bit enables wakeup event through an SDIO Card Interrupt assertion in the Normal Interrupt Status register. This bit can be set to 1 if FN_WUS (Wake Up Support) in CIS is set to 1. Values: - 0x0 (DISABLED): Disable - 0x1 (ENABLED): Enable

WUP_CARD_INT_W
Field WUP_CARD_INT writer - Wakeup Event Enable on SDIO Card Interrupt (through DAT[1]). This bit enables wakeup event through an SDIO Card Interrupt assertion in the Normal Interrupt Status register. This bit can be set to 1 if FN_WUS (Wake Up Support) in CIS is set to 1. Values: - 0x0 (DISABLED): Disable - 0x1 (ENABLED): Enable

WUP_CARD_REMOVAL_R
Field WUP_CARD_REMOVAL reader - Wakeup Event Enable on SD Card Removal This bit enables wakeup event through Card Removal assertion in the Normal Interrupt Status register. For the SDIO card, Wake Up Support (FN_WUS) in the Card Information Structure (CIS) register does not affect this bit. Values: - 0x0 (DISABLED): Disable - 0x1 (ENABLED): Enable

WUP_CARD_REMOVAL_W
Field WUP_CARD_REMOVAL writer - Wakeup Event Enable on SD Card Removal This bit enables wakeup event through Card Removal assertion in the Normal Interrupt Status register. For the SDIO card, Wake Up Support (FN_WUS) in the Card Information Structure (CIS) register does not affect this bit. Values: - 0x0 (DISABLED): Disable - 0x1 (ENABLED): Enable

WUP_CTRL_R_SPEC
Wakeup Control Register