Module psoc6_01_pac::smif0::device::addr_ctl

Address control

Structs

ADDR_CTL_SPEC
Address control

DIV2_R
Field DIV2 reader - Specifies if the AHB-Lite bus transfer address is divided by 2 or not: ‘0’: No divide by 2. ‘1’: Divide by 2. This functionality is used for read and write operation in XIP, dual quad SPI mode; i.e. this DIV2 must be set to ‘1’ in dual quad SPI mode. If the transfer request address is NOT a multiple of 2 or the requested number of Bytes is not a multiple of 2, the XIP_ALIGNMENT_ERROR interrupt cause is activated.

DIV2_W
Field DIV2 writer - Specifies if the AHB-Lite bus transfer address is divided by 2 or not: ‘0’: No divide by 2. ‘1’: Divide by 2. This functionality is used for read and write operation in XIP, dual quad SPI mode; i.e. this DIV2 must be set to ‘1’ in dual quad SPI mode. If the transfer request address is NOT a multiple of 2 or the requested number of Bytes is not a multiple of 2, the XIP_ALIGNMENT_ERROR interrupt cause is activated.

R
Register ADDR_CTL reader

SIZE2_R
Field SIZE2 reader - Specifies the size of the XIP device address in Bytes: ‘0’: 1 Byte address. ‘1’: 2 Byte address. ‘2’: 3 Byte address. ‘3’: 4 Byte address. The lower significant address Bytes of the transfer request are used as XIP address to the external device. Note that for dual quad SPI data transfer, the transfer request address is divided by 2. Therefore, the transfer request address needs to be a multiple of 2. If the trasnfer requestaddress is NOT a multiple of 2, the XIP_ALIGNMENT_ERROR interrupt cause is activated.

SIZE2_W
Field SIZE2 writer - Specifies the size of the XIP device address in Bytes: ‘0’: 1 Byte address. ‘1’: 2 Byte address. ‘2’: 3 Byte address. ‘3’: 4 Byte address. The lower significant address Bytes of the transfer request are used as XIP address to the external device. Note that for dual quad SPI data transfer, the transfer request address is divided by 2. Therefore, the transfer request address needs to be a multiple of 2. If the trasnfer requestaddress is NOT a multiple of 2, the XIP_ALIGNMENT_ERROR interrupt cause is activated.

W
Register ADDR_CTL writer