Module psoc6_01_pac::smif0::intr

Interrupt register

Structs

INTR_SPEC
Interrupt register

R
Register INTR reader

RX_DATA_FIFO_UNDERFLOW_R
Field RX_DATA_FIFO_UNDERFLOW reader - Activated in MMIO mode, on an AHB-Lite read transfer from the RX data FIFO (RX_DATA_FIFO_RD1, RX_DATA_FIFO_RD2, RX_DATA_FIFO_RD4) with not enough entries available. Only activated for NON test bus controller transfers.

RX_DATA_FIFO_UNDERFLOW_W
Field RX_DATA_FIFO_UNDERFLOW writer - Activated in MMIO mode, on an AHB-Lite read transfer from the RX data FIFO (RX_DATA_FIFO_RD1, RX_DATA_FIFO_RD2, RX_DATA_FIFO_RD4) with not enough entries available. Only activated for NON test bus controller transfers.

TR_RX_REQ_R
Field TR_RX_REQ reader - Activated in MMIO mode, when a RX data FIFO trigger ‘tr_rx_req’ is activated.

TR_RX_REQ_W
Field TR_RX_REQ writer - Activated in MMIO mode, when a RX data FIFO trigger ‘tr_rx_req’ is activated.

TR_TX_REQ_R
Field TR_TX_REQ reader - Activated in MMIO mode, when a TX data FIFO trigger ‘tr_tx_req’ is activated.

TR_TX_REQ_W
Field TR_TX_REQ writer - Activated in MMIO mode, when a TX data FIFO trigger ‘tr_tx_req’ is activated.

TX_CMD_FIFO_OVERFLOW_R
Field TX_CMD_FIFO_OVERFLOW reader - Activated in MMIO mode, on an AHB-Lite write transfer to the TX command FIFO (TX_CMD_FIFO_WR) with not enough free entries available.

TX_CMD_FIFO_OVERFLOW_W
Field TX_CMD_FIFO_OVERFLOW writer - Activated in MMIO mode, on an AHB-Lite write transfer to the TX command FIFO (TX_CMD_FIFO_WR) with not enough free entries available.

TX_DATA_FIFO_OVERFLOW_R
Field TX_DATA_FIFO_OVERFLOW reader - Activated in MMIO mode, on an AHB-Lite write transfer to the TX data FIFO (TX_DATA_FIFO_WR1, TX_DATA_FIFO_WR2, TX_DATA_FIFO_WR4) with not enough free entries available.

TX_DATA_FIFO_OVERFLOW_W
Field TX_DATA_FIFO_OVERFLOW writer - Activated in MMIO mode, on an AHB-Lite write transfer to the TX data FIFO (TX_DATA_FIFO_WR1, TX_DATA_FIFO_WR2, TX_DATA_FIFO_WR4) with not enough free entries available.

W
Register INTR writer

XIP_ALIGNMENT_ERROR_R
Field XIP_ALIGNMENT_ERROR reader - Activated in XIP mode, if: - The selected device’s ADDR_CTL.DIV2 is ‘1’ and the AHB-Lite bus transfer address is not a multiple of 2. - The selected device’s ADDR_CTL.DIV2 is ‘1’ and the XIP transfer request is NOT for a multiple of 2 Bytes. Note: In dual-quad SPI mode (ADDR_CTL.DIV is ‘1’), each memory device contributes a 4-bit nibble for read data or write data. This is only possible if the request address is a multiple of 2 and the number of requested Bytes is a multiple of 2.

XIP_ALIGNMENT_ERROR_W
Field XIP_ALIGNMENT_ERROR writer - Activated in XIP mode, if: - The selected device’s ADDR_CTL.DIV2 is ‘1’ and the AHB-Lite bus transfer address is not a multiple of 2. - The selected device’s ADDR_CTL.DIV2 is ‘1’ and the XIP transfer request is NOT for a multiple of 2 Bytes. Note: In dual-quad SPI mode (ADDR_CTL.DIV is ‘1’), each memory device contributes a 4-bit nibble for read data or write data. This is only possible if the request address is a multiple of 2 and the number of requested Bytes is a multiple of 2.