Struct psoc6_01_pac::srss::clk_dsi_select::DSI_MUX_W

pub struct DSI_MUX_W<'a> { /* fields omitted */ }

Field DSI_MUX writer - Selects a DSI source or low frequency clock for use in a clock path. The output of this mux can be selected for clock PATH using CLK_SELECT_PATH register. Using the output of this mux as HFCLK source will result in undefined behavior. It can be used to clocks to DSI or to the reference inputs of FLL/PLL, subject to the frequency limits of those circuits. This mux is not glitch free, so do not change the selection while it is an actively selected clock.

Implementations

impl<'a> DSI_MUX_W<'a>

pub fn variant(self, variant: DSI_MUX_A) -> &'a mut W

Writes variant to the field

pub fn dsi_out0(self) -> &'a mut W

DSI0 - dsi_out[0]

pub fn dsi_out1(self) -> &'a mut W

DSI1 - dsi_out[1]

pub fn dsi_out2(self) -> &'a mut W

DSI2 - dsi_out[2]

pub fn dsi_out3(self) -> &'a mut W

DSI3 - dsi_out[3]

pub fn dsi_out4(self) -> &'a mut W

DSI4 - dsi_out[4]

pub fn dsi_out5(self) -> &'a mut W

DSI5 - dsi_out[5]

pub fn dsi_out6(self) -> &'a mut W

DSI6 - dsi_out[6]

pub fn dsi_out7(self) -> &'a mut W

DSI7 - dsi_out[7]

pub fn dsi_out8(self) -> &'a mut W

DSI8 - dsi_out[8]

pub fn dsi_out9(self) -> &'a mut W

DSI9 - dsi_out[9]

pub fn dsi_out10(self) -> &'a mut W

DSI10 - dsi_out[10]

pub fn dsi_out11(self) -> &'a mut W

DSI11 - dsi_out[11]

pub fn dsi_out12(self) -> &'a mut W

DSI12 - dsi_out[12]

pub fn dsi_out13(self) -> &'a mut W

DSI13 - dsi_out[13]

pub fn dsi_out14(self) -> &'a mut W

DSI14 - dsi_out[14]

pub fn dsi_out15(self) -> &'a mut W

DSI15 - dsi_out[15]

pub fn ilo(self) -> &'a mut W

ILO - Internal Low-speed Oscillator

pub fn wco(self) -> &'a mut W

WCO - Watch-Crystal Oscillator

pub fn altlf(self) -> &'a mut W

ALTLF - Alternate Low-Frequency Clock

pub fn pilo(self) -> &'a mut W

PILO - Precision Internal Low-speed Oscillator

pub unsafe fn bits(self, value: u8) -> &'a mut W

Writes raw bits to the field