Module psoc6_01_pac::srss::clk_eco_config

ECO Configuration Register

Structs

AGC_EN_R
Field AGC_EN reader - Automatic Gain Control (AGC) enable. When set, the oscillation amplitude is controlled to the level selected by ECO_TRIM0.ATRIM. When low, the amplitude is not explicitly controlled and can be as high as the vddd supply. WARNING: use care when disabling AGC because driving a crystal beyond its rated limit can permanently damage the crystal.

AGC_EN_W
Field AGC_EN writer - Automatic Gain Control (AGC) enable. When set, the oscillation amplitude is controlled to the level selected by ECO_TRIM0.ATRIM. When low, the amplitude is not explicitly controlled and can be as high as the vddd supply. WARNING: use care when disabling AGC because driving a crystal beyond its rated limit can permanently damage the crystal.

CLK_ECO_CONFIG_SPEC
ECO Configuration Register

ECO_EN_R
Field ECO_EN reader - Master enable for ECO oscillator.

ECO_EN_W
Field ECO_EN writer - Master enable for ECO oscillator.

R
Register CLK_ECO_CONFIG reader

W
Register CLK_ECO_CONFIG writer