Module psoc6_01_pac::srss::clk_fll_config2

FLL Configuration Register 2

Structs

CLK_FLL_CONFIG2_SPEC
FLL Configuration Register 2

FLL_REF_DIV_R
Field FLL_REF_DIV reader - Control bits for reference divider. Set the divide value before enabling the FLL, and do not change it while FLL is enabled. 0: illegal (undefined behavior) 1: divide by 1 … 8191: divide by 8191

FLL_REF_DIV_W
Field FLL_REF_DIV writer - Control bits for reference divider. Set the divide value before enabling the FLL, and do not change it while FLL is enabled. 0: illegal (undefined behavior) 1: divide by 1 … 8191: divide by 8191

LOCK_TOL_R
Field LOCK_TOL reader - Lock tolerance sets the error threshold for when the FLL output is considered locked to the reference input. A high tolerance can be used to lock more quickly or to track a less accurate source. The tolerance should be set so that the FLL does not unlock under normal conditions. The tolerance is the allowed difference between the count value for the ideal formula and the measured value. 0: tolerate error of 1 count value 1: tolerate error of 2 count values … 511: tolerate error of 512 count values

LOCK_TOL_W
Field LOCK_TOL writer - Lock tolerance sets the error threshold for when the FLL output is considered locked to the reference input. A high tolerance can be used to lock more quickly or to track a less accurate source. The tolerance should be set so that the FLL does not unlock under normal conditions. The tolerance is the allowed difference between the count value for the ideal formula and the measured value. 0: tolerate error of 1 count value 1: tolerate error of 2 count values … 511: tolerate error of 512 count values

R
Register CLK_FLL_CONFIG2 reader

W
Register CLK_FLL_CONFIG2 writer