Module psoc6_01_pac::srss::clk_fll_config4

FLL Configuration Register 4

Structs

CCO_ENABLE_R
Field CCO_ENABLE reader - Enable the CCO. It is required to enable the CCO before using the FLL. 0: Block is powered off 1: Block is powered on

CCO_ENABLE_W
Field CCO_ENABLE writer - Enable the CCO. It is required to enable the CCO before using the FLL. 0: Block is powered off 1: Block is powered on

CCO_FREQ_R
Field CCO_FREQ reader - CCO frequency code. This is updated by HW when the FLL is enabled. It can be manually updated to use the CCO in an open loop configuration. The meaning of each frequency code depends on the range.

CCO_FREQ_W
Field CCO_FREQ writer - CCO frequency code. This is updated by HW when the FLL is enabled. It can be manually updated to use the CCO in an open loop configuration. The meaning of each frequency code depends on the range.

CCO_HW_UPDATE_DIS_R
Field CCO_HW_UPDATE_DIS reader - Disable CCO frequency update by FLL hardware 0: Hardware update of CCO settings is allowed. Use this setting for normal FLL operation. 1: Hardware update of CCO settings is disabled. Use this setting for open-loop FLL operation.

CCO_HW_UPDATE_DIS_W
Field CCO_HW_UPDATE_DIS writer - Disable CCO frequency update by FLL hardware 0: Hardware update of CCO settings is allowed. Use this setting for normal FLL operation. 1: Hardware update of CCO settings is disabled. Use this setting for open-loop FLL operation.

CCO_LIMIT_R
Field CCO_LIMIT reader - Maximum CCO offset allowed (used to prevent FLL dynamics from selecting an CCO frequency that the logic cannot support)

CCO_LIMIT_W
Field CCO_LIMIT writer - Maximum CCO offset allowed (used to prevent FLL dynamics from selecting an CCO frequency that the logic cannot support)

CCO_RANGE_R
Field CCO_RANGE reader - Frequency range of CCO

CCO_RANGE_W
Field CCO_RANGE writer - Frequency range of CCO

CLK_FLL_CONFIG4_SPEC
FLL Configuration Register 4

R
Register CLK_FLL_CONFIG4 reader

W
Register CLK_FLL_CONFIG4 writer

Enums

CCO_RANGE_A
Frequency range of CCO