Module psoc6_01_pac::srss::clk_output_fast

Fast Clock Output Select Register

Structs

CLK_OUTPUT_FAST_SPEC
Fast Clock Output Select Register

FAST_SEL0_R
Field FAST_SEL0 reader - Select signal for fast clock output #0

FAST_SEL0_W
Field FAST_SEL0 writer - Select signal for fast clock output #0

FAST_SEL1_R
Field FAST_SEL1 reader - Select signal for fast clock output #1

FAST_SEL1_W
Field FAST_SEL1 writer - Select signal for fast clock output #1

HFCLK_SEL0_R
Field HFCLK_SEL0 reader - Selects a HFCLK tree for use in fast clock output #0

HFCLK_SEL0_W
Field HFCLK_SEL0 writer - Selects a HFCLK tree for use in fast clock output #0

HFCLK_SEL1_R
Field HFCLK_SEL1 reader - Selects a HFCLK tree for use in fast clock output #1 logic

HFCLK_SEL1_W
Field HFCLK_SEL1 writer - Selects a HFCLK tree for use in fast clock output #1 logic

PATH_SEL0_R
Field PATH_SEL0 reader - Selects a clock path to use in fast clock output #0 logic. 0: FLL output 1-15: PLL output on path1-path15 (if available)

PATH_SEL0_W
Field PATH_SEL0 writer - Selects a clock path to use in fast clock output #0 logic. 0: FLL output 1-15: PLL output on path1-path15 (if available)

PATH_SEL1_R
Field PATH_SEL1 reader - Selects a clock path to use in fast clock output #1 logic. 0: FLL output 1-15: PLL output on path1-path15 (if available)

PATH_SEL1_W
Field PATH_SEL1 writer - Selects a clock path to use in fast clock output #1 logic. 0: FLL output 1-15: PLL output on path1-path15 (if available)

R
Register CLK_OUTPUT_FAST reader

W
Register CLK_OUTPUT_FAST writer

Enums

FAST_SEL0_A
Select signal for fast clock output #0

FAST_SEL1_A
Select signal for fast clock output #1