Module psoc6_01_pac::srss::clk_select

Clock selection register

Structs

CLK_SELECT_SPEC
Clock selection register

LFCLK_SEL_R
Field LFCLK_SEL reader - Select source for LFCLK. Note that not all products support all clock sources. Selecting a clock source that is not supported will result in undefined behavior. Writes to this field are ignored unless the WDT is unlocked using WDT_LOCK register.

LFCLK_SEL_W
Field LFCLK_SEL writer - Select source for LFCLK. Note that not all products support all clock sources. Selecting a clock source that is not supported will result in undefined behavior. Writes to this field are ignored unless the WDT is unlocked using WDT_LOCK register.

PUMP_DIV_R
Field PUMP_DIV reader - Division ratio for PUMPCLK. Uses selected PUMP_SEL clock as the source.

PUMP_DIV_W
Field PUMP_DIV writer - Division ratio for PUMPCLK. Uses selected PUMP_SEL clock as the source.

PUMP_ENABLE_R
Field PUMP_ENABLE reader - Enable the pump clock. PUMP_ENABLE and the PUMP_SEL mux are not glitch-free to minimize side-effects, avoid changing the PUMP_SEL and PUMP_DIV while changing PUMP_ENABLE. To change the settings, do the following: 1) If the pump clock is enabled, write PUMP_ENABLE=0 without changing PUMP_SEL and PUMP_DIV. 2) Change PUMP_SEL and PUMP_DIV to desired settings with PUMP_ENABLE=0. 3) Write PUMP_ENABLE=1 without changing PUMP_SEL and PUMP_DIV.

PUMP_ENABLE_W
Field PUMP_ENABLE writer - Enable the pump clock. PUMP_ENABLE and the PUMP_SEL mux are not glitch-free to minimize side-effects, avoid changing the PUMP_SEL and PUMP_DIV while changing PUMP_ENABLE. To change the settings, do the following: 1) If the pump clock is enabled, write PUMP_ENABLE=0 without changing PUMP_SEL and PUMP_DIV. 2) Change PUMP_SEL and PUMP_DIV to desired settings with PUMP_ENABLE=0. 3) Write PUMP_ENABLE=1 without changing PUMP_SEL and PUMP_DIV.

PUMP_SEL_R
Field PUMP_SEL reader - Selects clock PATH, where k=PUMP_SEL. The output of this mux goes to the PUMP_DIV to make PUMPCLK Each product has a specific number of available clock paths. Selecting a path that is not implemented on a product will result in undefined behavior. Note that this is not a glitch free mux.

PUMP_SEL_W
Field PUMP_SEL writer - Selects clock PATH, where k=PUMP_SEL. The output of this mux goes to the PUMP_DIV to make PUMPCLK Each product has a specific number of available clock paths. Selecting a path that is not implemented on a product will result in undefined behavior. Note that this is not a glitch free mux.

R
Register CLK_SELECT reader

W
Register CLK_SELECT writer

Enums

LFCLK_SEL_A
Select source for LFCLK. Note that not all products support all clock sources. Selecting a clock source that is not supported will result in undefined behavior. Writes to this field are ignored unless the WDT is unlocked using WDT_LOCK register.

PUMP_DIV_A
Division ratio for PUMPCLK. Uses selected PUMP_SEL clock as the source.