Module psoc6_01_pac::srss::clk_trim_eco_ctl

ECO Trim Register

Structs

ATRIM_R
Field ATRIM reader - Amplitude trim to set the crystal drive level when ECO_CONFIG.AGC_EN=1. WARNING: use care when setting this field because driving a crystal beyond its rated limit can permanently damage the crystal. 0x0 - 150mV 0x1 - 175mV 0x2 - 200mV 0x3 - 225mV 0x4 - 250mV 0x5 - 275mV 0x6 - 300mV 0x7 - 325mV 0x8 - 350mV 0x9 - 375mV 0xA - 400mV 0xB - 425mV 0xC - 450mV 0xD - 475mV 0xE - 500mV 0xF - 525mV

ATRIM_W
Field ATRIM writer - Amplitude trim to set the crystal drive level when ECO_CONFIG.AGC_EN=1. WARNING: use care when setting this field because driving a crystal beyond its rated limit can permanently damage the crystal. 0x0 - 150mV 0x1 - 175mV 0x2 - 200mV 0x3 - 225mV 0x4 - 250mV 0x5 - 275mV 0x6 - 300mV 0x7 - 325mV 0x8 - 350mV 0x9 - 375mV 0xA - 400mV 0xB - 425mV 0xC - 450mV 0xD - 475mV 0xE - 500mV 0xF - 525mV

CLK_TRIM_ECO_CTL_SPEC
ECO Trim Register

FTRIM_R
Field FTRIM reader - Filter Trim - 3rd harmonic oscillation

FTRIM_W
Field FTRIM writer - Filter Trim - 3rd harmonic oscillation

GTRIM_R
Field GTRIM reader - Gain Trim - Startup time

GTRIM_W
Field GTRIM writer - Gain Trim - Startup time

ITRIM_R
Field ITRIM reader - Current Trim

ITRIM_W
Field ITRIM writer - Current Trim

R
Register CLK_TRIM_ECO_CTL reader

RTRIM_R
Field RTRIM reader - Feedback resistor Trim

RTRIM_W
Field RTRIM writer - Feedback resistor Trim

W
Register CLK_TRIM_ECO_CTL writer

WDTRIM_R
Field WDTRIM reader - Watch Dog Trim - Delta voltage below steady state level 0x0 - 50mV 0x1 - 75mV 0x2 - 100mV 0x3 - 125mV 0x4 - 150mV 0x5 - 175mV 0x6 - 200mV 0x7 - 225mV

WDTRIM_W
Field WDTRIM writer - Watch Dog Trim - Delta voltage below steady state level 0x0 - 50mV 0x1 - 75mV 0x2 - 100mV 0x3 - 125mV 0x4 - 150mV 0x5 - 175mV 0x6 - 200mV 0x7 - 225mV