Module psoc6_01_pac::srss::mcwdt_struct::mcwdt_ctl

Multi-Counter Watchdog Counter Control

Structs

MCWDT_CTL_SPEC
Multi-Counter Watchdog Counter Control

R
Register MCWDT_CTL reader

W
Register MCWDT_CTL writer

WDT_ENABLE0_R
Field WDT_ENABLE0 reader - Enable subcounter 0. May take up to 2 LFCLK cycles to take effect. 0: Counter is disabled (not clocked) 1: Counter is enabled (counting up)

WDT_ENABLE0_W
Field WDT_ENABLE0 writer - Enable subcounter 0. May take up to 2 LFCLK cycles to take effect. 0: Counter is disabled (not clocked) 1: Counter is enabled (counting up)

WDT_ENABLE1_R
Field WDT_ENABLE1 reader - Enable subcounter 1. May take up to 2 LFCLK cycles to take effect. 0: Counter is disabled (not clocked) 1: Counter is enabled (counting up)

WDT_ENABLE1_W
Field WDT_ENABLE1 writer - Enable subcounter 1. May take up to 2 LFCLK cycles to take effect. 0: Counter is disabled (not clocked) 1: Counter is enabled (counting up)

WDT_ENABLE2_R
Field WDT_ENABLE2 reader - Enable subcounter 2. May take up to 2 LFCLK cycles to take effect. 0: Counter is disabled (not clocked) 1: Counter is enabled (counting up)

WDT_ENABLE2_W
Field WDT_ENABLE2 writer - Enable subcounter 2. May take up to 2 LFCLK cycles to take effect. 0: Counter is disabled (not clocked) 1: Counter is enabled (counting up)

WDT_ENABLED0_R
Field WDT_ENABLED0 reader - Indicates actual state of counter. May lag WDT_ENABLE0 by up to two LFCLK cycles.

WDT_ENABLED1_R
Field WDT_ENABLED1 reader - Indicates actual state of counter. May lag WDT_ENABLE1 by up to two LFCLK cycles.

WDT_ENABLED2_R
Field WDT_ENABLED2 reader - Indicates actual state of counter. May lag WDT_ENABLE2 by up to two LFCLK cycles.

WDT_RESET0_R
Field WDT_RESET0 reader - Resets counter 0 back to 0000. Hardware will reset this bit after counter was reset. This will take up to one LFCLK cycle to take effect.

WDT_RESET0_W
Field WDT_RESET0 writer - Resets counter 0 back to 0000. Hardware will reset this bit after counter was reset. This will take up to one LFCLK cycle to take effect.

WDT_RESET1_R
Field WDT_RESET1 reader - Resets counter 1 back to 0000. Hardware will reset this bit after counter was reset. This will take up to one LFCLK cycle to take effect.

WDT_RESET1_W
Field WDT_RESET1 writer - Resets counter 1 back to 0000. Hardware will reset this bit after counter was reset. This will take up to one LFCLK cycle to take effect.

WDT_RESET2_R
Field WDT_RESET2 reader - Resets counter 2 back to 0000. Hardware will reset this bit after counter was reset. This will take up to one LFCLK cycle to take effect.

WDT_RESET2_W
Field WDT_RESET2 writer - Resets counter 2 back to 0000. Hardware will reset this bit after counter was reset. This will take up to one LFCLK cycle to take effect.