Module psoc6_01_pac::srss::pwr_buck_ctl2

Buck Control Register 2

Structs

BUCK_OUT2_EN_R
Field BUCK_OUT2_EN reader - Enable for vccbuck2 output. The value in this register is ignored unless PWR_BUCK_CTL.BUCK_EN==1. The regulator takes up to 600us to charge the external capacitor. If there is additional load current while charging, this will increase the startup time.

BUCK_OUT2_EN_W
Field BUCK_OUT2_EN writer - Enable for vccbuck2 output. The value in this register is ignored unless PWR_BUCK_CTL.BUCK_EN==1. The regulator takes up to 600us to charge the external capacitor. If there is additional load current while charging, this will increase the startup time.

BUCK_OUT2_HW_SEL_R
Field BUCK_OUT2_HW_SEL reader - Hardware control for vccbuck2 output. When this bit is set, the value in BUCK_OUT2_EN is ignored and a hardware signal is used instead. If the product has supporting hardware, it can directly control the enable signal for vccbuck2. The same charging time in BUCK_OUT2_EN applies.

BUCK_OUT2_HW_SEL_W
Field BUCK_OUT2_HW_SEL writer - Hardware control for vccbuck2 output. When this bit is set, the value in BUCK_OUT2_EN is ignored and a hardware signal is used instead. If the product has supporting hardware, it can directly control the enable signal for vccbuck2. The same charging time in BUCK_OUT2_EN applies.

BUCK_OUT2_SEL_R
Field BUCK_OUT2_SEL reader - Voltage output selection for vccbuck2 output. When increasing the voltage, it can take up to 200us for the output voltage to settle. When decreasing the voltage, the settling time depends on the load current. 0: 1.15V 1: 1.20V 2: 1.25V 3: 1.30V 4: 1.35V 5: 1.40V 6: 1.45V 7: 1.50V

BUCK_OUT2_SEL_W
Field BUCK_OUT2_SEL writer - Voltage output selection for vccbuck2 output. When increasing the voltage, it can take up to 200us for the output voltage to settle. When decreasing the voltage, the settling time depends on the load current. 0: 1.15V 1: 1.20V 2: 1.25V 3: 1.30V 4: 1.35V 5: 1.40V 6: 1.45V 7: 1.50V

PWR_BUCK_CTL2_SPEC
Buck Control Register 2

R
Register PWR_BUCK_CTL2 reader

W
Register PWR_BUCK_CTL2 writer