Module psoc6_01_pac::srss::pwr_hibernate

HIBERNATE Mode Register

Structs

FREEZE_R
Field FREEZE reader - Firmware sets this bit to freeze the configuration, mode and state of all GPIOs and SIOs in the system. When entering HIBERNATE mode, the first write instructs DEEPSLEEP peripherals that they cannot ignore the upcoming freeze command. This occurs even in the illegal condition where UNLOCK is not set. If UNLOCK and HIBERNATE are properly set, the IOs actually freeze on the second write.

FREEZE_W
Field FREEZE writer - Firmware sets this bit to freeze the configuration, mode and state of all GPIOs and SIOs in the system. When entering HIBERNATE mode, the first write instructs DEEPSLEEP peripherals that they cannot ignore the upcoming freeze command. This occurs even in the illegal condition where UNLOCK is not set. If UNLOCK and HIBERNATE are properly set, the IOs actually freeze on the second write.

HIBERNATE_DISABLE_R
Field HIBERNATE_DISABLE reader - Hibernate disable bit. 0: Normal operation, HIBERNATE works as described 1: Further writes to this register are ignored Note: This bit is a write-once bit until the next reset. Avoid changing any other bits in this register while disabling HIBERNATE mode. Also, it is recommended to clear the UNLOCK code, if it was previously written..

HIBERNATE_DISABLE_W
Field HIBERNATE_DISABLE writer - Hibernate disable bit. 0: Normal operation, HIBERNATE works as described 1: Further writes to this register are ignored Note: This bit is a write-once bit until the next reset. Avoid changing any other bits in this register while disabling HIBERNATE mode. Also, it is recommended to clear the UNLOCK code, if it was previously written..

HIBERNATE_R
Field HIBERNATE reader - Firmware sets this bit to enter HIBERNATE mode. The system will enter HIBERNATE mode immediately after writing to this bit and will wakeup only in response to XRES or WAKEUP event. Both UNLOCK and FREEZE must have been set correctly in a previous write operations. Otherwise, it will not enter HIBERNATE. External supplies must have been stable for 250us before entering HIBERNATE mode.

HIBERNATE_W
Field HIBERNATE writer - Firmware sets this bit to enter HIBERNATE mode. The system will enter HIBERNATE mode immediately after writing to this bit and will wakeup only in response to XRES or WAKEUP event. Both UNLOCK and FREEZE must have been set correctly in a previous write operations. Otherwise, it will not enter HIBERNATE. External supplies must have been stable for 250us before entering HIBERNATE mode.

MASK_HIBALARM_R
Field MASK_HIBALARM reader - When set, HIBERNATE will wakeup for a RTC interrupt

MASK_HIBALARM_W
Field MASK_HIBALARM writer - When set, HIBERNATE will wakeup for a RTC interrupt

MASK_HIBPIN_R
Field MASK_HIBPIN reader - When set, HIBERNATE will wakeup if the corresponding pin input matches the POLARITY_HIBPIN setting. Each bit corresponds to one of the wakeup pins.

MASK_HIBPIN_W
Field MASK_HIBPIN writer - When set, HIBERNATE will wakeup if the corresponding pin input matches the POLARITY_HIBPIN setting. Each bit corresponds to one of the wakeup pins.

MASK_HIBWDT_R
Field MASK_HIBWDT reader - When set, HIBERNATE will wakeup if WDT matches

MASK_HIBWDT_W
Field MASK_HIBWDT writer - When set, HIBERNATE will wakeup if WDT matches

POLARITY_HIBPIN_R
Field POLARITY_HIBPIN reader - Each bit sets the active polarity of the corresponding wakeup pin. 0: Pin input of 0 will wakeup the part from HIBERNATE 1: Pin input of 1 will wakeup the part from HIBERNATE

POLARITY_HIBPIN_W
Field POLARITY_HIBPIN writer - Each bit sets the active polarity of the corresponding wakeup pin. 0: Pin input of 0 will wakeup the part from HIBERNATE 1: Pin input of 1 will wakeup the part from HIBERNATE

PWR_HIBERNATE_SPEC
HIBERNATE Mode Register

R
Register PWR_HIBERNATE reader

TOKEN_R
Field TOKEN reader - Contains a 8-bit token that is retained through a HIBERNATE/WAKEUP sequence that can be used by firmware to differentiate WAKEUP from a general RESET event. Note that waking up from HIBERNATE using XRES will reset this register.

TOKEN_W
Field TOKEN writer - Contains a 8-bit token that is retained through a HIBERNATE/WAKEUP sequence that can be used by firmware to differentiate WAKEUP from a general RESET event. Note that waking up from HIBERNATE using XRES will reset this register.

UNLOCK_R
Field UNLOCK reader - This byte must be set to 0x3A for FREEZE or HIBERNATE fields to operate. Any other value in this register will cause FREEZE/HIBERNATE to have no effect, except as noted in the FREEZE description.

UNLOCK_W
Field UNLOCK writer - This byte must be set to 0x3A for FREEZE or HIBERNATE fields to operate. Any other value in this register will cause FREEZE/HIBERNATE to have no effect, except as noted in the FREEZE description.

W
Register PWR_HIBERNATE writer