Module psoc6_01_pac::srss::wdt_cnt

Watchdog Counter Count Register

Structs

COUNTER_R
Field COUNTER reader - Current value of WDT Counter. The write feature of this register is for engineering use (DfV), have no synchronization, and can only be applied when the WDT is fully off. When writing, the value is updated immediately in the WDT counter, but it will read back as the old value until this register resynchronizes just after the negedge of ILO. Writes will be ignored if they occur when the WDT is enabled.

COUNTER_W
Field COUNTER writer - Current value of WDT Counter. The write feature of this register is for engineering use (DfV), have no synchronization, and can only be applied when the WDT is fully off. When writing, the value is updated immediately in the WDT counter, but it will read back as the old value until this register resynchronizes just after the negedge of ILO. Writes will be ignored if they occur when the WDT is enabled.

R
Register WDT_CNT reader

W
Register WDT_CNT writer

WDT_CNT_SPEC
Watchdog Counter Count Register