Struct psoc6_01_pac::tcpwm0::cnt::ctrl::R

pub struct R(_);

Register CTRL reader

Implementations

impl R

pub fn auto_reload_cc(&self) -> AUTO_RELOAD_CC_R

Bit 0 - Specifies switching of the CC and buffered CC values. This field has a function in TIMER, PWM, PWM_DT and PWM_PR modes. Timer mode: ‘0’: never switch. ‘1’: switch on a compare match event. PWM, PWM_DT, PWM_PR modes: ’0: never switch. ‘1’: switch on a terminal count event with an actively pending switch event.

pub fn auto_reload_period(&self) -> AUTO_RELOAD_PERIOD_R

Bit 1 - Specifies switching of the PERIOD and buffered PERIOD values. This field has a function in PWM, PWM_DT and PWM_PR modes. ‘0’: never switch. ‘1’: switch on a terminal count event with and actively pending switch event.

pub fn pwm_sync_kill(&self) -> PWM_SYNC_KILL_R

Bit 2 - Specifies asynchronous/synchronous kill behavior: ‘1’: synchronous kill mode: the kill event disables the ‘dt_line_out’ and ‘dt_line_compl_out’ signals till the next terminal count event (synchronous kill). In synchronous kill mode, STOP_EDGE should be RISING_EDGE. ‘0’: asynchronous kill mode: the kill event only disables the ‘dt_line_out’ and ‘dt_line_compl_out’ signals when present. In asynchronous kill mode, STOP_EDGE should be NO_EDGE_DET. This field has a function in PWM and PWM_DT modes only. This field is only used when PWM_STOP_ON_KILL is ‘0’.

pub fn pwm_stop_on_kill(&self) -> PWM_STOP_ON_KILL_R

Bit 3 - Specifies whether the counter stops on a kill events: ‘0’: kill event does NOT stop counter. ‘1’: kill event stops counter. This field has a function in PWM, PWM_DT and PWM_PR modes only.

pub fn generic(&self) -> GENERIC_R

Bits 8:15 - Generic 8-bit control field. In PWM_DT mode, this field is used to determine the dead time: amount of dead time cycles in the counter clock domain. In all other modes, the lower 3 bits of this field determine pre-scaling of the selected counter clock.

pub fn up_down_mode(&self) -> UP_DOWN_MODE_R

Bits 16:17 - Determines counter direction.

pub fn one_shot(&self) -> ONE_SHOT_R

Bit 18 - When ‘0’, counter runs continuous. When ‘1’, counter is turned off by hardware when a terminal count event is generated.

pub fn quadrature_mode(&self) -> QUADRATURE_MODE_R

Bits 20:21 - In QUAD mode selects quadrature encoding mode (X1/X2/X4). In PWM, PWM_DT and PWM_PR modes, these two bits can be used to invert ‘dt_line_out’ and ‘dt_line_compl_out’. Inversion is the last step in generation of ‘dt_line_out’ and ‘dt_line_compl_out’; i.e. a disabled output line ‘dt_line_out’ has the value QUADRATURE_MODE[0] and a disabled output line ‘dt_line_compl_out’ has the value QUADRATURE_MODE[1].

pub fn mode(&self) -> MODE_R

Bits 24:26 - Counter mode.