Module psoc6_01_pac::tcpwm0::cnt::tr_ctrl0

Counter trigger control register 0

Structs

CAPTURE_SEL_R
Field CAPTURE_SEL reader - Selects one of the 16 input triggers as a capture trigger. Input trigger 0 is always ‘0’ and input trigger is always ‘1’. In the PWM, PWM_DT and PWM_PR modes this trigger is used to switch the values if the compare and period registers with their buffer counterparts.

CAPTURE_SEL_W
Field CAPTURE_SEL writer - Selects one of the 16 input triggers as a capture trigger. Input trigger 0 is always ‘0’ and input trigger is always ‘1’. In the PWM, PWM_DT and PWM_PR modes this trigger is used to switch the values if the compare and period registers with their buffer counterparts.

COUNT_SEL_R
Field COUNT_SEL reader - Selects one of the 16 input triggers as a count trigger. In QUAD mode, this is the first phase (phi A). Default setting selects input trigger 1, which is always ‘1’.

COUNT_SEL_W
Field COUNT_SEL writer - Selects one of the 16 input triggers as a count trigger. In QUAD mode, this is the first phase (phi A). Default setting selects input trigger 1, which is always ‘1’.

R
Register TR_CTRL0 reader

RELOAD_SEL_R
Field RELOAD_SEL reader - Selects one of the 16 input triggers as a reload trigger. In QUAD mode, this is the index or revolution pulse. In this mode, it will update the counter with 0x8000 (counter midpoint).

RELOAD_SEL_W
Field RELOAD_SEL writer - Selects one of the 16 input triggers as a reload trigger. In QUAD mode, this is the index or revolution pulse. In this mode, it will update the counter with 0x8000 (counter midpoint).

START_SEL_R
Field START_SEL reader - Selects one of the 16 input triggers as a start trigger. In QUAD mode, this is the second phase (phi B).

START_SEL_W
Field START_SEL writer - Selects one of the 16 input triggers as a start trigger. In QUAD mode, this is the second phase (phi B).

STOP_SEL_R
Field STOP_SEL reader - Selects one of the 16 input triggers as a stop trigger. In PWM, PWM_DT and PWM_PR modes, this is the kill trigger. In these modes, the kill trigger is used to either temporarily block the PWM outputs (PWM_STOP_ON_KILL is ‘0’) or stop the functionality (PWM_STOP_ON_KILL is ‘1’). For the PWM and PWM_DT modes, the blocking of the output signals can be asynchronous (STOP_EDGE should be NO_EDGE_DET) in which case the blocking is as long as the trigger is ‘1’ or synchronous (STOP_EDGE should be RISING_EDGE) in which case it extends till the next terminal count event.

STOP_SEL_W
Field STOP_SEL writer - Selects one of the 16 input triggers as a stop trigger. In PWM, PWM_DT and PWM_PR modes, this is the kill trigger. In these modes, the kill trigger is used to either temporarily block the PWM outputs (PWM_STOP_ON_KILL is ‘0’) or stop the functionality (PWM_STOP_ON_KILL is ‘1’). For the PWM and PWM_DT modes, the blocking of the output signals can be asynchronous (STOP_EDGE should be NO_EDGE_DET) in which case the blocking is as long as the trigger is ‘1’ or synchronous (STOP_EDGE should be RISING_EDGE) in which case it extends till the next terminal count event.

TR_CTRL0_SPEC
Counter trigger control register 0

W
Register TR_CTRL0 writer