Module psoc6_01_pac::tcpwm0::cnt::tr_ctrl1

Counter trigger control register 1

Structs

CAPTURE_EDGE_R
Field CAPTURE_EDGE reader - A capture event will copy the counter value into the CC register.

CAPTURE_EDGE_W
Field CAPTURE_EDGE writer - A capture event will copy the counter value into the CC register.

COUNT_EDGE_R
Field COUNT_EDGE reader - A counter event will increase or decrease the counter by ‘1’.

COUNT_EDGE_W
Field COUNT_EDGE writer - A counter event will increase or decrease the counter by ‘1’.

R
Register TR_CTRL1 reader

RELOAD_EDGE_R
Field RELOAD_EDGE reader - A reload event will initialize the counter. When counting up, the counter is initialized to ‘0’. When counting down, the counter is initialized with PERIOD.

RELOAD_EDGE_W
Field RELOAD_EDGE writer - A reload event will initialize the counter. When counting up, the counter is initialized to ‘0’. When counting down, the counter is initialized with PERIOD.

START_EDGE_R
Field START_EDGE reader - A start event will start the counter; i.e. the counter will become running. Starting does NOT enable the counter. A start event will not initialize the counter whereas the reload event does.

START_EDGE_W
Field START_EDGE writer - A start event will start the counter; i.e. the counter will become running. Starting does NOT enable the counter. A start event will not initialize the counter whereas the reload event does.

STOP_EDGE_R
Field STOP_EDGE reader - A stop event, will stop the counter; i.e. it will no longer be running. Stopping will NOT disable the counter.

STOP_EDGE_W
Field STOP_EDGE writer - A stop event, will stop the counter; i.e. it will no longer be running. Stopping will NOT disable the counter.

TR_CTRL1_SPEC
Counter trigger control register 1

W
Register TR_CTRL1 writer

Enums

CAPTURE_EDGE_A
A capture event will copy the counter value into the CC register.

COUNT_EDGE_A
A counter event will increase or decrease the counter by ‘1’.

RELOAD_EDGE_A
A reload event will initialize the counter. When counting up, the counter is initialized to ‘0’. When counting down, the counter is initialized with PERIOD.

START_EDGE_A
A start event will start the counter; i.e. the counter will become running. Starting does NOT enable the counter. A start event will not initialize the counter whereas the reload event does.

STOP_EDGE_A
A stop event, will stop the counter; i.e. it will no longer be running. Stopping will NOT disable the counter.