Module psoc6_01_pac::tcpwm0::cnt::tr_ctrl2

Counter trigger control register 2

Structs

CC_MATCH_MODE_R
Field CC_MATCH_MODE reader - Determines the effect of a compare match event (COUNTER equals CC register) on the ‘line_out’ output signals. Note that INVERT is especially useful for center aligned pulse width modulation. To generate a duty cycle of 0 percent, the counter CC register should be set to ‘0’. For a 100 percent duty cycle, the counter CC register should be set to larger than the counter PERIOD register.

CC_MATCH_MODE_W
Field CC_MATCH_MODE writer - Determines the effect of a compare match event (COUNTER equals CC register) on the ‘line_out’ output signals. Note that INVERT is especially useful for center aligned pulse width modulation. To generate a duty cycle of 0 percent, the counter CC register should be set to ‘0’. For a 100 percent duty cycle, the counter CC register should be set to larger than the counter PERIOD register.

OVERFLOW_MODE_R
Field OVERFLOW_MODE reader - Determines the effect of a counter overflow event (COUNTER reaches PERIOD) on the ‘line_out’ output signals.

OVERFLOW_MODE_W
Field OVERFLOW_MODE writer - Determines the effect of a counter overflow event (COUNTER reaches PERIOD) on the ‘line_out’ output signals.

R
Register TR_CTRL2 reader

TR_CTRL2_SPEC
Counter trigger control register 2

UNDERFLOW_MODE_R
Field UNDERFLOW_MODE reader - Determines the effect of a counter underflow event (COUNTER reaches ‘0’) on the ‘line_out’ output signals.

UNDERFLOW_MODE_W
Field UNDERFLOW_MODE writer - Determines the effect of a counter underflow event (COUNTER reaches ‘0’) on the ‘line_out’ output signals.

W
Register TR_CTRL2 writer

Enums

CC_MATCH_MODE_A
Determines the effect of a compare match event (COUNTER equals CC register) on the ‘line_out’ output signals. Note that INVERT is especially useful for center aligned pulse width modulation. To generate a duty cycle of 0 percent, the counter CC register should be set to ‘0’. For a 100 percent duty cycle, the counter CC register should be set to larger than the counter PERIOD register.

OVERFLOW_MODE_A
Determines the effect of a counter overflow event (COUNTER reaches PERIOD) on the ‘line_out’ output signals.

UNDERFLOW_MODE_A
Determines the effect of a counter underflow event (COUNTER reaches ‘0’) on the ‘line_out’ output signals.