Module psoc6_01_pac::usbfs0::usbdev::arb_ep2_cfg

Endpoint Configuration Register *1

Structs

ARB_EP2_CFG_SPEC
Endpoint Configuration Register *1

CRC_BYPASS_R
Field CRC_BYPASS reader - Configuration Setting to prevent CRC bytes from being written to memory and being read by firmware

CRC_BYPASS_W
Field CRC_BYPASS writer - Configuration Setting to prevent CRC bytes from being written to memory and being read by firmware

DMA_REQ_R
Field DMA_REQ reader - Manual DMA Request for a particular (1 to 8) endpoint; changing this field from 0 to 1 causes a DMA request to be generated.

DMA_REQ_W
Field DMA_REQ writer - Manual DMA Request for a particular (1 to 8) endpoint; changing this field from 0 to 1 causes a DMA request to be generated.

IN_DATA_RDY_R
Field IN_DATA_RDY reader - Indication that Endpoint Packet Data is Ready in Main memory

IN_DATA_RDY_W
Field IN_DATA_RDY writer - Indication that Endpoint Packet Data is Ready in Main memory

R
Register ARB_EP2_CFG reader

RESET_PTR_R
Field RESET_PTR reader - Configuration Setting to Reset the RA and WA Pointers to their start values at the End of Packet transaction.

RESET_PTR_W
Field RESET_PTR writer - Configuration Setting to Reset the RA and WA Pointers to their start values at the End of Packet transaction.

W
Register ARB_EP2_CFG writer

Enums

CRC_BYPASS_A
Configuration Setting to prevent CRC bytes from being written to memory and being read by firmware

RESET_PTR_A
Configuration Setting to Reset the RA and WA Pointers to their start values at the End of Packet transaction.