Module psoc6_01_pac::usbfs0::usbdev::cr1

USB control 1 Register

Structs

BUS_ACTIVITY_R
Field BUS_ACTIVITY reader - The Bus Activity bit is a stickybit that detects any non-idle USB event that has occurred on the USB bus. Once set to High by the SIE to indicate the bus activity this bit retains its logical High value until firmware clears it.

BUS_ACTIVITY_W
Field BUS_ACTIVITY writer - The Bus Activity bit is a stickybit that detects any non-idle USB event that has occurred on the USB bus. Once set to High by the SIE to indicate the bus activity this bit retains its logical High value until firmware clears it.

CR1_SPEC
USB control 1 Register

ENABLE_LOCK_R
Field ENABLE_LOCK reader - This bit is set to turn on the automatic frequency locking of the internal oscillator to USB traffic. Unless an external clock is being provided this bit should remain set for proper USB operation.

ENABLE_LOCK_W
Field ENABLE_LOCK writer - This bit is set to turn on the automatic frequency locking of the internal oscillator to USB traffic. Unless an external clock is being provided this bit should remain set for proper USB operation.

R
Register CR1 reader

REG_ENABLE_R
Field REG_ENABLE reader - This bit controls the operation of the internal USB regulator. For applications with supply voltages in the 5V range this bit is set high to enable the internal regulator. For device supply voltage in the 3.3V range this bit is cleared to connect the transceiver directly to the supply.

REG_ENABLE_W
Field REG_ENABLE writer - This bit controls the operation of the internal USB regulator. For applications with supply voltages in the 5V range this bit is set high to enable the internal regulator. For device supply voltage in the 3.3V range this bit is cleared to connect the transceiver directly to the supply.

RSVD_3_R
Field RSVD_3 reader - N/A

RSVD_3_W
Field RSVD_3 writer - N/A

W
Register CR1 writer