Module psoc6_01_pac::usbfs0::usbdev::ep0_cnt

Endpoint0 count Register

Structs

BYTE_COUNT_R
Field BYTE_COUNT reader - These bits indicate the number of data bytes in a transaction. For IN transactions firmware loads the count with the number of bytes to be transmitted to the host from the endpoint FIFO. Valid values are 0 to 8. For OUT or SETUP transactions the count is updated by hardware to the number of data bytes received plus two for the CRC bytes. Valid values are 2 to 10.

BYTE_COUNT_W
Field BYTE_COUNT writer - These bits indicate the number of data bytes in a transaction. For IN transactions firmware loads the count with the number of bytes to be transmitted to the host from the endpoint FIFO. Valid values are 0 to 8. For OUT or SETUP transactions the count is updated by hardware to the number of data bytes received plus two for the CRC bytes. Valid values are 2 to 10.

DATA_TOGGLE_R
Field DATA_TOGGLE reader - This bit selects the DATA packet’s toggle state. For IN transactions firmware must set this bit to the expected state. For OUT transactions the hardware sets this bit to the state of the received Data Toggle bit.

DATA_TOGGLE_W
Field DATA_TOGGLE writer - This bit selects the DATA packet’s toggle state. For IN transactions firmware must set this bit to the expected state. For OUT transactions the hardware sets this bit to the state of the received Data Toggle bit.

DATA_VALID_R
Field DATA_VALID reader - This bit is used for OUT/SETUP transactions only and is read only. It is cleared to ‘0’ if CRC bit stuffing errors or PID errors occur. This bit does not update for some endpoint mode settings.

DATA_VALID_W
Field DATA_VALID writer - This bit is used for OUT/SETUP transactions only and is read only. It is cleared to ‘0’ if CRC bit stuffing errors or PID errors occur. This bit does not update for some endpoint mode settings.

EP0_CNT_SPEC
Endpoint0 count Register

R
Register EP0_CNT reader

W
Register EP0_CNT writer

Enums

DATA_VALID_A
This bit is used for OUT/SETUP transactions only and is read only. It is cleared to ‘0’ if CRC bit stuffing errors or PID errors occur. This bit does not update for some endpoint mode settings.