Module psoc6_01_pac::usbfs0::usbdev::usbio_cr1

USBIO control 1 Register

Structs

DMO_R
Field DMO reader - This read only bit gives the state of the D- pin when IOMODE bit is ‘0’ and USB doesn’t transmit. This bit is ‘0’ when USB transmits SE0, and this bit is ‘1’ when USB transmits other than SE0. This bit is valid if USB Device.

DPO_R
Field DPO reader - This read only bit gives the state of the D+ pin when IOMODE bit is ‘0’ and USB doesn’t transmit. This bit displays the output value of D+ pin when USB transmits SE0 or data. This bit is valid if USB Device.

IOMODE_R
Field IOMODE reader - This bit allows the D+ and D- pins to be configured for either USB mode or bit-banged modes. If this bit is set the DMI and DPI bits are used to drive the D- and D+ pins.

IOMODE_W
Field IOMODE writer - This bit allows the D+ and D- pins to be configured for either USB mode or bit-banged modes. If this bit is set the DMI and DPI bits are used to drive the D- and D+ pins.

R
Register USBIO_CR1 reader

RSVD_2_R
Field RSVD_2 reader - N/A

RSVD_2_W
Field RSVD_2 writer - N/A

USBIO_CR1_SPEC
USBIO control 1 Register

W
Register USBIO_CR1 writer