Module psoc6_01_pac::usbfs0::usbdev::usbio_cr2

USBIO control 2 Register

Structs

R
Register USBIO_CR2 reader

RSVD_5_0_R
Field RSVD_5_0 reader - N/A

RSVD_7_R
Field RSVD_7 reader - N/A

RSVD_7_W
Field RSVD_7 writer - N/A

TEST_PKT_R
Field TEST_PKT reader - This bit enables the device to transmit a packet in response to an internally generated IN packet. When set, one packet will be generated.

TEST_PKT_W
Field TEST_PKT writer - This bit enables the device to transmit a packet in response to an internally generated IN packet. When set, one packet will be generated.

USBIO_CR2_SPEC
USBIO control 2 Register

W
Register USBIO_CR2 writer