Module psoc6_01_pac::usbfs0::usbhost::host_ctl2

Host Control 2 Register.

Structs

ALIVE_R
Field ALIVE reader - This bit is used to specify the keep-alive function in the low-speed mode. If this bit it set to ‘1’ while the CLKSEL bit of the Host Control 1 Register (HOST_CTL1) is ‘0’, SE0 is output instead of SOF. This bit is only effective when the CLKSEL bit is ‘0’. If the CLKSEL bit is ‘1’ (Full-Speed mode), SOF is output regardless of the setting of the ALIVE bit. ‘0’ : SOF output. ‘1’ : SE0 output (Keep alive)

ALIVE_W
Field ALIVE writer - This bit is used to specify the keep-alive function in the low-speed mode. If this bit it set to ‘1’ while the CLKSEL bit of the Host Control 1 Register (HOST_CTL1) is ‘0’, SE0 is output instead of SOF. This bit is only effective when the CLKSEL bit is ‘0’. If the CLKSEL bit is ‘1’ (Full-Speed mode), SOF is output regardless of the setting of the ALIVE bit. ‘0’ : SOF output. ‘1’ : SE0 output (Keep alive)

CANCEL_R
Field CANCEL reader - When this bit is set to ‘1’, if the target token is written to the Host Token Endpoint Register (HOST_TOKEN) in the EOF area (specified in the Host EOF Setup Register), its sending is canceled. When this bit is set to ‘0’, token sending is not canceled even if the target token is written to the register. The cancellation of token sending is detected by reading the TCAN bit of the Interrupt USB Host Register (INTR_USBHOST). ‘0’ : Continues a token. ‘1’ : Cancels a token.

CANCEL_W
Field CANCEL writer - When this bit is set to ‘1’, if the target token is written to the Host Token Endpoint Register (HOST_TOKEN) in the EOF area (specified in the Host EOF Setup Register), its sending is canceled. When this bit is set to ‘0’, token sending is not canceled even if the target token is written to the register. The cancellation of token sending is detected by reading the TCAN bit of the Interrupt USB Host Register (INTR_USBHOST). ‘0’ : Continues a token. ‘1’ : Cancels a token.

HOST_CTL2_SPEC
Host Control 2 Register.

R
Register HOST_CTL2 reader

RETRY_R
Field RETRY reader - If this bit is set to ‘1’, the target token is retried if a NAK or error* occurs. Retry processing is performed after the time that is specified in the Host Retry Timer Setup Register (HOST_RTIMER). * : HOST_ERR.RERR=‘1’, HOST_ERR.TOUT=‘1’, HOST_ERR.CRC=‘1’, HOST_ERR.TGERR=‘1’, HOST_ERR.STUFF=‘1’ ‘0’ : Doesn’t retry token sending. ‘1’ : Retries token sending Note: - This bit isn’t initialized even if the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

RETRY_W
Field RETRY writer - If this bit is set to ‘1’, the target token is retried if a NAK or error* occurs. Retry processing is performed after the time that is specified in the Host Retry Timer Setup Register (HOST_RTIMER). * : HOST_ERR.RERR=‘1’, HOST_ERR.TOUT=‘1’, HOST_ERR.CRC=‘1’, HOST_ERR.TGERR=‘1’, HOST_ERR.STUFF=‘1’ ‘0’ : Doesn’t retry token sending. ‘1’ : Retries token sending Note: - This bit isn’t initialized even if the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

RSVD_4_R
Field RSVD_4 reader - N/A

RSVD_4_W
Field RSVD_4 writer - N/A

RSVD_5_R
Field RSVD_5 reader - N/A

RSVD_5_W
Field RSVD_5 writer - N/A

SOFSTEP_R
Field SOFSTEP reader - If this bit is set to ‘1’, the SOF interrupt flag (INTR_USBHOST.SOFIRQ) is set to ‘1’ each time SOF is sent. If this bit is set to ‘0’, the set value of the Host SOF Interrupt Frame Compare Register (HOST_FCOMP) is compared with the low-order eight bits of the SOF frame number. If they match, the SOF interrupt flag (INTR_USBHOST.SOFIRQ) is set to ‘1’. ‘0’ : An interrupt occurred due to the HOST_HFCOMP setting. ‘1’ : An interrupt occurred. Notes: - If a SOF token (TKNEN=‘001’) is sent by the setting of the Host Token Endpoint Register (HOST_TOKEN), the SOF interrupt flag (INTR_USBHOST.SOFIRQ) is not set to ‘1’ regardless of the setting of this bit.

SOFSTEP_W
Field SOFSTEP writer - If this bit is set to ‘1’, the SOF interrupt flag (INTR_USBHOST.SOFIRQ) is set to ‘1’ each time SOF is sent. If this bit is set to ‘0’, the set value of the Host SOF Interrupt Frame Compare Register (HOST_FCOMP) is compared with the low-order eight bits of the SOF frame number. If they match, the SOF interrupt flag (INTR_USBHOST.SOFIRQ) is set to ‘1’. ‘0’ : An interrupt occurred due to the HOST_HFCOMP setting. ‘1’ : An interrupt occurred. Notes: - If a SOF token (TKNEN=‘001’) is sent by the setting of the Host Token Endpoint Register (HOST_TOKEN), the SOF interrupt flag (INTR_USBHOST.SOFIRQ) is not set to ‘1’ regardless of the setting of this bit.

TTEST_R
Field TTEST reader - N/A

TTEST_W
Field TTEST writer - N/A

W
Register HOST_CTL2 writer