Struct psoc6_01_pac::usbfs0::usbhost::host_ctl2::R

pub struct R(_);

Register HOST_CTL2 reader

Implementations

impl R

pub fn retry(&self) -> RETRY_R

Bit 0 - If this bit is set to ‘1’, the target token is retried if a NAK or error* occurs. Retry processing is performed after the time that is specified in the Host Retry Timer Setup Register (HOST_RTIMER). * : HOST_ERR.RERR=‘1’, HOST_ERR.TOUT=‘1’, HOST_ERR.CRC=‘1’, HOST_ERR.TGERR=‘1’, HOST_ERR.STUFF=‘1’ ‘0’ : Doesn’t retry token sending. ‘1’ : Retries token sending Note: - This bit isn’t initialized even if the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

pub fn cancel(&self) -> CANCEL_R

Bit 1 - When this bit is set to ‘1’, if the target token is written to the Host Token Endpoint Register (HOST_TOKEN) in the EOF area (specified in the Host EOF Setup Register), its sending is canceled. When this bit is set to ‘0’, token sending is not canceled even if the target token is written to the register. The cancellation of token sending is detected by reading the TCAN bit of the Interrupt USB Host Register (INTR_USBHOST). ‘0’ : Continues a token. ‘1’ : Cancels a token.

pub fn sofstep(&self) -> SOFSTEP_R

Bit 2 - If this bit is set to ‘1’, the SOF interrupt flag (INTR_USBHOST.SOFIRQ) is set to ‘1’ each time SOF is sent. If this bit is set to ‘0’, the set value of the Host SOF Interrupt Frame Compare Register (HOST_FCOMP) is compared with the low-order eight bits of the SOF frame number. If they match, the SOF interrupt flag (INTR_USBHOST.SOFIRQ) is set to ‘1’. ‘0’ : An interrupt occurred due to the HOST_HFCOMP setting. ‘1’ : An interrupt occurred. Notes: - If a SOF token (TKNEN=‘001’) is sent by the setting of the Host Token Endpoint Register (HOST_TOKEN), the SOF interrupt flag (INTR_USBHOST.SOFIRQ) is not set to ‘1’ regardless of the setting of this bit.

pub fn alive(&self) -> ALIVE_R

Bit 3 - This bit is used to specify the keep-alive function in the low-speed mode. If this bit it set to ‘1’ while the CLKSEL bit of the Host Control 1 Register (HOST_CTL1) is ‘0’, SE0 is output instead of SOF. This bit is only effective when the CLKSEL bit is ‘0’. If the CLKSEL bit is ‘1’ (Full-Speed mode), SOF is output regardless of the setting of the ALIVE bit. ‘0’ : SOF output. ‘1’ : SE0 output (Keep alive)

pub fn rsvd_4(&self) -> RSVD_4_R

Bit 4 - N/A

pub fn rsvd_5(&self) -> RSVD_5_R

Bit 5 - N/A

pub fn ttest(&self) -> TTEST_R

Bits 6:7 - N/A