Module psoc6_01_pac::usbfs0::usbhost::host_ep1_status

Host Endpoint 1 Status Register

Structs

HOST_EP1_STATUS_SPEC
Host Endpoint 1 Status Register

INI_ST_R
Field INI_ST reader - This bit shows that EP1 is initialized. If the init bit of the Host Endpoint 1 Control Register (HOST_EP1_CTL) is set to ‘1’ and EP1 is initialized, this bit is to ‘1’. ‘0’ : Not initiatialized ‘1’ : Initialized Note: - This bit isn’t set to ‘0’ or ‘1’ immediately even if BFINI bit of the Host Endpoint 1 Control Register (HOST_EP1_CTL) is set to ‘0’ or ‘1’. Read this bit to confirm the transition.

R
Register HOST_EP1_STATUS reader

RSVD_18_R
Field RSVD_18 reader - N/A

SIZE1_R
Field SIZE1 reader - These bits indicate the number of data bytes written to the receive buffer when IN packet transfer of EP1 has finished. The indication range is from 0x000 to 0x100. Note : - These bits are set to the data size transferred in the IN direction and written to the buffer. Therefore, a value read during transfer in the OUT direction has no effect.

VAL_DATA_R
Field VAL_DATA reader - This bit shows that there is valid data in the EP1 buffer. ‘0’ : Invalid data in the buffer ‘1’ : Valid data in the buffer