Module psoc6_01_pac::usbfs0::usbhost::host_err

Host Error Status Register.

Structs

CRC_R
Field CRC reader - If this bit is set to ‘1’, it means that a CRC error is detected in the USB Host. When this bit is ‘0’, it means that no error is detected. If a CRC error is detected, bit5 (TOUT) of this register is also set to ‘1’. Write ‘1’ to clear, a write of ‘0’ is ignored. ‘0’ : No CRC error. ‘1’ : CRC error detected. Note: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

CRC_W
Field CRC writer - If this bit is set to ‘1’, it means that a CRC error is detected in the USB Host. When this bit is ‘0’, it means that no error is detected. If a CRC error is detected, bit5 (TOUT) of this register is also set to ‘1’. Write ‘1’ to clear, a write of ‘0’ is ignored. ‘0’ : No CRC error. ‘1’ : CRC error detected. Note: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

HOST_ERR_SPEC
Host Error Status Register.

HS_R
Field HS reader - These flags indicate the status of a handshake packet to be sent or received. These flags are set to ‘NULL’ when no handshake occurs due to an error or when a SOF token has been ended with the TKNEN bit of the Host Token Endpoint Register (HOST_TOKEN). These bits are updated when sending or receiving has been ended. Write ‘11’ to set the status back to ‘NULL’, all other write values are ignored. Note: This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

HS_W
Field HS writer - These flags indicate the status of a handshake packet to be sent or received. These flags are set to ‘NULL’ when no handshake occurs due to an error or when a SOF token has been ended with the TKNEN bit of the Host Token Endpoint Register (HOST_TOKEN). These bits are updated when sending or receiving has been ended. Write ‘11’ to set the status back to ‘NULL’, all other write values are ignored. Note: This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

LSTSOF_R
Field LSTSOF reader - If this bit is set to ‘1’, it means that the SOF token can’t be sent in the USB Host because other token is in process. When this bit is ‘0’, it means that SOF token was sent with no error. Write ‘1’ to clear, a write of ‘0’ is ignored. ‘0’ : SOF sent without error. ‘1’ : SOF error detected. Note: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

LSTSOF_W
Field LSTSOF writer - If this bit is set to ‘1’, it means that the SOF token can’t be sent in the USB Host because other token is in process. When this bit is ‘0’, it means that SOF token was sent with no error. Write ‘1’ to clear, a write of ‘0’ is ignored. ‘0’ : SOF sent without error. ‘1’ : SOF error detected. Note: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

R
Register HOST_ERR reader

RERR_R
Field RERR reader - When this bit is set to ‘1’, it means that the received data exceeds the specified maximum number of packets in the USB Host. If a receive error is detected, bit5 (TOUT) of this register is also set to ‘1’. When this bit is ‘0’, it means that no error is detected. Write ‘1’ to clear, a write of ‘0’ is ignored. ‘0’ : No receive error. ‘1’ : Maximum packet receive error detected. - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

RERR_W
Field RERR writer - When this bit is set to ‘1’, it means that the received data exceeds the specified maximum number of packets in the USB Host. If a receive error is detected, bit5 (TOUT) of this register is also set to ‘1’. When this bit is ‘0’, it means that no error is detected. Write ‘1’ to clear, a write of ‘0’ is ignored. ‘0’ : No receive error. ‘1’ : Maximum packet receive error detected. - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

STUFF_R
Field STUFF reader - If this bit is set to ‘1’, it means that a bit stuffing error has been detected. When this bit is ‘0’, it means that no error is detected. If a stuffing error is detected, bit5 (TOUT) of this register is also set to ‘1’. Write ‘1’ to clear, a write of ‘0’ is ignored. ‘0’ : No stuffing error. ‘1’ : Stuffing error detected. Note: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

STUFF_W
Field STUFF writer - If this bit is set to ‘1’, it means that a bit stuffing error has been detected. When this bit is ‘0’, it means that no error is detected. If a stuffing error is detected, bit5 (TOUT) of this register is also set to ‘1’. Write ‘1’ to clear, a write of ‘0’ is ignored. ‘0’ : No stuffing error. ‘1’ : Stuffing error detected. Note: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

TGERR_R
Field TGERR reader - If this bit is set to ‘1’, it means that the data does not match the TGGL data. When this bit is ‘0’, it means that no error is detected. Write ‘1’ to clear, a write of ‘0’ is ignored. ‘0’ : No toggle error. ‘1’ : Toggle error detected. Note: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

TGERR_W
Field TGERR writer - If this bit is set to ‘1’, it means that the data does not match the TGGL data. When this bit is ‘0’, it means that no error is detected. Write ‘1’ to clear, a write of ‘0’ is ignored. ‘0’ : No toggle error. ‘1’ : Toggle error detected. Note: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

TOUT_R
Field TOUT reader - If this bit is set to ‘1’, it means that no response is returned from the device within the specified time after a token has been sent in the USB Host. When this bit is ‘0’, it means that no timeout is detected. Write ‘1’ to clear, a write of ‘0’ is ignored. ‘0’ : No timeout. ‘1’ : Timeout has detected. Note: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

TOUT_W
Field TOUT writer - If this bit is set to ‘1’, it means that no response is returned from the device within the specified time after a token has been sent in the USB Host. When this bit is ‘0’, it means that no timeout is detected. Write ‘1’ to clear, a write of ‘0’ is ignored. ‘0’ : No timeout. ‘1’ : Timeout has detected. Note: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

W
Register HOST_ERR writer

Enums

HS_A
These flags indicate the status of a handshake packet to be sent or received. These flags are set to ‘NULL’ when no handshake occurs due to an error or when a SOF token has been ended with the TKNEN bit of the Host Token Endpoint Register (HOST_TOKEN). These bits are updated when sending or receiving has been ended. Write ‘11’ to set the status back to ‘NULL’, all other write values are ignored. Note: This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.