Module psoc6_01_pac::usbfs0::usbhost::host_status

Host Status Register.

Structs

CLKSEL_ST_R
Field CLKSEL_ST reader - This bit shows whether it is full-speed or not. If the CLKSEL bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’, this bit is set to ‘1’. ‘0’ : Low speed ‘1’ : Full speed Note: - If this bit is different from the CLKSEL bit, The execution of the token and bus reset must wait these bits match. - This bit takes time to be initialized by the RST bit of the Host Control 1 Resgiter (HOST_CTL1). Read this bit to confirm the operation is complete.

CSTAT_R
Field CSTAT reader - When this bit is ‘1’, it means that the device is connected. When this bit is ‘0’, it means that the device is disconnected. ‘0’ : Device is disconnected. ‘1’ : Device is connected. Notes: - This bit is set to the default value if the RST bit of the Host Control 1 Register (Host_CTL1) is set to ‘1’. - The transition to disconnected on RST isn’t immediate. Read this bit to confirm the transition is complete.

HOST_STATUS_SPEC
Host Status Register.

HOST_ST_R
Field HOST_ST reader - This bit shows whether the device is in USB Host mode. If the HOST bit of the Host Control Register (HOST_CTL0) is set to ‘1’, this bit is set to ‘1’. ‘0’ : USB Device ‘1’ : USB Host Notes: - If this bit is different from the HOST bit, The execution of a token must wait these bits match. - This bit takes time to be initialized by the RST bit of the Host Control 1 Resgiter (HOST_CTL1). Read this bit to confirm the operation is complete.

R
Register HOST_STATUS reader

RSTBUSY_R
Field RSTBUSY reader - This bit shows that USB Host is being reset internally. If the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’, this bit is set to ‘1’. If the RST bit of Host Control 1 Register (HOST_CTL1) is set to ‘0’, this bit is set to ‘0’. ‘0’ : USB Host isn’t being reset. ‘1’ : USB Host is being reset. Notes: - If this bit is ‘1’, the a token must not be executed. - This bit isn’t set to ‘0’ or ‘1’ immediately even if the RST bit of Host Control 1 Register (HOST_CTL1) is set to ‘0’ or ‘1’. Read this bit to confirm the operation is complete.

RSVD_5_R
Field RSVD_5 reader - N/A

SOFBUSY_R
Field SOFBUSY reader - When a SOF token is sent using the Host Token Endpoint Register (HOST_TOKEN), this bit is set to ‘1’, which means that the SOF timer is active. When this bit is ‘0’, it means that the SOF timer is under suspension. To stop the active SOF timer, write ‘0’ to this bit. However, if this bit is written with ‘1’, its value is ignored. ‘0’ : The SOF timer is stopped. ‘1’ : The SOF timer is active. Notes: - This bit is set to the initial value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’. - This bit takes time to be initialized by the RST bit of the Host Control 1 Resgiter (HOST_CTL1). - The SOF timer does not stop immediately after this bit has been set to ‘0’ to stop the SOF timer. To check whether or not the SOF timer is stopped, read this bit.

SOFBUSY_W
Field SOFBUSY writer - When a SOF token is sent using the Host Token Endpoint Register (HOST_TOKEN), this bit is set to ‘1’, which means that the SOF timer is active. When this bit is ‘0’, it means that the SOF timer is under suspension. To stop the active SOF timer, write ‘0’ to this bit. However, if this bit is written with ‘1’, its value is ignored. ‘0’ : The SOF timer is stopped. ‘1’ : The SOF timer is active. Notes: - This bit is set to the initial value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’. - This bit takes time to be initialized by the RST bit of the Host Control 1 Resgiter (HOST_CTL1). - The SOF timer does not stop immediately after this bit has been set to ‘0’ to stop the SOF timer. To check whether or not the SOF timer is stopped, read this bit.

SUSP_R
Field SUSP reader - If this bit is set to ‘1’, the USB Host is placed into the suspend state. If this bit is set to ‘0’ while it is ‘1’ or the USB bus is placed into the k-state mode, then suspend state is released, and the RWIRQ bit of the Interrupt USB Host Register (INTR_USBHOST) is set to ‘1’. Set to ‘1’ : Suspend. Set ‘0’ when this bit is ‘1’ : Resume. Other conditions : Holds the status. Notes: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’. - The transition to disconnected on RST isn’t immediate. Read this bit to confirm the transition is complete. - If this bit is set to ‘1’, this bit must not be set to ‘1’ until the RWIRQ bit of the Interrupt USB Host Register (INTR_USBHOST) is set to ‘1’. - Do not set this bit to ‘1’ while the USB is active (during USB bus resetting, data transfer, or SOF timer running). - If the value of this bit is changed, it is not immediately reflected on the state of the USB bus. To check whether or not the state is updated, read this bit.

SUSP_W
Field SUSP writer - If this bit is set to ‘1’, the USB Host is placed into the suspend state. If this bit is set to ‘0’ while it is ‘1’ or the USB bus is placed into the k-state mode, then suspend state is released, and the RWIRQ bit of the Interrupt USB Host Register (INTR_USBHOST) is set to ‘1’. Set to ‘1’ : Suspend. Set ‘0’ when this bit is ‘1’ : Resume. Other conditions : Holds the status. Notes: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’. - The transition to disconnected on RST isn’t immediate. Read this bit to confirm the transition is complete. - If this bit is set to ‘1’, this bit must not be set to ‘1’ until the RWIRQ bit of the Interrupt USB Host Register (INTR_USBHOST) is set to ‘1’. - Do not set this bit to ‘1’ while the USB is active (during USB bus resetting, data transfer, or SOF timer running). - If the value of this bit is changed, it is not immediately reflected on the state of the USB bus. To check whether or not the state is updated, read this bit.

TMODE_R
Field TMODE reader - If this bit is ‘1’, it means that the device is connected in the full-speed mode. When this bit is ‘0’, it means that the device is connected in the low-speed mode. This bit is valid when the CSTAT bit of the Host Status Register (HOST_STATUS) is ‘1’. ‘0’ : Low-speed. ‘1’ : Full-speed. Notes: - This bit is set to the default value if the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’. - The transition to disconnected on RST isn’t immediate. Read this bit to confirm the transition is complete.

URST_R
Field URST reader - When this bit is set to ‘1’, the USB bus is reset. This bit remains a ‘1’ during USB bus resetting, and changes to ‘0’ when USB bus resetting is ended. If this bit is set to ‘0’, the USB bus reset is complete

URST_W
Field URST writer - When this bit is set to ‘1’, the USB bus is reset. This bit remains a ‘1’ during USB bus resetting, and changes to ‘0’ when USB bus resetting is ended. If this bit is set to ‘0’, the USB bus reset is complete

W
Register HOST_STATUS writer