Module psoc6_01_pac::usbfs0::usbhost::host_token

Host Token Endpoint Register

Structs

ENDPT_R
Field ENDPT reader - These bits are used to specify an endpoint to send or receive data to or from the device. Note: - This bit isn’t reset to default even if the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

ENDPT_W
Field ENDPT writer - These bits are used to specify an endpoint to send or receive data to or from the device. Note: - This bit isn’t reset to default even if the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’.

HOST_TOKEN_SPEC
Host Token Endpoint Register

R
Register HOST_TOKEN reader

TGGL_R
Field TGGL reader - This bit is used to set toggle data. Toggle data is sent depending on the setting of this bit. When receiving toggle data, received toggle data is compared with the toggle data of this bit to verify whether or not an error occurs. ‘0’ : DATA0 ‘1’ : DATA1 Notes: - This bit isn’t reset to the default value even if the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’. - Set this bit when the TKNEN bit of the Host Token Endpoint Register (HOST_TOKEN) is ‘000’.

TGGL_W
Field TGGL writer - This bit is used to set toggle data. Toggle data is sent depending on the setting of this bit. When receiving toggle data, received toggle data is compared with the toggle data of this bit to verify whether or not an error occurs. ‘0’ : DATA0 ‘1’ : DATA1 Notes: - This bit isn’t reset to the default value even if the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’. - Set this bit when the TKNEN bit of the Host Token Endpoint Register (HOST_TOKEN) is ‘000’.

TKNEN_R
Field TKNEN reader - These bits send a token according to the current settings. After operation is complete, the TKNEN bit is set to ‘000’, and the CMPIRQ bit of the Interrupt USB Host Register (INTR_USBHOST) is set to ‘1’. The settings of the TGGL and ENDPT bits are ignored when sending a SOF token. Notes: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’. - The PRE packet isn’t supported. - Do not set ‘100’ to the TKNEN bit when the SOFBUSY bit of the Host Status Register (HOST_STATUS) is ‘1’ - Mode should be USB Host before writing data to this bit. - When issuing a token again after the token interrupt flag (CMPIRQ) has been set to ‘1’, wait for 3 cycles or more after a USB transfer clock (12 MHz in the full-speed mode, 1.5 MHz in the low-speed mode) was output, then write data to this bit. - Read the value of TKNEN bit if a new value is written in it .Continue writing in this bit until a retrieved value equals a new value written in. During this checking process, it is needed to prevent any interrupt. - Take the following steps when CMPIRQ bit of Interrupt USB Host Register (INTR_USBHOST) is set to ‘1’ by finishing IN token or Isochronous IN token. 1. Read HS bit of Host Error Status Register (HOST_ERR), then set CMPIRQ bit to ‘0’. 2. Set EPn bit of Host DMA Enable Register (HOST_DMA_ENBL) (n=1 or 2) to ‘1’ if HS bit of Host Error Status Register (HOST_ERR) is equal to ‘00’ and wait until EPn bit of Host DMA Data Request Register (HOST_DMA_DREQ) changes to ‘1’. Finish the IN token processing if HS bit is not equal to ‘00’. 3. Read the received data if EPn bit of Host DMA Data Requet (HOST_DMA_DREQ) (n=1 or 2) changes to ‘1’.

TKNEN_W
Field TKNEN writer - These bits send a token according to the current settings. After operation is complete, the TKNEN bit is set to ‘000’, and the CMPIRQ bit of the Interrupt USB Host Register (INTR_USBHOST) is set to ‘1’. The settings of the TGGL and ENDPT bits are ignored when sending a SOF token. Notes: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’. - The PRE packet isn’t supported. - Do not set ‘100’ to the TKNEN bit when the SOFBUSY bit of the Host Status Register (HOST_STATUS) is ‘1’ - Mode should be USB Host before writing data to this bit. - When issuing a token again after the token interrupt flag (CMPIRQ) has been set to ‘1’, wait for 3 cycles or more after a USB transfer clock (12 MHz in the full-speed mode, 1.5 MHz in the low-speed mode) was output, then write data to this bit. - Read the value of TKNEN bit if a new value is written in it .Continue writing in this bit until a retrieved value equals a new value written in. During this checking process, it is needed to prevent any interrupt. - Take the following steps when CMPIRQ bit of Interrupt USB Host Register (INTR_USBHOST) is set to ‘1’ by finishing IN token or Isochronous IN token. 1. Read HS bit of Host Error Status Register (HOST_ERR), then set CMPIRQ bit to ‘0’. 2. Set EPn bit of Host DMA Enable Register (HOST_DMA_ENBL) (n=1 or 2) to ‘1’ if HS bit of Host Error Status Register (HOST_ERR) is equal to ‘00’ and wait until EPn bit of Host DMA Data Request Register (HOST_DMA_DREQ) changes to ‘1’. Finish the IN token processing if HS bit is not equal to ‘00’. 3. Read the received data if EPn bit of Host DMA Data Requet (HOST_DMA_DREQ) (n=1 or 2) changes to ‘1’.

W
Register HOST_TOKEN writer

Enums

TKNEN_A
These bits send a token according to the current settings. After operation is complete, the TKNEN bit is set to ‘000’, and the CMPIRQ bit of the Interrupt USB Host Register (INTR_USBHOST) is set to ‘1’. The settings of the TGGL and ENDPT bits are ignored when sending a SOF token. Notes: - This bit is set to the default value when the RST bit of the Host Control 1 Register (HOST_CTL1) is set to ‘1’. - The PRE packet isn’t supported. - Do not set ‘100’ to the TKNEN bit when the SOFBUSY bit of the Host Status Register (HOST_STATUS) is ‘1’ - Mode should be USB Host before writing data to this bit. - When issuing a token again after the token interrupt flag (CMPIRQ) has been set to ‘1’, wait for 3 cycles or more after a USB transfer clock (12 MHz in the full-speed mode, 1.5 MHz in the low-speed mode) was output, then write data to this bit. - Read the value of TKNEN bit if a new value is written in it .Continue writing in this bit until a retrieved value equals a new value written in. During this checking process, it is needed to prevent any interrupt. - Take the following steps when CMPIRQ bit of Interrupt USB Host Register (INTR_USBHOST) is set to ‘1’ by finishing IN token or Isochronous IN token. 1. Read HS bit of Host Error Status Register (HOST_ERR), then set CMPIRQ bit to ‘0’. 2. Set EPn bit of Host DMA Enable Register (HOST_DMA_ENBL) (n=1 or 2) to ‘1’ if HS bit of Host Error Status Register (HOST_ERR) is equal to ‘00’ and wait until EPn bit of Host DMA Data Request Register (HOST_DMA_DREQ) changes to ‘1’. Finish the IN token processing if HS bit is not equal to ‘00’. 3. Read the received data if EPn bit of Host DMA Data Requet (HOST_DMA_DREQ) (n=1 or 2) changes to ‘1’.