Module psoc6_01_pac::usbfs0::usblpm::flow_ctl

Flow Control Register

Structs

EP1_ERR_RESP_R
Field EP1_ERR_RESP reader - End Point 1 error response 0: do nothing (backward compatibility mode) 1: if this is an IN EP and an underflow occurs then cause a CRC error, if this is an OUT EP and an overflow occurs then send a NAK

EP1_ERR_RESP_W
Field EP1_ERR_RESP writer - End Point 1 error response 0: do nothing (backward compatibility mode) 1: if this is an IN EP and an underflow occurs then cause a CRC error, if this is an OUT EP and an overflow occurs then send a NAK

EP2_ERR_RESP_R
Field EP2_ERR_RESP reader - End Point 2 error response

EP2_ERR_RESP_W
Field EP2_ERR_RESP writer - End Point 2 error response

EP3_ERR_RESP_R
Field EP3_ERR_RESP reader - End Point 3 error response

EP3_ERR_RESP_W
Field EP3_ERR_RESP writer - End Point 3 error response

EP4_ERR_RESP_R
Field EP4_ERR_RESP reader - End Point 4 error response

EP4_ERR_RESP_W
Field EP4_ERR_RESP writer - End Point 4 error response

EP5_ERR_RESP_R
Field EP5_ERR_RESP reader - End Point 5 error response

EP5_ERR_RESP_W
Field EP5_ERR_RESP writer - End Point 5 error response

EP6_ERR_RESP_R
Field EP6_ERR_RESP reader - End Point 6 error response

EP6_ERR_RESP_W
Field EP6_ERR_RESP writer - End Point 6 error response

EP7_ERR_RESP_R
Field EP7_ERR_RESP reader - End Point 7 error response

EP7_ERR_RESP_W
Field EP7_ERR_RESP writer - End Point 7 error response

EP8_ERR_RESP_R
Field EP8_ERR_RESP reader - End Point 8 error response

EP8_ERR_RESP_W
Field EP8_ERR_RESP writer - End Point 8 error response

FLOW_CTL_SPEC
Flow Control Register

R
Register FLOW_CTL reader

W
Register FLOW_CTL writer