Module psoc6_01_pac::usbfs0::usblpm::power_ctl

Power Control Register

Structs

DM_BIG_R
Field DM_BIG reader - Select the resister value if POWER_CTL.DM_EN=‘1’. This bit is valid in GPIO. ‘0’ : The resister value is from 900 to1575Ohmpull up on the DM. ‘1’ : The resister value is from 1425 to 3090Ohmpull up on the DM

DM_BIG_W
Field DM_BIG writer - Select the resister value if POWER_CTL.DM_EN=‘1’. This bit is valid in GPIO. ‘0’ : The resister value is from 900 to1575Ohmpull up on the DM. ‘1’ : The resister value is from 1425 to 3090Ohmpull up on the DM

DM_DOWN_EN_R
Field DM_DOWN_EN reader - Enables the ~15k pull down on the DP.

DM_DOWN_EN_W
Field DM_DOWN_EN writer - Enables the ~15k pull down on the DP.

DM_UP_EN_R
Field DM_UP_EN reader - Enables the pull up on the DM. The bit is valid in GPIO. The pull up resistor is disabled in not GPIO. ‘0’ : Disable. ‘1’ : Enable.

DM_UP_EN_W
Field DM_UP_EN writer - Enables the pull up on the DM. The bit is valid in GPIO. The pull up resistor is disabled in not GPIO. ‘0’ : Disable. ‘1’ : Enable.

DP_BIG_R
Field DP_BIG reader - Select the resister value if POWER_CTL.DP_EN=‘1’. This bit is valid in GPIO. ‘0’ : The resister value is from 900 to1575Ohmpull up on the DP. ‘1’ : The resister value is from 1425 to 3090Ohmpull up on the DP

DP_BIG_W
Field DP_BIG writer - Select the resister value if POWER_CTL.DP_EN=‘1’. This bit is valid in GPIO. ‘0’ : The resister value is from 900 to1575Ohmpull up on the DP. ‘1’ : The resister value is from 1425 to 3090Ohmpull up on the DP

DP_DOWN_EN_R
Field DP_DOWN_EN reader - Enables the ~15k pull down on the DP.

DP_DOWN_EN_W
Field DP_DOWN_EN writer - Enables the ~15k pull down on the DP.

DP_UP_EN_R
Field DP_UP_EN reader - Enables the pull up on the DP. ‘0’ : Disable. ‘1’ : Enable.

DP_UP_EN_W
Field DP_UP_EN writer - Enables the pull up on the DP. ‘0’ : Disable. ‘1’ : Enable.

ENABLE_DMO_R
Field ENABLE_DMO reader - Enables the signle ended receiver on D-.

ENABLE_DMO_W
Field ENABLE_DMO writer - Enables the signle ended receiver on D-.

ENABLE_DPO_R
Field ENABLE_DPO reader - Enables the single ended receiver on D+.

ENABLE_DPO_W
Field ENABLE_DPO writer - Enables the single ended receiver on D+.

POWER_CTL_SPEC
Power Control Register

R
Register POWER_CTL reader

SUSPEND_R
Field SUSPEND reader - Put PHY into Suspend mode. If the PHY is enabled, this bit MUST be set before entering a low power mode (DeepSleep). Note: - This bit is invalid if the HOST bit of the Host Control 0 Register (HOST_CTL0) is ‘1’.

SUSPEND_W
Field SUSPEND writer - Put PHY into Suspend mode. If the PHY is enabled, this bit MUST be set before entering a low power mode (DeepSleep). Note: - This bit is invalid if the HOST bit of the Host Control 0 Register (HOST_CTL0) is ‘1’.

W
Register POWER_CTL writer