Struct psoc6_01_pac::usbfs0::usblpm::power_ctl::W

pub struct W(_);

Register POWER_CTL writer

Implementations

impl W

pub fn suspend(&mut self) -> SUSPEND_W<'_>

Bit 2 - Put PHY into Suspend mode. If the PHY is enabled, this bit MUST be set before entering a low power mode (DeepSleep). Note: - This bit is invalid if the HOST bit of the Host Control 0 Register (HOST_CTL0) is ‘1’.

pub fn dp_up_en(&mut self) -> DP_UP_EN_W<'_>

Bit 16 - Enables the pull up on the DP. ‘0’ : Disable. ‘1’ : Enable.

pub fn dp_big(&mut self) -> DP_BIG_W<'_>

Bit 17 - Select the resister value if POWER_CTL.DP_EN=‘1’. This bit is valid in GPIO. ‘0’ : The resister value is from 900 to1575Ohmpull up on the DP. ‘1’ : The resister value is from 1425 to 3090Ohmpull up on the DP

pub fn dp_down_en(&mut self) -> DP_DOWN_EN_W<'_>

Bit 18 - Enables the ~15k pull down on the DP.

pub fn dm_up_en(&mut self) -> DM_UP_EN_W<'_>

Bit 19 - Enables the pull up on the DM. The bit is valid in GPIO. The pull up resistor is disabled in not GPIO. ‘0’ : Disable. ‘1’ : Enable.

pub fn dm_big(&mut self) -> DM_BIG_W<'_>

Bit 20 - Select the resister value if POWER_CTL.DM_EN=‘1’. This bit is valid in GPIO. ‘0’ : The resister value is from 900 to1575Ohmpull up on the DM. ‘1’ : The resister value is from 1425 to 3090Ohmpull up on the DM

pub fn dm_down_en(&mut self) -> DM_DOWN_EN_W<'_>

Bit 21 - Enables the ~15k pull down on the DP.

pub fn enable_dpo(&mut self) -> ENABLE_DPO_W<'_>

Bit 28 - Enables the single ended receiver on D+.

pub fn enable_dmo(&mut self) -> ENABLE_DMO_W<'_>

Bit 29 - Enables the signle ended receiver on D-.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.